rp2040_pac::adc
Module cs

ADC Control and Status

Structs

CS_SPEC

ADC Control and Status

Type Aliases

AINSEL_R

Field AINSEL reader - Select analog mux input. Updated automatically in round-robin mode.

AINSEL_W

Field AINSEL writer - Select analog mux input. Updated automatically in round-robin mode.

EN_R

Field EN reader - Power on ADC and enable its clock.
1 - enabled. 0 - disabled.

EN_W

Field EN writer - Power on ADC and enable its clock.
1 - enabled. 0 - disabled.

ERR_R

Field ERR reader - The most recent ADC conversion encountered an error; result is undefined or noisy.

ERR_STICKY_R

Field ERR_STICKY reader - Some past ADC conversion encountered an error. Write 1 to clear.

ERR_STICKY_W

Field ERR_STICKY writer - Some past ADC conversion encountered an error. Write 1 to clear.

R

Register CS reader

READY_R

Field READY reader - 1 if the ADC is ready to start a new conversion. Implies any previous conversion has completed.
0 whilst conversion in progress.

RROBIN_R

Field RROBIN reader - Round-robin sampling. 1 bit per channel. Set all bits to 0 to disable.
Otherwise, the ADC will cycle through each enabled channel in a round-robin fashion.
The first channel to be sampled will be the one currently indicated by AINSEL.
AINSEL will be updated after each conversion with the newly-selected channel.

RROBIN_W

Field RROBIN writer - Round-robin sampling. 1 bit per channel. Set all bits to 0 to disable.
Otherwise, the ADC will cycle through each enabled channel in a round-robin fashion.
The first channel to be sampled will be the one currently indicated by AINSEL.
AINSEL will be updated after each conversion with the newly-selected channel.

START_MANY_R

Field START_MANY reader - Continuously perform conversions whilst this bit is 1. A new conversion will start immediately after the previous finishes.

START_MANY_W

Field START_MANY writer - Continuously perform conversions whilst this bit is 1. A new conversion will start immediately after the previous finishes.

START_ONCE_R

Field START_ONCE reader - Start a single conversion. Self-clearing. Ignored if start_many is asserted.

START_ONCE_W

Field START_ONCE writer - Start a single conversion. Self-clearing. Ignored if start_many is asserted.

TS_EN_R

Field TS_EN reader - Power on temperature sensor. 1 - enabled. 0 - disabled.

TS_EN_W

Field TS_EN writer - Power on temperature sensor. 1 - enabled. 0 - disabled.

W

Register CS writer