rp2040_pac::busctrl
Module bus_priority

Set the priority of each master for bus arbitration.

Structs

BUS_PRIORITY_SPEC

Set the priority of each master for bus arbitration.

Type Aliases

DMA_R_R

Field DMA_R reader - 0 - low priority, 1 - high priority

DMA_R_W

Field DMA_R writer - 0 - low priority, 1 - high priority

DMA_W_R

Field DMA_W reader - 0 - low priority, 1 - high priority

DMA_W_W

Field DMA_W writer - 0 - low priority, 1 - high priority

PROC0_R

Field PROC0 reader - 0 - low priority, 1 - high priority

PROC0_W

Field PROC0 writer - 0 - low priority, 1 - high priority

PROC1_R

Field PROC1 reader - 0 - low priority, 1 - high priority

PROC1_W

Field PROC1 writer - 0 - low priority, 1 - high priority

R

Register BUS_PRIORITY reader

W

Register BUS_PRIORITY writer