rp2040_pac::clocks

Module clk_sys_ctrl

Clock control, can be changed on-the-fly (except for auxsrc)

Structs

CLK_SYS_CTRL_SPEC

Clock control, can be changed on-the-fly (except for auxsrc)

Enums

AUXSRC_A

Selects the auxiliary clock source, will glitch when switching

SRC_A

Selects the clock source glitchlessly, can be changed on-the-fly

Type Aliases

AUXSRC_R

Field AUXSRC reader - Selects the auxiliary clock source, will glitch when switching

AUXSRC_W

Field AUXSRC writer - Selects the auxiliary clock source, will glitch when switching

R

Register CLK_SYS_CTRL reader

SRC_R

Field SRC reader - Selects the clock source glitchlessly, can be changed on-the-fly

SRC_W

Field SRC writer - Selects the clock source glitchlessly, can be changed on-the-fly

W

Register CLK_SYS_CTRL writer