Type Alias rp2040_pac::clocks::clk_usb_ctrl::W

pub type W = W<CLK_USB_CTRL_SPEC>;

Register CLK_USB_CTRL writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn auxsrc(&mut self) -> AUXSRC_W<'_, CLK_USB_CTRL_SPEC>

Bits 5:7 - Selects the auxiliary clock source, will glitch when switching

pub fn kill(&mut self) -> KILL_W<'_, CLK_USB_CTRL_SPEC>

Bit 10 - Asynchronously kills the clock generator

pub fn enable(&mut self) -> ENABLE_W<'_, CLK_USB_CTRL_SPEC>

Bit 11 - Starts and stops the clock generator cleanly

pub fn phase(&mut self) -> PHASE_W<'_, CLK_USB_CTRL_SPEC>

Bits 16:17 - This delays the enable signal by up to 3 cycles of the input clock
This must be set before the clock is enabled to have any effect

pub fn nudge(&mut self) -> NUDGE_W<'_, CLK_USB_CTRL_SPEC>

Bit 20 - An edge on this signal shifts the phase of the output by 1 cycle of the input clock
This can be done at any time

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual