rp2040_pac::clocks
Module fc0_status

Frequency counter status

Structs

FC0_STATUS_SPEC

Frequency counter status

Type Aliases

DIED_R

Field DIED reader - Test clock stopped during test

DONE_R

Field DONE reader - Test complete

FAIL_R

Field FAIL reader - Test failed

FAST_R

Field FAST reader - Test clock faster than expected, only valid when status_done=1

PASS_R

Field PASS reader - Test passed

R

Register FC0_STATUS reader

RUNNING_R

Field RUNNING reader - Test running

SLOW_R

Field SLOW reader - Test clock slower than expected, only valid when status_done=1

WAITING_R

Field WAITING reader - Waiting for test clock to start