Struct rp2040_pac::clocks::fields::ClkRtcCtrl

#[repr(transparent)]pub struct ClkRtcCtrl(_);

Clock control, can be changed on-the-fly (except for auxsrc)

Implementations

impl ClkRtcCtrl

pub const fn to_bits(&self) -> u32

pub const fn from_bits(val: u32) -> ClkRtcCtrl

pub const fn nudge(&self) -> bool

An edge on this signal shifts the phase of the output by 1 cycle of the input clock This can be done at any time

pub fn set_nudge(&mut self, val: bool)

An edge on this signal shifts the phase of the output by 1 cycle of the input clock This can be done at any time

pub const fn phase(&self) -> u8

This delays the enable signal by up to 3 cycles of the input clock This must be set before the clock is enabled to have any effect

pub fn set_phase(&mut self, val: u8)

This delays the enable signal by up to 3 cycles of the input clock This must be set before the clock is enabled to have any effect

pub const fn enable(&self) -> bool

Starts and stops the clock generator cleanly

pub fn set_enable(&mut self, val: bool)

Starts and stops the clock generator cleanly

pub const fn kill(&self) -> bool

Asynchronously kills the clock generator

pub fn set_kill(&mut self, val: bool)

Asynchronously kills the clock generator

pub const fn auxsrc(&self) -> ClkRtcCtrlAuxsrc

Selects the auxiliary clock source, will glitch when switching

pub fn set_auxsrc(&mut self, val: ClkRtcCtrlAuxsrc)

Selects the auxiliary clock source, will glitch when switching