Struct rp2040_pac::clocks::Clocks

pub struct Clocks(_);

Implementations

impl Clocks

pub const fn from_ptr(ptr: *mut u8) -> Self

pub fn clk_gpout0_ctrl(self) -> Reg<ClkGpout0Ctrl, RW>

Clock control, can be changed on-the-fly (except for auxsrc)

pub fn clk_gpout0_div(self) -> Reg<ClkGpout0Div, RW>

Clock divisor, can be changed on-the-fly

pub fn clk_gpout0_selected(self) -> Reg<u32, R>

Indicates which src is currently selected (one-hot)

pub fn clk_gpout1_ctrl(self) -> Reg<ClkGpout1Ctrl, RW>

Clock control, can be changed on-the-fly (except for auxsrc)

pub fn clk_gpout1_div(self) -> Reg<ClkGpout1Div, RW>

Clock divisor, can be changed on-the-fly

pub fn clk_gpout1_selected(self) -> Reg<u32, R>

Indicates which src is currently selected (one-hot)

pub fn clk_gpout2_ctrl(self) -> Reg<ClkGpout2Ctrl, RW>

Clock control, can be changed on-the-fly (except for auxsrc)

pub fn clk_gpout2_div(self) -> Reg<ClkGpout2Div, RW>

Clock divisor, can be changed on-the-fly

pub fn clk_gpout2_selected(self) -> Reg<u32, R>

Indicates which src is currently selected (one-hot)

pub fn clk_gpout3_ctrl(self) -> Reg<ClkGpout3Ctrl, RW>

Clock control, can be changed on-the-fly (except for auxsrc)

pub fn clk_gpout3_div(self) -> Reg<ClkGpout3Div, RW>

Clock divisor, can be changed on-the-fly

pub fn clk_gpout3_selected(self) -> Reg<u32, R>

Indicates which src is currently selected (one-hot)

pub fn clk_ref_ctrl(self) -> Reg<ClkRefCtrl, RW>

Clock control, can be changed on-the-fly (except for auxsrc)

pub fn clk_ref_div(self) -> Reg<ClkRefDiv, RW>

Clock divisor, can be changed on-the-fly

pub fn clk_ref_selected(self) -> Reg<u32, R>

Indicates which src is currently selected (one-hot)

pub fn clk_sys_ctrl(self) -> Reg<ClkSysCtrl, RW>

Clock control, can be changed on-the-fly (except for auxsrc)

pub fn clk_sys_div(self) -> Reg<ClkSysDiv, RW>

Clock divisor, can be changed on-the-fly

pub fn clk_sys_selected(self) -> Reg<u32, R>

Indicates which src is currently selected (one-hot)

pub fn clk_peri_ctrl(self) -> Reg<ClkPeriCtrl, RW>

Clock control, can be changed on-the-fly (except for auxsrc)

pub fn clk_peri_selected(self) -> Reg<u32, R>

Indicates which src is currently selected (one-hot)

pub fn clk_usb_ctrl(self) -> Reg<ClkUsbCtrl, RW>

Clock control, can be changed on-the-fly (except for auxsrc)

pub fn clk_usb_div(self) -> Reg<ClkUsbDiv, RW>

Clock divisor, can be changed on-the-fly

pub fn clk_usb_selected(self) -> Reg<u32, R>

Indicates which src is currently selected (one-hot)

pub fn clk_adc_ctrl(self) -> Reg<ClkAdcCtrl, RW>

Clock control, can be changed on-the-fly (except for auxsrc)

pub fn clk_adc_div(self) -> Reg<ClkAdcDiv, RW>

Clock divisor, can be changed on-the-fly

pub fn clk_adc_selected(self) -> Reg<u32, R>

Indicates which src is currently selected (one-hot)

pub fn clk_rtc_ctrl(self) -> Reg<ClkRtcCtrl, RW>

Clock control, can be changed on-the-fly (except for auxsrc)

pub fn clk_rtc_div(self) -> Reg<ClkRtcDiv, RW>

Clock divisor, can be changed on-the-fly

pub fn clk_rtc_selected(self) -> Reg<u32, R>

Indicates which src is currently selected (one-hot)

pub fn clk_sys_resus_ctrl(self) -> Reg<ClkSysResusCtrl, RW>

pub fn clk_sys_resus_status(self) -> Reg<ClkSysResusStatus, RW>

pub fn fc0_ref_khz(self) -> Reg<Fc0RefKhz, RW>

Reference clock frequency in kHz

pub fn fc0_min_khz(self) -> Reg<Fc0MinKhz, RW>

Minimum pass frequency in kHz. This is optional. Set to 0 if you are not using the pass/fail flags

pub fn fc0_max_khz(self) -> Reg<Fc0MaxKhz, RW>

Maximum pass frequency in kHz. This is optional. Set to 0x1ffffff if you are not using the pass/fail flags

pub fn fc0_delay(self) -> Reg<Fc0Delay, RW>

Delays the start of frequency counting to allow the mux to settle Delay is measured in multiples of the reference clock period

pub fn fc0_interval(self) -> Reg<Fc0Interval, RW>

The test interval is 0.98us * 2interval, but let's call it 1us * 2interval The default gives a test interval of 250us

pub fn fc0_src(self) -> Reg<Fc0Src, RW>

Clock sent to frequency counter, set to 0 when not required Writing to this register initiates the frequency count

pub fn fc0_status(self) -> Reg<Fc0Status, RW>

Frequency counter status

pub fn fc0_result(self) -> Reg<Fc0Result, RW>

Result of frequency measurement, only valid when status_done=1

pub fn wake_en0(self) -> Reg<WakeEn0, RW>

enable clock in wake mode

pub fn wake_en1(self) -> Reg<WakeEn1, RW>

enable clock in wake mode

pub fn sleep_en0(self) -> Reg<SleepEn0, RW>

enable clock in sleep mode

pub fn sleep_en1(self) -> Reg<SleepEn1, RW>

enable clock in sleep mode

pub fn enabled0(self) -> Reg<Enabled0, RW>

indicates the state of the clock enable

pub fn enabled1(self) -> Reg<Enabled1, RW>

indicates the state of the clock enable

pub fn intr(self) -> Reg<Intr, RW>

Raw Interrupts

pub fn inte(self) -> Reg<Inte, RW>

Interrupt Enable

pub fn intf(self) -> Reg<Intf, RW>

Interrupt Force

pub fn ints(self) -> Reg<Ints, RW>

Interrupt status after masking & forcing