Struct rp2040_pac::clocks::RegisterBlock

#[repr(C)]
pub struct RegisterBlock {

    pub clk_gpout0_ctrl: Reg<CLK_GPOUT0_CTRL_SPEC>,
    pub clk_gpout0_div: Reg<CLK_GPOUT0_DIV_SPEC>,
    pub clk_gpout0_selected: Reg<CLK_GPOUT0_SELECTED_SPEC>,
    pub clk_gpout1_ctrl: Reg<CLK_GPOUT1_CTRL_SPEC>,
    pub clk_gpout1_div: Reg<CLK_GPOUT1_DIV_SPEC>,
    pub clk_gpout1_selected: Reg<CLK_GPOUT1_SELECTED_SPEC>,
    pub clk_gpout2_ctrl: Reg<CLK_GPOUT2_CTRL_SPEC>,
    pub clk_gpout2_div: Reg<CLK_GPOUT2_DIV_SPEC>,
    pub clk_gpout2_selected: Reg<CLK_GPOUT2_SELECTED_SPEC>,
    pub clk_gpout3_ctrl: Reg<CLK_GPOUT3_CTRL_SPEC>,
    pub clk_gpout3_div: Reg<CLK_GPOUT3_DIV_SPEC>,
    pub clk_gpout3_selected: Reg<CLK_GPOUT3_SELECTED_SPEC>,
    pub clk_ref_ctrl: Reg<CLK_REF_CTRL_SPEC>,
    pub clk_ref_div: Reg<CLK_REF_DIV_SPEC>,
    pub clk_ref_selected: Reg<CLK_REF_SELECTED_SPEC>,
    pub clk_sys_ctrl: Reg<CLK_SYS_CTRL_SPEC>,
    pub clk_sys_div: Reg<CLK_SYS_DIV_SPEC>,
    pub clk_sys_selected: Reg<CLK_SYS_SELECTED_SPEC>,
    pub clk_peri_ctrl: Reg<CLK_PERI_CTRL_SPEC>,
    pub clk_peri_selected: Reg<CLK_PERI_SELECTED_SPEC>,
    pub clk_usb_ctrl: Reg<CLK_USB_CTRL_SPEC>,
    pub clk_usb_div: Reg<CLK_USB_DIV_SPEC>,
    pub clk_usb_selected: Reg<CLK_USB_SELECTED_SPEC>,
    pub clk_adc_ctrl: Reg<CLK_ADC_CTRL_SPEC>,
    pub clk_adc_div: Reg<CLK_ADC_DIV_SPEC>,
    pub clk_adc_selected: Reg<CLK_ADC_SELECTED_SPEC>,
    pub clk_rtc_ctrl: Reg<CLK_RTC_CTRL_SPEC>,
    pub clk_rtc_div: Reg<CLK_RTC_DIV_SPEC>,
    pub clk_rtc_selected: Reg<CLK_RTC_SELECTED_SPEC>,
    pub clk_sys_resus_ctrl: Reg<CLK_SYS_RESUS_CTRL_SPEC>,
    pub clk_sys_resus_status: Reg<CLK_SYS_RESUS_STATUS_SPEC>,
    pub fc0_ref_khz: Reg<FC0_REF_KHZ_SPEC>,
    pub fc0_min_khz: Reg<FC0_MIN_KHZ_SPEC>,
    pub fc0_max_khz: Reg<FC0_MAX_KHZ_SPEC>,
    pub fc0_delay: Reg<FC0_DELAY_SPEC>,
    pub fc0_interval: Reg<FC0_INTERVAL_SPEC>,
    pub fc0_src: Reg<FC0_SRC_SPEC>,
    pub fc0_status: Reg<FC0_STATUS_SPEC>,
    pub fc0_result: Reg<FC0_RESULT_SPEC>,
    pub wake_en0: Reg<WAKE_EN0_SPEC>,
    pub wake_en1: Reg<WAKE_EN1_SPEC>,
    pub sleep_en0: Reg<SLEEP_EN0_SPEC>,
    pub sleep_en1: Reg<SLEEP_EN1_SPEC>,
    pub enabled0: Reg<ENABLED0_SPEC>,
    pub enabled1: Reg<ENABLED1_SPEC>,
    pub intr: Reg<INTR_SPEC>,
    pub inte: Reg<INTE_SPEC>,
    pub intf: Reg<INTF_SPEC>,
    pub ints: Reg<INTS_SPEC>,
    /* private fields */

}

Register block

Fields
clk_gpout0_ctrl: Reg<CLK_GPOUT0_CTRL_SPEC>

0x00 - Clock control, can be changed on-the-fly (except for auxsrc)

clk_gpout0_div: Reg<CLK_GPOUT0_DIV_SPEC>

0x04 - Clock divisor, can be changed on-the-fly

clk_gpout0_selected: Reg<CLK_GPOUT0_SELECTED_SPEC>

0x08 - Indicates which SRC is currently selected by the glitchless mux (one-hot).
This slice does not have a glitchless mux (only the AUX_SRC field is present, not SRC) so this register is hardwired to 0x1.

clk_gpout1_ctrl: Reg<CLK_GPOUT1_CTRL_SPEC>

0x0c - Clock control, can be changed on-the-fly (except for auxsrc)

clk_gpout1_div: Reg<CLK_GPOUT1_DIV_SPEC>

0x10 - Clock divisor, can be changed on-the-fly

clk_gpout1_selected: Reg<CLK_GPOUT1_SELECTED_SPEC>

0x14 - Indicates which SRC is currently selected by the glitchless mux (one-hot).
This slice does not have a glitchless mux (only the AUX_SRC field is present, not SRC) so this register is hardwired to 0x1.

clk_gpout2_ctrl: Reg<CLK_GPOUT2_CTRL_SPEC>

0x18 - Clock control, can be changed on-the-fly (except for auxsrc)

clk_gpout2_div: Reg<CLK_GPOUT2_DIV_SPEC>

0x1c - Clock divisor, can be changed on-the-fly

clk_gpout2_selected: Reg<CLK_GPOUT2_SELECTED_SPEC>

0x20 - Indicates which SRC is currently selected by the glitchless mux (one-hot).
This slice does not have a glitchless mux (only the AUX_SRC field is present, not SRC) so this register is hardwired to 0x1.

clk_gpout3_ctrl: Reg<CLK_GPOUT3_CTRL_SPEC>

0x24 - Clock control, can be changed on-the-fly (except for auxsrc)

clk_gpout3_div: Reg<CLK_GPOUT3_DIV_SPEC>

0x28 - Clock divisor, can be changed on-the-fly

clk_gpout3_selected: Reg<CLK_GPOUT3_SELECTED_SPEC>

0x2c - Indicates which SRC is currently selected by the glitchless mux (one-hot).
This slice does not have a glitchless mux (only the AUX_SRC field is present, not SRC) so this register is hardwired to 0x1.

clk_ref_ctrl: Reg<CLK_REF_CTRL_SPEC>

0x30 - Clock control, can be changed on-the-fly (except for auxsrc)

clk_ref_div: Reg<CLK_REF_DIV_SPEC>

0x34 - Clock divisor, can be changed on-the-fly

clk_ref_selected: Reg<CLK_REF_SELECTED_SPEC>

0x38 - Indicates which SRC is currently selected by the glitchless mux (one-hot).
The glitchless multiplexer does not switch instantaneously (to avoid glitches), so software should poll this register to wait for the switch to complete. This register contains one decoded bit for each of the clock sources enumerated in the CTRL SRC field. At most one of these bits will be set at any time, indicating that clock is currently present at the output of the glitchless mux. Whilst switching is in progress, this register may briefly show all-0s.

clk_sys_ctrl: Reg<CLK_SYS_CTRL_SPEC>

0x3c - Clock control, can be changed on-the-fly (except for auxsrc)

clk_sys_div: Reg<CLK_SYS_DIV_SPEC>

0x40 - Clock divisor, can be changed on-the-fly

clk_sys_selected: Reg<CLK_SYS_SELECTED_SPEC>

0x44 - Indicates which SRC is currently selected by the glitchless mux (one-hot).
The glitchless multiplexer does not switch instantaneously (to avoid glitches), so software should poll this register to wait for the switch to complete. This register contains one decoded bit for each of the clock sources enumerated in the CTRL SRC field. At most one of these bits will be set at any time, indicating that clock is currently present at the output of the glitchless mux. Whilst switching is in progress, this register may briefly show all-0s.

clk_peri_ctrl: Reg<CLK_PERI_CTRL_SPEC>

0x48 - Clock control, can be changed on-the-fly (except for auxsrc)

clk_peri_selected: Reg<CLK_PERI_SELECTED_SPEC>

0x50 - Indicates which SRC is currently selected by the glitchless mux (one-hot).
This slice does not have a glitchless mux (only the AUX_SRC field is present, not SRC) so this register is hardwired to 0x1.

clk_usb_ctrl: Reg<CLK_USB_CTRL_SPEC>

0x54 - Clock control, can be changed on-the-fly (except for auxsrc)

clk_usb_div: Reg<CLK_USB_DIV_SPEC>

0x58 - Clock divisor, can be changed on-the-fly

clk_usb_selected: Reg<CLK_USB_SELECTED_SPEC>

0x5c - Indicates which SRC is currently selected by the glitchless mux (one-hot).
This slice does not have a glitchless mux (only the AUX_SRC field is present, not SRC) so this register is hardwired to 0x1.

clk_adc_ctrl: Reg<CLK_ADC_CTRL_SPEC>

0x60 - Clock control, can be changed on-the-fly (except for auxsrc)

clk_adc_div: Reg<CLK_ADC_DIV_SPEC>

0x64 - Clock divisor, can be changed on-the-fly

clk_adc_selected: Reg<CLK_ADC_SELECTED_SPEC>

0x68 - Indicates which SRC is currently selected by the glitchless mux (one-hot).
This slice does not have a glitchless mux (only the AUX_SRC field is present, not SRC) so this register is hardwired to 0x1.

clk_rtc_ctrl: Reg<CLK_RTC_CTRL_SPEC>

0x6c - Clock control, can be changed on-the-fly (except for auxsrc)

clk_rtc_div: Reg<CLK_RTC_DIV_SPEC>

0x70 - Clock divisor, can be changed on-the-fly

clk_rtc_selected: Reg<CLK_RTC_SELECTED_SPEC>

0x74 - Indicates which SRC is currently selected by the glitchless mux (one-hot).
This slice does not have a glitchless mux (only the AUX_SRC field is present, not SRC) so this register is hardwired to 0x1.

clk_sys_resus_ctrl: Reg<CLK_SYS_RESUS_CTRL_SPEC>

0x78 -

clk_sys_resus_status: Reg<CLK_SYS_RESUS_STATUS_SPEC>

0x7c -

fc0_ref_khz: Reg<FC0_REF_KHZ_SPEC>

0x80 - Reference clock frequency in kHz

fc0_min_khz: Reg<FC0_MIN_KHZ_SPEC>

0x84 - Minimum pass frequency in kHz. This is optional. Set to 0 if you are not using the pass/fail flags

fc0_max_khz: Reg<FC0_MAX_KHZ_SPEC>

0x88 - Maximum pass frequency in kHz. This is optional. Set to 0x1ffffff if you are not using the pass/fail flags

fc0_delay: Reg<FC0_DELAY_SPEC>

0x8c - Delays the start of frequency counting to allow the mux to settle
Delay is measured in multiples of the reference clock period

fc0_interval: Reg<FC0_INTERVAL_SPEC>

0x90 - The test interval is 0.98us * 2interval, but let’s call it 1us * 2interval
The default gives a test interval of 250us

fc0_src: Reg<FC0_SRC_SPEC>

0x94 - Clock sent to frequency counter, set to 0 when not required
Writing to this register initiates the frequency count

fc0_status: Reg<FC0_STATUS_SPEC>

0x98 - Frequency counter status

fc0_result: Reg<FC0_RESULT_SPEC>

0x9c - Result of frequency measurement, only valid when status_done=1

wake_en0: Reg<WAKE_EN0_SPEC>

0xa0 - enable clock in wake mode

wake_en1: Reg<WAKE_EN1_SPEC>

0xa4 - enable clock in wake mode

sleep_en0: Reg<SLEEP_EN0_SPEC>

0xa8 - enable clock in sleep mode

sleep_en1: Reg<SLEEP_EN1_SPEC>

0xac - enable clock in sleep mode

enabled0: Reg<ENABLED0_SPEC>

0xb0 - indicates the state of the clock enable

enabled1: Reg<ENABLED1_SPEC>

0xb4 - indicates the state of the clock enable

intr: Reg<INTR_SPEC>

0xb8 - Raw Interrupts

inte: Reg<INTE_SPEC>

0xbc - Interrupt Enable

intf: Reg<INTF_SPEC>

0xc0 - Interrupt Force

ints: Reg<INTS_SPEC>

0xc4 - Interrupt status after masking & forcing