rp2040_pac::dma::ch::ch_ctrl_trig

Type Alias RING_SIZE_R

pub type RING_SIZE_R = FieldReader<RING_SIZE_A>;

Field RING_SIZE reader - Size of address wrap region. If 0, don’t wrap. For values n > 0, only the lower n bits of the address will change. This wraps the address on a (1 << n) byte boundary, facilitating access to naturally-aligned ring buffers.

Ring sizes between 2 and 32768 bytes are possible. This can apply to either read or write addresses, based on value of RING_SEL.

Aliased Type

struct RING_SIZE_R { /* private fields */ }

Implementations

impl RING_SIZE_R

pub const fn variant(&self) -> Option<RING_SIZE_A>

Get enumerated values variant

pub fn is_ring_none(&self) -> bool

0