Module rp2040_pac::dma::fields

Structs

Ch0CtrlTrig
DMA Channel 0 Control and Status

Ch0DbgCtdreq
Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

Ch1CtrlTrig
DMA Channel 1 Control and Status

Ch1DbgCtdreq
Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

Ch2CtrlTrig
DMA Channel 2 Control and Status

Ch2DbgCtdreq
Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

Ch3CtrlTrig
DMA Channel 3 Control and Status

Ch3DbgCtdreq
Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

Ch4CtrlTrig
DMA Channel 4 Control and Status

Ch4DbgCtdreq
Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

Ch5CtrlTrig
DMA Channel 5 Control and Status

Ch5DbgCtdreq
Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

Ch6CtrlTrig
DMA Channel 6 Control and Status

Ch6DbgCtdreq
Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

Ch7CtrlTrig
DMA Channel 7 Control and Status

Ch7DbgCtdreq
Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

Ch8CtrlTrig
DMA Channel 8 Control and Status

Ch8DbgCtdreq
Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

Ch9CtrlTrig
DMA Channel 9 Control and Status

Ch9DbgCtdreq
Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

Ch10CtrlTrig
DMA Channel 10 Control and Status

Ch10DbgCtdreq
Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

Ch11CtrlTrig
DMA Channel 11 Control and Status

Ch11DbgCtdreq
Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

ChanAbort
Abort an in-progress transfer sequence on one or more channels

FifoLevels
Debug RAF, WAF, TDF levels

Inte0
Interrupt Enables for IRQ 0

Inte1
Interrupt Enables for IRQ 1

Intf0
Force Interrupts

Intf1
Force Interrupts for IRQ 1

Intr
Interrupt Status (raw)

Ints0
Interrupt Status for IRQ 0

Ints1
Interrupt Status (masked) for IRQ 1

MultiChanTrigger
Trigger one or more channels simultaneously

NChannels
The number of channels this DMA instance is equipped with. This DMA supports up to 16 hardware channels, but can be configured with as few as one, to minimise silicon area.

SniffCtrl
Sniffer Control

Timer0
Pacing (X/Y) Fractional Timer The pacing timer produces TREQ assertions at a rate set by ((X/Y) * sys_clk). This equation is evaluated every sys_clk cycles and therefore can only generate TREQs at a rate of 1 per sys_clk (i.e. permanent TREQ) or less.

Timer1
Pacing (X/Y) Fractional Timer The pacing timer produces TREQ assertions at a rate set by ((X/Y) * sys_clk). This equation is evaluated every sys_clk cycles and therefore can only generate TREQs at a rate of 1 per sys_clk (i.e. permanent TREQ) or less.