rp2040_pac::dma::ints0
Type Alias R

pub type R = R<INTS0_SPEC>;

Register INTS0 reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn ints0(&self) -> INTS0_R

Bits 0:15 - Indicates active channel interrupt requests which are currently causing IRQ 0 to be asserted.
Channel interrupts can be cleared by writing a bit mask here.