rp2040_pac::dma::sniff_ctrl

Type Alias R

pub type R = R<SNIFF_CTRL_SPEC>;

Register SNIFF_CTRL reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn en(&self) -> EN_R

Bit 0 - Enable sniffer

pub fn dmach(&self) -> DMACH_R

Bits 1:4 - DMA channel for Sniffer to observe

pub fn calc(&self) -> CALC_R

Bits 5:8

pub fn bswap(&self) -> BSWAP_R

Bit 9 - Locally perform a byte reverse on the sniffed data, before feeding into checksum.

Note that the sniff hardware is downstream of the DMA channel byteswap performed in the read master: if channel CTRL_BSWAP and SNIFF_CTRL_BSWAP are both enabled, their effects cancel from the sniffer’s point of view.

pub fn out_rev(&self) -> OUT_REV_R

Bit 10 - If set, the result appears bit-reversed when read. This does not affect the way the checksum is calculated; the result is transformed on-the-fly between the result register and the bus.

pub fn out_inv(&self) -> OUT_INV_R

Bit 11 - If set, the result appears inverted (bitwise complement) when read. This does not affect the way the checksum is calculated; the result is transformed on-the-fly between the result register and the bus.