Struct rp2040_pac::dma::Dma

pub struct Dma(_);

DMA with separate read and write masters

Implementations

impl Dma

pub const fn from_ptr(ptr: *mut u8) -> Self

pub fn ch0_read_addr(self) -> Reg<u32, RW>

DMA Channel 0 Read Address pointer This register updates automatically each time a read completes. The current value is the next address to be read by this channel.

pub fn ch0_write_addr(self) -> Reg<u32, RW>

DMA Channel 0 Write Address pointer This register updates automatically each time a write completes. The current value is the next address to be written by this channel.

pub fn ch0_trans_count(self) -> Reg<u32, RW>

DMA Channel 0 Transfer Count Program the number of bus transfers a channel will perform before halting. Note that, if transfers are larger than one byte in size, this is not equal to the number of bytes transferred (see CTRL_DATA_SIZE). When the channel is active, reading this register shows the number of transfers remaining, updating automatically each time a write transfer completes. Writing this register sets the RELOAD value for the transfer counter. Each time this channel is triggered, the RELOAD value is copied into the live transfer counter. The channel can be started multiple times, and will perform the same number of transfers each time, as programmed by most recent write. The RELOAD value can be observed at CHx_DBG_TCR. If TRANS_COUNT is used as a trigger, the written value is used immediately as the length of the new transfer sequence, as well as being written to RELOAD.

pub fn ch0_ctrl_trig(self) -> Reg<Ch0CtrlTrig, RW>

DMA Channel 0 Control and Status

pub fn ch0_al1_ctrl(self) -> Reg<u32, R>

Alias for channel 0 CTRL register

pub fn ch0_al1_read_addr(self) -> Reg<u32, R>

Alias for channel 0 READ_ADDR register

pub fn ch0_al1_write_addr(self) -> Reg<u32, R>

Alias for channel 0 WRITE_ADDR register

pub fn ch0_al1_trans_count_trig(self) -> Reg<u32, R>

Alias for channel 0 TRANS_COUNT register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch0_al2_ctrl(self) -> Reg<u32, R>

Alias for channel 0 CTRL register

pub fn ch0_al2_trans_count(self) -> Reg<u32, R>

Alias for channel 0 TRANS_COUNT register

pub fn ch0_al2_read_addr(self) -> Reg<u32, R>

Alias for channel 0 READ_ADDR register

pub fn ch0_al2_write_addr_trig(self) -> Reg<u32, R>

Alias for channel 0 WRITE_ADDR register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch0_al3_ctrl(self) -> Reg<u32, R>

Alias for channel 0 CTRL register

pub fn ch0_al3_write_addr(self) -> Reg<u32, R>

Alias for channel 0 WRITE_ADDR register

pub fn ch0_al3_trans_count(self) -> Reg<u32, R>

Alias for channel 0 TRANS_COUNT register

pub fn ch0_al3_read_addr_trig(self) -> Reg<u32, R>

Alias for channel 0 READ_ADDR register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch1_read_addr(self) -> Reg<u32, RW>

DMA Channel 1 Read Address pointer This register updates automatically each time a read completes. The current value is the next address to be read by this channel.

pub fn ch1_write_addr(self) -> Reg<u32, RW>

DMA Channel 1 Write Address pointer This register updates automatically each time a write completes. The current value is the next address to be written by this channel.

pub fn ch1_trans_count(self) -> Reg<u32, RW>

DMA Channel 1 Transfer Count Program the number of bus transfers a channel will perform before halting. Note that, if transfers are larger than one byte in size, this is not equal to the number of bytes transferred (see CTRL_DATA_SIZE). When the channel is active, reading this register shows the number of transfers remaining, updating automatically each time a write transfer completes. Writing this register sets the RELOAD value for the transfer counter. Each time this channel is triggered, the RELOAD value is copied into the live transfer counter. The channel can be started multiple times, and will perform the same number of transfers each time, as programmed by most recent write. The RELOAD value can be observed at CHx_DBG_TCR. If TRANS_COUNT is used as a trigger, the written value is used immediately as the length of the new transfer sequence, as well as being written to RELOAD.

pub fn ch1_ctrl_trig(self) -> Reg<Ch1CtrlTrig, RW>

DMA Channel 1 Control and Status

pub fn ch1_al1_ctrl(self) -> Reg<u32, R>

Alias for channel 1 CTRL register

pub fn ch1_al1_read_addr(self) -> Reg<u32, R>

Alias for channel 1 READ_ADDR register

pub fn ch1_al1_write_addr(self) -> Reg<u32, R>

Alias for channel 1 WRITE_ADDR register

pub fn ch1_al1_trans_count_trig(self) -> Reg<u32, R>

Alias for channel 1 TRANS_COUNT register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch1_al2_ctrl(self) -> Reg<u32, R>

Alias for channel 1 CTRL register

pub fn ch1_al2_trans_count(self) -> Reg<u32, R>

Alias for channel 1 TRANS_COUNT register

pub fn ch1_al2_read_addr(self) -> Reg<u32, R>

Alias for channel 1 READ_ADDR register

pub fn ch1_al2_write_addr_trig(self) -> Reg<u32, R>

Alias for channel 1 WRITE_ADDR register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch1_al3_ctrl(self) -> Reg<u32, R>

Alias for channel 1 CTRL register

pub fn ch1_al3_write_addr(self) -> Reg<u32, R>

Alias for channel 1 WRITE_ADDR register

pub fn ch1_al3_trans_count(self) -> Reg<u32, R>

Alias for channel 1 TRANS_COUNT register

pub fn ch1_al3_read_addr_trig(self) -> Reg<u32, R>

Alias for channel 1 READ_ADDR register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch2_read_addr(self) -> Reg<u32, RW>

DMA Channel 2 Read Address pointer This register updates automatically each time a read completes. The current value is the next address to be read by this channel.

pub fn ch2_write_addr(self) -> Reg<u32, RW>

DMA Channel 2 Write Address pointer This register updates automatically each time a write completes. The current value is the next address to be written by this channel.

pub fn ch2_trans_count(self) -> Reg<u32, RW>

DMA Channel 2 Transfer Count Program the number of bus transfers a channel will perform before halting. Note that, if transfers are larger than one byte in size, this is not equal to the number of bytes transferred (see CTRL_DATA_SIZE). When the channel is active, reading this register shows the number of transfers remaining, updating automatically each time a write transfer completes. Writing this register sets the RELOAD value for the transfer counter. Each time this channel is triggered, the RELOAD value is copied into the live transfer counter. The channel can be started multiple times, and will perform the same number of transfers each time, as programmed by most recent write. The RELOAD value can be observed at CHx_DBG_TCR. If TRANS_COUNT is used as a trigger, the written value is used immediately as the length of the new transfer sequence, as well as being written to RELOAD.

pub fn ch2_ctrl_trig(self) -> Reg<Ch2CtrlTrig, RW>

DMA Channel 2 Control and Status

pub fn ch2_al1_ctrl(self) -> Reg<u32, R>

Alias for channel 2 CTRL register

pub fn ch2_al1_read_addr(self) -> Reg<u32, R>

Alias for channel 2 READ_ADDR register

pub fn ch2_al1_write_addr(self) -> Reg<u32, R>

Alias for channel 2 WRITE_ADDR register

pub fn ch2_al1_trans_count_trig(self) -> Reg<u32, R>

Alias for channel 2 TRANS_COUNT register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch2_al2_ctrl(self) -> Reg<u32, R>

Alias for channel 2 CTRL register

pub fn ch2_al2_trans_count(self) -> Reg<u32, R>

Alias for channel 2 TRANS_COUNT register

pub fn ch2_al2_read_addr(self) -> Reg<u32, R>

Alias for channel 2 READ_ADDR register

pub fn ch2_al2_write_addr_trig(self) -> Reg<u32, R>

Alias for channel 2 WRITE_ADDR register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch2_al3_ctrl(self) -> Reg<u32, R>

Alias for channel 2 CTRL register

pub fn ch2_al3_write_addr(self) -> Reg<u32, R>

Alias for channel 2 WRITE_ADDR register

pub fn ch2_al3_trans_count(self) -> Reg<u32, R>

Alias for channel 2 TRANS_COUNT register

pub fn ch2_al3_read_addr_trig(self) -> Reg<u32, R>

Alias for channel 2 READ_ADDR register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch3_read_addr(self) -> Reg<u32, RW>

DMA Channel 3 Read Address pointer This register updates automatically each time a read completes. The current value is the next address to be read by this channel.

pub fn ch3_write_addr(self) -> Reg<u32, RW>

DMA Channel 3 Write Address pointer This register updates automatically each time a write completes. The current value is the next address to be written by this channel.

pub fn ch3_trans_count(self) -> Reg<u32, RW>

DMA Channel 3 Transfer Count Program the number of bus transfers a channel will perform before halting. Note that, if transfers are larger than one byte in size, this is not equal to the number of bytes transferred (see CTRL_DATA_SIZE). When the channel is active, reading this register shows the number of transfers remaining, updating automatically each time a write transfer completes. Writing this register sets the RELOAD value for the transfer counter. Each time this channel is triggered, the RELOAD value is copied into the live transfer counter. The channel can be started multiple times, and will perform the same number of transfers each time, as programmed by most recent write. The RELOAD value can be observed at CHx_DBG_TCR. If TRANS_COUNT is used as a trigger, the written value is used immediately as the length of the new transfer sequence, as well as being written to RELOAD.

pub fn ch3_ctrl_trig(self) -> Reg<Ch3CtrlTrig, RW>

DMA Channel 3 Control and Status

pub fn ch3_al1_ctrl(self) -> Reg<u32, R>

Alias for channel 3 CTRL register

pub fn ch3_al1_read_addr(self) -> Reg<u32, R>

Alias for channel 3 READ_ADDR register

pub fn ch3_al1_write_addr(self) -> Reg<u32, R>

Alias for channel 3 WRITE_ADDR register

pub fn ch3_al1_trans_count_trig(self) -> Reg<u32, R>

Alias for channel 3 TRANS_COUNT register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch3_al2_ctrl(self) -> Reg<u32, R>

Alias for channel 3 CTRL register

pub fn ch3_al2_trans_count(self) -> Reg<u32, R>

Alias for channel 3 TRANS_COUNT register

pub fn ch3_al2_read_addr(self) -> Reg<u32, R>

Alias for channel 3 READ_ADDR register

pub fn ch3_al2_write_addr_trig(self) -> Reg<u32, R>

Alias for channel 3 WRITE_ADDR register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch3_al3_ctrl(self) -> Reg<u32, R>

Alias for channel 3 CTRL register

pub fn ch3_al3_write_addr(self) -> Reg<u32, R>

Alias for channel 3 WRITE_ADDR register

pub fn ch3_al3_trans_count(self) -> Reg<u32, R>

Alias for channel 3 TRANS_COUNT register

pub fn ch3_al3_read_addr_trig(self) -> Reg<u32, R>

Alias for channel 3 READ_ADDR register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch4_read_addr(self) -> Reg<u32, RW>

DMA Channel 4 Read Address pointer This register updates automatically each time a read completes. The current value is the next address to be read by this channel.

pub fn ch4_write_addr(self) -> Reg<u32, RW>

DMA Channel 4 Write Address pointer This register updates automatically each time a write completes. The current value is the next address to be written by this channel.

pub fn ch4_trans_count(self) -> Reg<u32, RW>

DMA Channel 4 Transfer Count Program the number of bus transfers a channel will perform before halting. Note that, if transfers are larger than one byte in size, this is not equal to the number of bytes transferred (see CTRL_DATA_SIZE). When the channel is active, reading this register shows the number of transfers remaining, updating automatically each time a write transfer completes. Writing this register sets the RELOAD value for the transfer counter. Each time this channel is triggered, the RELOAD value is copied into the live transfer counter. The channel can be started multiple times, and will perform the same number of transfers each time, as programmed by most recent write. The RELOAD value can be observed at CHx_DBG_TCR. If TRANS_COUNT is used as a trigger, the written value is used immediately as the length of the new transfer sequence, as well as being written to RELOAD.

pub fn ch4_ctrl_trig(self) -> Reg<Ch4CtrlTrig, RW>

DMA Channel 4 Control and Status

pub fn ch4_al1_ctrl(self) -> Reg<u32, R>

Alias for channel 4 CTRL register

pub fn ch4_al1_read_addr(self) -> Reg<u32, R>

Alias for channel 4 READ_ADDR register

pub fn ch4_al1_write_addr(self) -> Reg<u32, R>

Alias for channel 4 WRITE_ADDR register

pub fn ch4_al1_trans_count_trig(self) -> Reg<u32, R>

Alias for channel 4 TRANS_COUNT register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch4_al2_ctrl(self) -> Reg<u32, R>

Alias for channel 4 CTRL register

pub fn ch4_al2_trans_count(self) -> Reg<u32, R>

Alias for channel 4 TRANS_COUNT register

pub fn ch4_al2_read_addr(self) -> Reg<u32, R>

Alias for channel 4 READ_ADDR register

pub fn ch4_al2_write_addr_trig(self) -> Reg<u32, R>

Alias for channel 4 WRITE_ADDR register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch4_al3_ctrl(self) -> Reg<u32, R>

Alias for channel 4 CTRL register

pub fn ch4_al3_write_addr(self) -> Reg<u32, R>

Alias for channel 4 WRITE_ADDR register

pub fn ch4_al3_trans_count(self) -> Reg<u32, R>

Alias for channel 4 TRANS_COUNT register

pub fn ch4_al3_read_addr_trig(self) -> Reg<u32, R>

Alias for channel 4 READ_ADDR register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch5_read_addr(self) -> Reg<u32, RW>

DMA Channel 5 Read Address pointer This register updates automatically each time a read completes. The current value is the next address to be read by this channel.

pub fn ch5_write_addr(self) -> Reg<u32, RW>

DMA Channel 5 Write Address pointer This register updates automatically each time a write completes. The current value is the next address to be written by this channel.

pub fn ch5_trans_count(self) -> Reg<u32, RW>

DMA Channel 5 Transfer Count Program the number of bus transfers a channel will perform before halting. Note that, if transfers are larger than one byte in size, this is not equal to the number of bytes transferred (see CTRL_DATA_SIZE). When the channel is active, reading this register shows the number of transfers remaining, updating automatically each time a write transfer completes. Writing this register sets the RELOAD value for the transfer counter. Each time this channel is triggered, the RELOAD value is copied into the live transfer counter. The channel can be started multiple times, and will perform the same number of transfers each time, as programmed by most recent write. The RELOAD value can be observed at CHx_DBG_TCR. If TRANS_COUNT is used as a trigger, the written value is used immediately as the length of the new transfer sequence, as well as being written to RELOAD.

pub fn ch5_ctrl_trig(self) -> Reg<Ch5CtrlTrig, RW>

DMA Channel 5 Control and Status

pub fn ch5_al1_ctrl(self) -> Reg<u32, R>

Alias for channel 5 CTRL register

pub fn ch5_al1_read_addr(self) -> Reg<u32, R>

Alias for channel 5 READ_ADDR register

pub fn ch5_al1_write_addr(self) -> Reg<u32, R>

Alias for channel 5 WRITE_ADDR register

pub fn ch5_al1_trans_count_trig(self) -> Reg<u32, R>

Alias for channel 5 TRANS_COUNT register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch5_al2_ctrl(self) -> Reg<u32, R>

Alias for channel 5 CTRL register

pub fn ch5_al2_trans_count(self) -> Reg<u32, R>

Alias for channel 5 TRANS_COUNT register

pub fn ch5_al2_read_addr(self) -> Reg<u32, R>

Alias for channel 5 READ_ADDR register

pub fn ch5_al2_write_addr_trig(self) -> Reg<u32, R>

Alias for channel 5 WRITE_ADDR register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch5_al3_ctrl(self) -> Reg<u32, R>

Alias for channel 5 CTRL register

pub fn ch5_al3_write_addr(self) -> Reg<u32, R>

Alias for channel 5 WRITE_ADDR register

pub fn ch5_al3_trans_count(self) -> Reg<u32, R>

Alias for channel 5 TRANS_COUNT register

pub fn ch5_al3_read_addr_trig(self) -> Reg<u32, R>

Alias for channel 5 READ_ADDR register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch6_read_addr(self) -> Reg<u32, RW>

DMA Channel 6 Read Address pointer This register updates automatically each time a read completes. The current value is the next address to be read by this channel.

pub fn ch6_write_addr(self) -> Reg<u32, RW>

DMA Channel 6 Write Address pointer This register updates automatically each time a write completes. The current value is the next address to be written by this channel.

pub fn ch6_trans_count(self) -> Reg<u32, RW>

DMA Channel 6 Transfer Count Program the number of bus transfers a channel will perform before halting. Note that, if transfers are larger than one byte in size, this is not equal to the number of bytes transferred (see CTRL_DATA_SIZE). When the channel is active, reading this register shows the number of transfers remaining, updating automatically each time a write transfer completes. Writing this register sets the RELOAD value for the transfer counter. Each time this channel is triggered, the RELOAD value is copied into the live transfer counter. The channel can be started multiple times, and will perform the same number of transfers each time, as programmed by most recent write. The RELOAD value can be observed at CHx_DBG_TCR. If TRANS_COUNT is used as a trigger, the written value is used immediately as the length of the new transfer sequence, as well as being written to RELOAD.

pub fn ch6_ctrl_trig(self) -> Reg<Ch6CtrlTrig, RW>

DMA Channel 6 Control and Status

pub fn ch6_al1_ctrl(self) -> Reg<u32, R>

Alias for channel 6 CTRL register

pub fn ch6_al1_read_addr(self) -> Reg<u32, R>

Alias for channel 6 READ_ADDR register

pub fn ch6_al1_write_addr(self) -> Reg<u32, R>

Alias for channel 6 WRITE_ADDR register

pub fn ch6_al1_trans_count_trig(self) -> Reg<u32, R>

Alias for channel 6 TRANS_COUNT register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch6_al2_ctrl(self) -> Reg<u32, R>

Alias for channel 6 CTRL register

pub fn ch6_al2_trans_count(self) -> Reg<u32, R>

Alias for channel 6 TRANS_COUNT register

pub fn ch6_al2_read_addr(self) -> Reg<u32, R>

Alias for channel 6 READ_ADDR register

pub fn ch6_al2_write_addr_trig(self) -> Reg<u32, R>

Alias for channel 6 WRITE_ADDR register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch6_al3_ctrl(self) -> Reg<u32, R>

Alias for channel 6 CTRL register

pub fn ch6_al3_write_addr(self) -> Reg<u32, R>

Alias for channel 6 WRITE_ADDR register

pub fn ch6_al3_trans_count(self) -> Reg<u32, R>

Alias for channel 6 TRANS_COUNT register

pub fn ch6_al3_read_addr_trig(self) -> Reg<u32, R>

Alias for channel 6 READ_ADDR register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch7_read_addr(self) -> Reg<u32, RW>

DMA Channel 7 Read Address pointer This register updates automatically each time a read completes. The current value is the next address to be read by this channel.

pub fn ch7_write_addr(self) -> Reg<u32, RW>

DMA Channel 7 Write Address pointer This register updates automatically each time a write completes. The current value is the next address to be written by this channel.

pub fn ch7_trans_count(self) -> Reg<u32, RW>

DMA Channel 7 Transfer Count Program the number of bus transfers a channel will perform before halting. Note that, if transfers are larger than one byte in size, this is not equal to the number of bytes transferred (see CTRL_DATA_SIZE). When the channel is active, reading this register shows the number of transfers remaining, updating automatically each time a write transfer completes. Writing this register sets the RELOAD value for the transfer counter. Each time this channel is triggered, the RELOAD value is copied into the live transfer counter. The channel can be started multiple times, and will perform the same number of transfers each time, as programmed by most recent write. The RELOAD value can be observed at CHx_DBG_TCR. If TRANS_COUNT is used as a trigger, the written value is used immediately as the length of the new transfer sequence, as well as being written to RELOAD.

pub fn ch7_ctrl_trig(self) -> Reg<Ch7CtrlTrig, RW>

DMA Channel 7 Control and Status

pub fn ch7_al1_ctrl(self) -> Reg<u32, R>

Alias for channel 7 CTRL register

pub fn ch7_al1_read_addr(self) -> Reg<u32, R>

Alias for channel 7 READ_ADDR register

pub fn ch7_al1_write_addr(self) -> Reg<u32, R>

Alias for channel 7 WRITE_ADDR register

pub fn ch7_al1_trans_count_trig(self) -> Reg<u32, R>

Alias for channel 7 TRANS_COUNT register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch7_al2_ctrl(self) -> Reg<u32, R>

Alias for channel 7 CTRL register

pub fn ch7_al2_trans_count(self) -> Reg<u32, R>

Alias for channel 7 TRANS_COUNT register

pub fn ch7_al2_read_addr(self) -> Reg<u32, R>

Alias for channel 7 READ_ADDR register

pub fn ch7_al2_write_addr_trig(self) -> Reg<u32, R>

Alias for channel 7 WRITE_ADDR register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch7_al3_ctrl(self) -> Reg<u32, R>

Alias for channel 7 CTRL register

pub fn ch7_al3_write_addr(self) -> Reg<u32, R>

Alias for channel 7 WRITE_ADDR register

pub fn ch7_al3_trans_count(self) -> Reg<u32, R>

Alias for channel 7 TRANS_COUNT register

pub fn ch7_al3_read_addr_trig(self) -> Reg<u32, R>

Alias for channel 7 READ_ADDR register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch8_read_addr(self) -> Reg<u32, RW>

DMA Channel 8 Read Address pointer This register updates automatically each time a read completes. The current value is the next address to be read by this channel.

pub fn ch8_write_addr(self) -> Reg<u32, RW>

DMA Channel 8 Write Address pointer This register updates automatically each time a write completes. The current value is the next address to be written by this channel.

pub fn ch8_trans_count(self) -> Reg<u32, RW>

DMA Channel 8 Transfer Count Program the number of bus transfers a channel will perform before halting. Note that, if transfers are larger than one byte in size, this is not equal to the number of bytes transferred (see CTRL_DATA_SIZE). When the channel is active, reading this register shows the number of transfers remaining, updating automatically each time a write transfer completes. Writing this register sets the RELOAD value for the transfer counter. Each time this channel is triggered, the RELOAD value is copied into the live transfer counter. The channel can be started multiple times, and will perform the same number of transfers each time, as programmed by most recent write. The RELOAD value can be observed at CHx_DBG_TCR. If TRANS_COUNT is used as a trigger, the written value is used immediately as the length of the new transfer sequence, as well as being written to RELOAD.

pub fn ch8_ctrl_trig(self) -> Reg<Ch8CtrlTrig, RW>

DMA Channel 8 Control and Status

pub fn ch8_al1_ctrl(self) -> Reg<u32, R>

Alias for channel 8 CTRL register

pub fn ch8_al1_read_addr(self) -> Reg<u32, R>

Alias for channel 8 READ_ADDR register

pub fn ch8_al1_write_addr(self) -> Reg<u32, R>

Alias for channel 8 WRITE_ADDR register

pub fn ch8_al1_trans_count_trig(self) -> Reg<u32, R>

Alias for channel 8 TRANS_COUNT register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch8_al2_ctrl(self) -> Reg<u32, R>

Alias for channel 8 CTRL register

pub fn ch8_al2_trans_count(self) -> Reg<u32, R>

Alias for channel 8 TRANS_COUNT register

pub fn ch8_al2_read_addr(self) -> Reg<u32, R>

Alias for channel 8 READ_ADDR register

pub fn ch8_al2_write_addr_trig(self) -> Reg<u32, R>

Alias for channel 8 WRITE_ADDR register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch8_al3_ctrl(self) -> Reg<u32, R>

Alias for channel 8 CTRL register

pub fn ch8_al3_write_addr(self) -> Reg<u32, R>

Alias for channel 8 WRITE_ADDR register

pub fn ch8_al3_trans_count(self) -> Reg<u32, R>

Alias for channel 8 TRANS_COUNT register

pub fn ch8_al3_read_addr_trig(self) -> Reg<u32, R>

Alias for channel 8 READ_ADDR register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch9_read_addr(self) -> Reg<u32, RW>

DMA Channel 9 Read Address pointer This register updates automatically each time a read completes. The current value is the next address to be read by this channel.

pub fn ch9_write_addr(self) -> Reg<u32, RW>

DMA Channel 9 Write Address pointer This register updates automatically each time a write completes. The current value is the next address to be written by this channel.

pub fn ch9_trans_count(self) -> Reg<u32, RW>

DMA Channel 9 Transfer Count Program the number of bus transfers a channel will perform before halting. Note that, if transfers are larger than one byte in size, this is not equal to the number of bytes transferred (see CTRL_DATA_SIZE). When the channel is active, reading this register shows the number of transfers remaining, updating automatically each time a write transfer completes. Writing this register sets the RELOAD value for the transfer counter. Each time this channel is triggered, the RELOAD value is copied into the live transfer counter. The channel can be started multiple times, and will perform the same number of transfers each time, as programmed by most recent write. The RELOAD value can be observed at CHx_DBG_TCR. If TRANS_COUNT is used as a trigger, the written value is used immediately as the length of the new transfer sequence, as well as being written to RELOAD.

pub fn ch9_ctrl_trig(self) -> Reg<Ch9CtrlTrig, RW>

DMA Channel 9 Control and Status

pub fn ch9_al1_ctrl(self) -> Reg<u32, R>

Alias for channel 9 CTRL register

pub fn ch9_al1_read_addr(self) -> Reg<u32, R>

Alias for channel 9 READ_ADDR register

pub fn ch9_al1_write_addr(self) -> Reg<u32, R>

Alias for channel 9 WRITE_ADDR register

pub fn ch9_al1_trans_count_trig(self) -> Reg<u32, R>

Alias for channel 9 TRANS_COUNT register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch9_al2_ctrl(self) -> Reg<u32, R>

Alias for channel 9 CTRL register

pub fn ch9_al2_trans_count(self) -> Reg<u32, R>

Alias for channel 9 TRANS_COUNT register

pub fn ch9_al2_read_addr(self) -> Reg<u32, R>

Alias for channel 9 READ_ADDR register

pub fn ch9_al2_write_addr_trig(self) -> Reg<u32, R>

Alias for channel 9 WRITE_ADDR register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch9_al3_ctrl(self) -> Reg<u32, R>

Alias for channel 9 CTRL register

pub fn ch9_al3_write_addr(self) -> Reg<u32, R>

Alias for channel 9 WRITE_ADDR register

pub fn ch9_al3_trans_count(self) -> Reg<u32, R>

Alias for channel 9 TRANS_COUNT register

pub fn ch9_al3_read_addr_trig(self) -> Reg<u32, R>

Alias for channel 9 READ_ADDR register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch10_read_addr(self) -> Reg<u32, RW>

DMA Channel 10 Read Address pointer This register updates automatically each time a read completes. The current value is the next address to be read by this channel.

pub fn ch10_write_addr(self) -> Reg<u32, RW>

DMA Channel 10 Write Address pointer This register updates automatically each time a write completes. The current value is the next address to be written by this channel.

pub fn ch10_trans_count(self) -> Reg<u32, RW>

DMA Channel 10 Transfer Count Program the number of bus transfers a channel will perform before halting. Note that, if transfers are larger than one byte in size, this is not equal to the number of bytes transferred (see CTRL_DATA_SIZE). When the channel is active, reading this register shows the number of transfers remaining, updating automatically each time a write transfer completes. Writing this register sets the RELOAD value for the transfer counter. Each time this channel is triggered, the RELOAD value is copied into the live transfer counter. The channel can be started multiple times, and will perform the same number of transfers each time, as programmed by most recent write. The RELOAD value can be observed at CHx_DBG_TCR. If TRANS_COUNT is used as a trigger, the written value is used immediately as the length of the new transfer sequence, as well as being written to RELOAD.

pub fn ch10_ctrl_trig(self) -> Reg<Ch10CtrlTrig, RW>

DMA Channel 10 Control and Status

pub fn ch10_al1_ctrl(self) -> Reg<u32, R>

Alias for channel 10 CTRL register

pub fn ch10_al1_read_addr(self) -> Reg<u32, R>

Alias for channel 10 READ_ADDR register

pub fn ch10_al1_write_addr(self) -> Reg<u32, R>

Alias for channel 10 WRITE_ADDR register

pub fn ch10_al1_trans_count_trig(self) -> Reg<u32, R>

Alias for channel 10 TRANS_COUNT register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch10_al2_ctrl(self) -> Reg<u32, R>

Alias for channel 10 CTRL register

pub fn ch10_al2_trans_count(self) -> Reg<u32, R>

Alias for channel 10 TRANS_COUNT register

pub fn ch10_al2_read_addr(self) -> Reg<u32, R>

Alias for channel 10 READ_ADDR register

pub fn ch10_al2_write_addr_trig(self) -> Reg<u32, R>

Alias for channel 10 WRITE_ADDR register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch10_al3_ctrl(self) -> Reg<u32, R>

Alias for channel 10 CTRL register

pub fn ch10_al3_write_addr(self) -> Reg<u32, R>

Alias for channel 10 WRITE_ADDR register

pub fn ch10_al3_trans_count(self) -> Reg<u32, R>

Alias for channel 10 TRANS_COUNT register

pub fn ch10_al3_read_addr_trig(self) -> Reg<u32, R>

Alias for channel 10 READ_ADDR register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch11_read_addr(self) -> Reg<u32, RW>

DMA Channel 11 Read Address pointer This register updates automatically each time a read completes. The current value is the next address to be read by this channel.

pub fn ch11_write_addr(self) -> Reg<u32, RW>

DMA Channel 11 Write Address pointer This register updates automatically each time a write completes. The current value is the next address to be written by this channel.

pub fn ch11_trans_count(self) -> Reg<u32, RW>

DMA Channel 11 Transfer Count Program the number of bus transfers a channel will perform before halting. Note that, if transfers are larger than one byte in size, this is not equal to the number of bytes transferred (see CTRL_DATA_SIZE). When the channel is active, reading this register shows the number of transfers remaining, updating automatically each time a write transfer completes. Writing this register sets the RELOAD value for the transfer counter. Each time this channel is triggered, the RELOAD value is copied into the live transfer counter. The channel can be started multiple times, and will perform the same number of transfers each time, as programmed by most recent write. The RELOAD value can be observed at CHx_DBG_TCR. If TRANS_COUNT is used as a trigger, the written value is used immediately as the length of the new transfer sequence, as well as being written to RELOAD.

pub fn ch11_ctrl_trig(self) -> Reg<Ch11CtrlTrig, RW>

DMA Channel 11 Control and Status

pub fn ch11_al1_ctrl(self) -> Reg<u32, R>

Alias for channel 11 CTRL register

pub fn ch11_al1_read_addr(self) -> Reg<u32, R>

Alias for channel 11 READ_ADDR register

pub fn ch11_al1_write_addr(self) -> Reg<u32, R>

Alias for channel 11 WRITE_ADDR register

pub fn ch11_al1_trans_count_trig(self) -> Reg<u32, R>

Alias for channel 11 TRANS_COUNT register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch11_al2_ctrl(self) -> Reg<u32, R>

Alias for channel 11 CTRL register

pub fn ch11_al2_trans_count(self) -> Reg<u32, R>

Alias for channel 11 TRANS_COUNT register

pub fn ch11_al2_read_addr(self) -> Reg<u32, R>

Alias for channel 11 READ_ADDR register

pub fn ch11_al2_write_addr_trig(self) -> Reg<u32, R>

Alias for channel 11 WRITE_ADDR register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn ch11_al3_ctrl(self) -> Reg<u32, R>

Alias for channel 11 CTRL register

pub fn ch11_al3_write_addr(self) -> Reg<u32, R>

Alias for channel 11 WRITE_ADDR register

pub fn ch11_al3_trans_count(self) -> Reg<u32, R>

Alias for channel 11 TRANS_COUNT register

pub fn ch11_al3_read_addr_trig(self) -> Reg<u32, R>

Alias for channel 11 READ_ADDR register This is a trigger register (0xc). Writing a nonzero value will reload the channel counter and start the channel.

pub fn intr(self) -> Reg<Intr, RW>

Interrupt Status (raw)

pub fn inte0(self) -> Reg<Inte0, RW>

Interrupt Enables for IRQ 0

pub fn intf0(self) -> Reg<Intf0, RW>

Force Interrupts

pub fn ints0(self) -> Reg<Ints0, RW>

Interrupt Status for IRQ 0

pub fn inte1(self) -> Reg<Inte1, RW>

Interrupt Enables for IRQ 1

pub fn intf1(self) -> Reg<Intf1, RW>

Force Interrupts for IRQ 1

pub fn ints1(self) -> Reg<Ints1, RW>

Interrupt Status (masked) for IRQ 1

pub fn timer0(self) -> Reg<Timer0, RW>

Pacing (X/Y) Fractional Timer The pacing timer produces TREQ assertions at a rate set by ((X/Y) * sys_clk). This equation is evaluated every sys_clk cycles and therefore can only generate TREQs at a rate of 1 per sys_clk (i.e. permanent TREQ) or less.

pub fn timer1(self) -> Reg<Timer1, RW>

Pacing (X/Y) Fractional Timer The pacing timer produces TREQ assertions at a rate set by ((X/Y) * sys_clk). This equation is evaluated every sys_clk cycles and therefore can only generate TREQs at a rate of 1 per sys_clk (i.e. permanent TREQ) or less.

pub fn multi_chan_trigger(self) -> Reg<MultiChanTrigger, RW>

Trigger one or more channels simultaneously

pub fn sniff_ctrl(self) -> Reg<SniffCtrl, RW>

Sniffer Control

pub fn sniff_data(self) -> Reg<u32, RW>

Data accumulator for sniff hardware Write an initial seed value here before starting a DMA transfer on the channel indicated by SNIFF_CTRL_DMACH. The hardware will update this register each time it observes a read from the indicated channel. Once the channel completes, the final result can be read from this register.

pub fn fifo_levels(self) -> Reg<FifoLevels, RW>

Debug RAF, WAF, TDF levels

pub fn chan_abort(self) -> Reg<ChanAbort, RW>

Abort an in-progress transfer sequence on one or more channels

pub fn n_channels(self) -> Reg<NChannels, RW>

The number of channels this DMA instance is equipped with. This DMA supports up to 16 hardware channels, but can be configured with as few as one, to minimise silicon area.

pub fn ch0_dbg_ctdreq(self) -> Reg<Ch0DbgCtdreq, RW>

Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

pub fn ch0_dbg_tcr(self) -> Reg<u32, R>

Read to get channel TRANS_COUNT reload value, i.e. the length of the next transfer

pub fn ch1_dbg_ctdreq(self) -> Reg<Ch1DbgCtdreq, RW>

Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

pub fn ch1_dbg_tcr(self) -> Reg<u32, R>

Read to get channel TRANS_COUNT reload value, i.e. the length of the next transfer

pub fn ch2_dbg_ctdreq(self) -> Reg<Ch2DbgCtdreq, RW>

Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

pub fn ch2_dbg_tcr(self) -> Reg<u32, R>

Read to get channel TRANS_COUNT reload value, i.e. the length of the next transfer

pub fn ch3_dbg_ctdreq(self) -> Reg<Ch3DbgCtdreq, RW>

Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

pub fn ch3_dbg_tcr(self) -> Reg<u32, R>

Read to get channel TRANS_COUNT reload value, i.e. the length of the next transfer

pub fn ch4_dbg_ctdreq(self) -> Reg<Ch4DbgCtdreq, RW>

Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

pub fn ch4_dbg_tcr(self) -> Reg<u32, R>

Read to get channel TRANS_COUNT reload value, i.e. the length of the next transfer

pub fn ch5_dbg_ctdreq(self) -> Reg<Ch5DbgCtdreq, RW>

Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

pub fn ch5_dbg_tcr(self) -> Reg<u32, R>

Read to get channel TRANS_COUNT reload value, i.e. the length of the next transfer

pub fn ch6_dbg_ctdreq(self) -> Reg<Ch6DbgCtdreq, RW>

Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

pub fn ch6_dbg_tcr(self) -> Reg<u32, R>

Read to get channel TRANS_COUNT reload value, i.e. the length of the next transfer

pub fn ch7_dbg_ctdreq(self) -> Reg<Ch7DbgCtdreq, RW>

Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

pub fn ch7_dbg_tcr(self) -> Reg<u32, R>

Read to get channel TRANS_COUNT reload value, i.e. the length of the next transfer

pub fn ch8_dbg_ctdreq(self) -> Reg<Ch8DbgCtdreq, RW>

Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

pub fn ch8_dbg_tcr(self) -> Reg<u32, R>

Read to get channel TRANS_COUNT reload value, i.e. the length of the next transfer

pub fn ch9_dbg_ctdreq(self) -> Reg<Ch9DbgCtdreq, RW>

Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

pub fn ch9_dbg_tcr(self) -> Reg<u32, R>

Read to get channel TRANS_COUNT reload value, i.e. the length of the next transfer

pub fn ch10_dbg_ctdreq(self) -> Reg<Ch10DbgCtdreq, RW>

Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

pub fn ch10_dbg_tcr(self) -> Reg<u32, R>

Read to get channel TRANS_COUNT reload value, i.e. the length of the next transfer

pub fn ch11_dbg_ctdreq(self) -> Reg<Ch11DbgCtdreq, RW>

Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

pub fn ch11_dbg_tcr(self) -> Reg<u32, R>

Read to get channel TRANS_COUNT reload value, i.e. the length of the next transfer