Struct rp2040_pac::dma::RegisterBlock

#[repr(C)]
pub struct RegisterBlock {

    pub ch: [CH; 12],
    pub intr: Reg<INTR_SPEC>,
    pub inte0: Reg<INTE0_SPEC>,
    pub intf0: Reg<INTF0_SPEC>,
    pub ints0: Reg<INTS0_SPEC>,
    pub inte1: Reg<INTE1_SPEC>,
    pub intf1: Reg<INTF1_SPEC>,
    pub ints1: Reg<INTS1_SPEC>,
    pub timer0: Reg<TIMER0_SPEC>,
    pub timer1: Reg<TIMER1_SPEC>,
    pub timer2: Reg<TIMER2_SPEC>,
    pub timer3: Reg<TIMER3_SPEC>,
    pub multi_chan_trigger: Reg<MULTI_CHAN_TRIGGER_SPEC>,
    pub sniff_ctrl: Reg<SNIFF_CTRL_SPEC>,
    pub sniff_data: Reg<SNIFF_DATA_SPEC>,
    pub fifo_levels: Reg<FIFO_LEVELS_SPEC>,
    pub chan_abort: Reg<CHAN_ABORT_SPEC>,
    pub n_channels: Reg<N_CHANNELS_SPEC>,
    pub ch0_dbg_ctdreq: Reg<CH0_DBG_CTDREQ_SPEC>,
    pub ch0_dbg_tcr: Reg<CH0_DBG_TCR_SPEC>,
    pub ch1_dbg_ctdreq: Reg<CH1_DBG_CTDREQ_SPEC>,
    pub ch1_dbg_tcr: Reg<CH1_DBG_TCR_SPEC>,
    pub ch2_dbg_ctdreq: Reg<CH2_DBG_CTDREQ_SPEC>,
    pub ch2_dbg_tcr: Reg<CH2_DBG_TCR_SPEC>,
    pub ch3_dbg_ctdreq: Reg<CH3_DBG_CTDREQ_SPEC>,
    pub ch3_dbg_tcr: Reg<CH3_DBG_TCR_SPEC>,
    pub ch4_dbg_ctdreq: Reg<CH4_DBG_CTDREQ_SPEC>,
    pub ch4_dbg_tcr: Reg<CH4_DBG_TCR_SPEC>,
    pub ch5_dbg_ctdreq: Reg<CH5_DBG_CTDREQ_SPEC>,
    pub ch5_dbg_tcr: Reg<CH5_DBG_TCR_SPEC>,
    pub ch6_dbg_ctdreq: Reg<CH6_DBG_CTDREQ_SPEC>,
    pub ch6_dbg_tcr: Reg<CH6_DBG_TCR_SPEC>,
    pub ch7_dbg_ctdreq: Reg<CH7_DBG_CTDREQ_SPEC>,
    pub ch7_dbg_tcr: Reg<CH7_DBG_TCR_SPEC>,
    pub ch8_dbg_ctdreq: Reg<CH8_DBG_CTDREQ_SPEC>,
    pub ch8_dbg_tcr: Reg<CH8_DBG_TCR_SPEC>,
    pub ch9_dbg_ctdreq: Reg<CH9_DBG_CTDREQ_SPEC>,
    pub ch9_dbg_tcr: Reg<CH9_DBG_TCR_SPEC>,
    pub ch10_dbg_ctdreq: Reg<CH10_DBG_CTDREQ_SPEC>,
    pub ch10_dbg_tcr: Reg<CH10_DBG_TCR_SPEC>,
    pub ch11_dbg_ctdreq: Reg<CH11_DBG_CTDREQ_SPEC>,
    pub ch11_dbg_tcr: Reg<CH11_DBG_TCR_SPEC>,
    /* private fields */

}

Register block

Fields
ch: [CH; 12]

0x00..0x300 - Cluster CH%s, containing CH?_READ_ADDR,CH??_READ_ADDR, CH?_WRITE_ADDR,CH??_WRITE_ADDR, CH?_TRANS_COUNT,CH??_TRANS_COUNT, CH?_CTRL_TRIG,CH??_CTRL_TRIG, CH?_AL1_CTRL,CH??_AL1_CTRL, CH?_AL1_READ_ADDR,CH??_AL1_READ_ADDR, CH?_AL1_WRITE_ADDR,CH??_AL1_WRITE_ADDR, CH?_AL1_TRANS_COUNT_TRIG,CH??_AL1_TRANS_COUNT_TRIG, CH?_AL2_CTRL,CH??_AL2_CTRL, CH?_AL2_TRANS_COUNT,CH??_AL2_TRANS_COUNT, CH?_AL2_READ_ADDR,CH??_AL2_READ_ADDR, CH?_AL2_WRITE_ADDR_TRIG,CH??_AL2_WRITE_ADDR_TRIG, CH?_AL3_CTRL,CH??_AL3_CTRL, CH?_AL3_WRITE_ADDR,CH??_AL3_WRITE_ADDR, CH?_AL3_TRANS_COUNT,CH??_AL3_TRANS_COUNT, CH?_AL3_READ_ADDR_TRIG,CH??_AL3_READ_ADDR_TRIG

intr: Reg<INTR_SPEC>

0x400 - Interrupt Status (raw)

inte0: Reg<INTE0_SPEC>

0x404 - Interrupt Enables for IRQ 0

intf0: Reg<INTF0_SPEC>

0x408 - Force Interrupts

ints0: Reg<INTS0_SPEC>

0x40c - Interrupt Status for IRQ 0

inte1: Reg<INTE1_SPEC>

0x414 - Interrupt Enables for IRQ 1

intf1: Reg<INTF1_SPEC>

0x418 - Force Interrupts for IRQ 1

ints1: Reg<INTS1_SPEC>

0x41c - Interrupt Status (masked) for IRQ 1

timer0: Reg<TIMER0_SPEC>

0x420 - Pacing (X/Y) Fractional Timer
The pacing timer produces TREQ assertions at a rate set by ((X/Y) * sys_clk). This equation is evaluated every sys_clk cycles and therefore can only generate TREQs at a rate of 1 per sys_clk (i.e. permanent TREQ) or less.

timer1: Reg<TIMER1_SPEC>

0x424 - Pacing (X/Y) Fractional Timer
The pacing timer produces TREQ assertions at a rate set by ((X/Y) * sys_clk). This equation is evaluated every sys_clk cycles and therefore can only generate TREQs at a rate of 1 per sys_clk (i.e. permanent TREQ) or less.

timer2: Reg<TIMER2_SPEC>

0x428 - Pacing (X/Y) Fractional Timer
The pacing timer produces TREQ assertions at a rate set by ((X/Y) * sys_clk). This equation is evaluated every sys_clk cycles and therefore can only generate TREQs at a rate of 1 per sys_clk (i.e. permanent TREQ) or less.

timer3: Reg<TIMER3_SPEC>

0x42c - Pacing (X/Y) Fractional Timer
The pacing timer produces TREQ assertions at a rate set by ((X/Y) * sys_clk). This equation is evaluated every sys_clk cycles and therefore can only generate TREQs at a rate of 1 per sys_clk (i.e. permanent TREQ) or less.

multi_chan_trigger: Reg<MULTI_CHAN_TRIGGER_SPEC>

0x430 - Trigger one or more channels simultaneously

sniff_ctrl: Reg<SNIFF_CTRL_SPEC>

0x434 - Sniffer Control

sniff_data: Reg<SNIFF_DATA_SPEC>

0x438 - Data accumulator for sniff hardware
Write an initial seed value here before starting a DMA transfer on the channel indicated by SNIFF_CTRL_DMACH. The hardware will update this register each time it observes a read from the indicated channel. Once the channel completes, the final result can be read from this register.

fifo_levels: Reg<FIFO_LEVELS_SPEC>

0x440 - Debug RAF, WAF, TDF levels

chan_abort: Reg<CHAN_ABORT_SPEC>

0x444 - Abort an in-progress transfer sequence on one or more channels

n_channels: Reg<N_CHANNELS_SPEC>

0x448 - The number of channels this DMA instance is equipped with. This DMA supports up to 16 hardware channels, but can be configured with as few as one, to minimise silicon area.

ch0_dbg_ctdreq: Reg<CH0_DBG_CTDREQ_SPEC>

0x800 - Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

ch0_dbg_tcr: Reg<CH0_DBG_TCR_SPEC>

0x804 - Read to get channel TRANS_COUNT reload value, i.e. the length of the next transfer

ch1_dbg_ctdreq: Reg<CH1_DBG_CTDREQ_SPEC>

0x840 - Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

ch1_dbg_tcr: Reg<CH1_DBG_TCR_SPEC>

0x844 - Read to get channel TRANS_COUNT reload value, i.e. the length of the next transfer

ch2_dbg_ctdreq: Reg<CH2_DBG_CTDREQ_SPEC>

0x880 - Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

ch2_dbg_tcr: Reg<CH2_DBG_TCR_SPEC>

0x884 - Read to get channel TRANS_COUNT reload value, i.e. the length of the next transfer

ch3_dbg_ctdreq: Reg<CH3_DBG_CTDREQ_SPEC>

0x8c0 - Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

ch3_dbg_tcr: Reg<CH3_DBG_TCR_SPEC>

0x8c4 - Read to get channel TRANS_COUNT reload value, i.e. the length of the next transfer

ch4_dbg_ctdreq: Reg<CH4_DBG_CTDREQ_SPEC>

0x900 - Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

ch4_dbg_tcr: Reg<CH4_DBG_TCR_SPEC>

0x904 - Read to get channel TRANS_COUNT reload value, i.e. the length of the next transfer

ch5_dbg_ctdreq: Reg<CH5_DBG_CTDREQ_SPEC>

0x940 - Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

ch5_dbg_tcr: Reg<CH5_DBG_TCR_SPEC>

0x944 - Read to get channel TRANS_COUNT reload value, i.e. the length of the next transfer

ch6_dbg_ctdreq: Reg<CH6_DBG_CTDREQ_SPEC>

0x980 - Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

ch6_dbg_tcr: Reg<CH6_DBG_TCR_SPEC>

0x984 - Read to get channel TRANS_COUNT reload value, i.e. the length of the next transfer

ch7_dbg_ctdreq: Reg<CH7_DBG_CTDREQ_SPEC>

0x9c0 - Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

ch7_dbg_tcr: Reg<CH7_DBG_TCR_SPEC>

0x9c4 - Read to get channel TRANS_COUNT reload value, i.e. the length of the next transfer

ch8_dbg_ctdreq: Reg<CH8_DBG_CTDREQ_SPEC>

0xa00 - Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

ch8_dbg_tcr: Reg<CH8_DBG_TCR_SPEC>

0xa04 - Read to get channel TRANS_COUNT reload value, i.e. the length of the next transfer

ch9_dbg_ctdreq: Reg<CH9_DBG_CTDREQ_SPEC>

0xa40 - Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

ch9_dbg_tcr: Reg<CH9_DBG_TCR_SPEC>

0xa44 - Read to get channel TRANS_COUNT reload value, i.e. the length of the next transfer

ch10_dbg_ctdreq: Reg<CH10_DBG_CTDREQ_SPEC>

0xa80 - Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

ch10_dbg_tcr: Reg<CH10_DBG_TCR_SPEC>

0xa84 - Read to get channel TRANS_COUNT reload value, i.e. the length of the next transfer

ch11_dbg_ctdreq: Reg<CH11_DBG_CTDREQ_SPEC>

0xac0 - Read: get channel DREQ counter (i.e. how many accesses the DMA expects it can perform on the peripheral without overflow/underflow. Write any value: clears the counter, and cause channel to re-initiate DREQ handshake.

ch11_dbg_tcr: Reg<CH11_DBG_TCR_SPEC>

0xac4 - Read to get channel TRANS_COUNT reload value, i.e. the length of the next transfer