rp2040_pac::dma

Type Alias SNIFF_DATA

pub type SNIFF_DATA = Reg<SNIFF_DATA_SPEC>;

SNIFF_DATA (rw) register accessor: Data accumulator for sniff hardware
Write an initial seed value here before starting a DMA transfer on the channel indicated by SNIFF_CTRL_DMACH. The hardware will update this register each time it observes a read from the indicated channel. Once the channel completes, the final result can be read from this register.

You can read this register and get sniff_data::R. You can reset, write, write_with_zero this register using sniff_data::W. You can also modify this register. See API.

For information about available fields see sniff_data module

Aliased Type

struct SNIFF_DATA { /* private fields */ }