rp2040_pac::generic

Trait Writable

pub trait Writable: RegisterSpec {
    const ZERO_TO_MODIFY_FIELDS_BITMAP: Self::Ux;
    const ONE_TO_MODIFY_FIELDS_BITMAP: Self::Ux;
}

Trait implemented by writeable registers.

This enables the write, write_with_zero and reset methods.

Registers marked with Readable can be also be modify’ed.

Required Associated Constants

const ZERO_TO_MODIFY_FIELDS_BITMAP: Self::Ux

Specifies the register bits that are not changed if you pass 1 and are changed if you pass 0

const ONE_TO_MODIFY_FIELDS_BITMAP: Self::Ux

Specifies the register bits that are not changed if you pass 0 and are changed if you pass 1

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.