rp2040_pac::i2c0::ic_con

Type Alias IC_RESTART_EN_W

pub type IC_RESTART_EN_W<'a, REG> = BitWriter<'a, REG, IC_RESTART_EN_A>;

Field IC_RESTART_EN writer - Determines whether RESTART conditions may be sent when acting as a master. Some older slaves do not support handling RESTART conditions; however, RESTART conditions are used in several DW_apb_i2c operations. When RESTART is disabled, the master is prohibited from performing the following functions: - Sending a START BYTE - Performing any high-speed mode operation - High-speed mode operation - Performing direction changes in combined format mode - Performing a read operation with a 10-bit address By replacing RESTART condition followed by a STOP and a subsequent START condition, split operations are broken down into multiple DW_apb_i2c transfers. If the above operations are performed, it will result in setting bit 6 (TX_ABRT) of the IC_RAW_INTR_STAT register.

Reset value: ENABLED

Aliased Type

struct IC_RESTART_EN_W<'a, REG> { /* private fields */ }

Implementations

impl<'a, REG> IC_RESTART_EN_W<'a, REG>
where REG: Writable + RegisterSpec,

pub fn disabled(self) -> &'a mut W<REG>

Master restart disabled

pub fn enabled(self) -> &'a mut W<REG>

Master restart enabled