rp2040_pac::i2c0::ic_con

Type Alias IC_SLAVE_DISABLE_R

pub type IC_SLAVE_DISABLE_R = BitReader<IC_SLAVE_DISABLE_A>;

Field IC_SLAVE_DISABLE reader - This bit controls whether I2C has its slave disabled, which means once the presetn signal is applied, then this bit is set and the slave is disabled.

If this bit is set (slave is disabled), DW_apb_i2c functions only as a master and does not perform any action that requires a slave.

NOTE: Software should ensure that if this bit is written with 0, then bit 0 should also be written with a 0.

Aliased Type

struct IC_SLAVE_DISABLE_R { /* private fields */ }

Implementations

impl IC_SLAVE_DISABLE_R

pub const fn variant(&self) -> IC_SLAVE_DISABLE_A

Get enumerated values variant

pub fn is_slave_enabled(&self) -> bool

Slave mode is enabled

pub fn is_slave_disabled(&self) -> bool

Slave mode is disabled