rp2040_pac::i2c0::ic_con
Type Alias STOP_DET_IFADDRESSED_W

pub type STOP_DET_IFADDRESSED_W<'a, REG> = BitWriter<'a, REG, STOP_DET_IFADDRESSED_A>;

Field STOP_DET_IFADDRESSED writer - In slave mode: - 1’b1: issues the STOP_DET interrupt only when it is addressed. - 1’b0: issues the STOP_DET irrespective of whether it’s addressed or not. Reset value: 0x0

NOTE: During a general call address, this slave does not issue the STOP_DET interrupt if STOP_DET_IF_ADDRESSED = 1’b1, even if the slave responds to the general call address by generating ACK. The STOP_DET interrupt is generated only when the transmitted address matches the slave address (SAR).

Aliased Type

struct STOP_DET_IFADDRESSED_W<'a, REG> { /* private fields */ }

Implementations

impl<'a, REG> STOP_DET_IFADDRESSED_W<'a, REG>
where REG: Writable + RegisterSpec,

pub fn disabled(self) -> &'a mut W<REG>

slave issues STOP_DET intr always

pub fn enabled(self) -> &'a mut W<REG>

slave issues STOP_DET intr only if addressed