rp2040_pac::i2c0::ic_data_cmd

Type Alias FIRST_DATA_BYTE_R

pub type FIRST_DATA_BYTE_R = BitReader<FIRST_DATA_BYTE_A>;

Field FIRST_DATA_BYTE reader - Indicates the first data byte received after the address phase for receive transfer in Master receiver or Slave receiver mode.

Reset value : 0x0

NOTE: In case of APB_DATA_WIDTH=8,

The user has to perform two APB Reads to IC_DATA_CMD in order to get status on 11 bit.

In order to read the 11 bit, the user has to perform the first data byte read [7:0] (offset 0x10) and then perform the second read [15:8] (offset 0x11) in order to know the status of 11 bit (whether the data received in previous read is a first data byte or not).

The 11th bit is an optional read field, user can ignore 2nd byte read [15:8] (offset 0x11) if not interested in FIRST_DATA_BYTE status.

Aliased Type

struct FIRST_DATA_BYTE_R { /* private fields */ }

Implementations

impl FIRST_DATA_BYTE_R

pub const fn variant(&self) -> FIRST_DATA_BYTE_A

Get enumerated values variant

pub fn is_inactive(&self) -> bool

Sequential data byte received

pub fn is_active(&self) -> bool

Non sequential data byte received