rp2040_pac::i2c0::ic_data_cmd

Type Alias RESTART_R

pub type RESTART_R = BitReader<RESTART_A>;

Field RESTART reader - This bit controls whether a RESTART is issued before the byte is sent or received.

1 - If IC_RESTART_EN is 1, a RESTART is issued before the data is sent/received (according to the value of CMD), regardless of whether or not the transfer direction is changing from the previous command; if IC_RESTART_EN is 0, a STOP followed by a START is issued instead.

0 - If IC_RESTART_EN is 1, a RESTART is issued only if the transfer direction is changing from the previous command; if IC_RESTART_EN is 0, a STOP followed by a START is issued instead.

Reset value: 0x0

Aliased Type

struct RESTART_R { /* private fields */ }

Implementations

impl RESTART_R

pub const fn variant(&self) -> RESTART_A

Get enumerated values variant

pub fn is_disable(&self) -> bool

Don’t Issue RESTART before this command

pub fn is_enable(&self) -> bool

Issue RESTART before this command