rp2040_pac::i2c0::ic_data_cmd

Type Alias STOP_R

pub type STOP_R = BitReader<STOP_A>;

Field STOP reader - This bit controls whether a STOP is issued after the byte is sent or received.

1 - STOP is issued after this byte, regardless of whether or not the Tx FIFO is empty. If the Tx FIFO is not empty, the master immediately tries to start a new transfer by issuing a START and arbitrating for the bus. - 0 - STOP is not issued after this byte, regardless of whether or not the Tx FIFO is empty. If the Tx FIFO is not empty, the master continues the current transfer by sending/receiving data bytes according to the value of the CMD bit. If the Tx FIFO is empty, the master holds the SCL line low and stalls the bus until a new command is available in the Tx FIFO. Reset value: 0x0

Aliased Type

struct STOP_R { /* private fields */ }

Implementations

impl STOP_R

pub const fn variant(&self) -> STOP_A

Get enumerated values variant

pub fn is_disable(&self) -> bool

Don’t Issue STOP after this command

pub fn is_enable(&self) -> bool

Issue STOP after this command