Module rp2040_pac::i2c0::ic_dma_tdlr

DMA Transmit Data Level Register

Structs

DMATDL_R

Field DMATDL reader - Transmit Data Level. This bit field controls the level at which a DMA request is made by the transmit logic. It is equal to the watermark level; that is, the dma_tx_req signal is generated when the number of valid data entries in the transmit FIFO is equal to or below this field value, and TDMAE = 1.

DMATDL_W

Field DMATDL writer - Transmit Data Level. This bit field controls the level at which a DMA request is made by the transmit logic. It is equal to the watermark level; that is, the dma_tx_req signal is generated when the number of valid data entries in the transmit FIFO is equal to or below this field value, and TDMAE = 1.

IC_DMA_TDLR_SPEC

DMA Transmit Data Level Register

R

Register IC_DMA_TDLR reader

W

Register IC_DMA_TDLR writer