rp2040_pac::i2c0::ic_enable

Type Alias ABORT_W

pub type ABORT_W<'a, REG> = BitWriter<'a, REG, ABORT_A>;

Field ABORT writer - When set, the controller initiates the transfer abort. - 0: ABORT not initiated or ABORT done - 1: ABORT operation in progress The software can abort the I2C transfer in master mode by setting this bit. The software can set this bit only when ENABLE is already set; otherwise, the controller ignores any write to ABORT bit. The software cannot clear the ABORT bit once set. In response to an ABORT, the controller issues a STOP and flushes the Tx FIFO after completing the current transfer, then sets the TX_ABORT interrupt after the abort operation. The ABORT bit is cleared automatically after the abort operation.

For a detailed description on how to abort I2C transfers, refer to ‘Aborting I2C Transfers’.

Reset value: 0x0

Aliased Type

struct ABORT_W<'a, REG> { /* private fields */ }

Implementations

impl<'a, REG> ABORT_W<'a, REG>
where REG: Writable + RegisterSpec,

pub fn disable(self) -> &'a mut W<REG>

ABORT operation not in progress

pub fn enabled(self) -> &'a mut W<REG>

ABORT operation in progress