rp2040_pac::i2c0::ic_enable_status

Type Alias IC_EN_R

pub type IC_EN_R = BitReader<IC_EN_A>;

Field IC_EN reader - ic_en Status. This bit always reflects the value driven on the output port ic_en. - When read as 1, DW_apb_i2c is deemed to be in an enabled state. - When read as 0, DW_apb_i2c is deemed completely inactive. Note: The CPU can safely read this bit anytime. When this bit is read as 0, the CPU can safely read SLV_RX_DATA_LOST (bit 2) and SLV_DISABLED_WHILE_BUSY (bit 1).

Reset value: 0x0

Aliased Type

struct IC_EN_R { /* private fields */ }

Implementations

impl IC_EN_R

pub const fn variant(&self) -> IC_EN_A

Get enumerated values variant

pub fn is_disabled(&self) -> bool

I2C disabled

pub fn is_enabled(&self) -> bool

I2C enabled