rp2040_pac::i2c0::ic_fs_spklen

Type Alias W

pub type W = W<IC_FS_SPKLEN_SPEC>;

Register IC_FS_SPKLEN writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn ic_fs_spklen(&mut self) -> IC_FS_SPKLEN_W<'_, IC_FS_SPKLEN_SPEC>

Bits 0:7 - This register must be set before any I2C bus transaction can take place to ensure stable operation. This register sets the duration, measured in ic_clk cycles, of the longest spike in the SCL or SDA lines that will be filtered out by the spike suppression logic. This register can be written only when the I2C interface is disabled which corresponds to the IC_ENABLE[0] register being set to 0. Writes at other times have no effect. The minimum valid value is 1; hardware prevents values less than this being written, and if attempted results in 1 being set. or more information, refer to ‘Spike Suppression’.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual