rp2040_pac::i2c0::ic_raw_intr_stat

Type Alias ACTIVITY_R

pub type ACTIVITY_R = BitReader<ACTIVITY_A>;

Field ACTIVITY reader - This bit captures DW_apb_i2c activity and stays set until it is cleared. There are four ways to clear it: - Disabling the DW_apb_i2c - Reading the IC_CLR_ACTIVITY register - Reading the IC_CLR_INTR register - System reset Once this bit is set, it stays set unless one of the four methods is used to clear it. Even if the DW_apb_i2c module is idle, this bit remains set until cleared, indicating that there was activity on the bus.

Reset value: 0x0

Aliased Type

struct ACTIVITY_R { /* private fields */ }

Implementations

impl ACTIVITY_R

pub const fn variant(&self) -> ACTIVITY_A

Get enumerated values variant

pub fn is_inactive(&self) -> bool

RAW_INTR_ACTIVITY interrupt is inactive

pub fn is_active(&self) -> bool

RAW_INTR_ACTIVITY interrupt is active