rp2040_pac::i2c0::ic_raw_intr_stat

Type Alias RX_OVER_R

pub type RX_OVER_R = BitReader<RX_OVER_A>;

Field RX_OVER reader - Set if the receive buffer is completely filled to IC_RX_BUFFER_DEPTH and an additional byte is received from an external I2C device. The DW_apb_i2c acknowledges this, but any data bytes received after the FIFO is full are lost. If the module is disabled (IC_ENABLE[0]=0), this bit keeps its level until the master or slave state machines go into idle, and when ic_en goes to 0, this interrupt is cleared.

Note: If bit 9 of the IC_CON register (RX_FIFO_FULL_HLD_CTRL) is programmed to HIGH, then the RX_OVER interrupt never occurs, because the Rx FIFO never overflows.

Reset value: 0x0

Aliased Type

struct RX_OVER_R { /* private fields */ }

Implementations

impl RX_OVER_R

pub const fn variant(&self) -> RX_OVER_A

Get enumerated values variant

pub fn is_inactive(&self) -> bool

RX_OVER interrupt is inactive

pub fn is_active(&self) -> bool

RX_OVER interrupt is active