Type Alias rp2040_pac::i2c0::ic_raw_intr_stat::TX_EMPTY_R

pub type TX_EMPTY_R = BitReader<TX_EMPTY_A>;

Field TX_EMPTY reader - The behavior of the TX_EMPTY interrupt status differs based on the TX_EMPTY_CTRL selection in the IC_CON register. - When TX_EMPTY_CTRL = 0: This bit is set to 1 when the transmit buffer is at or below the threshold value set in the IC_TX_TL register. - When TX_EMPTY_CTRL = 1: This bit is set to 1 when the transmit buffer is at or below the threshold value set in the IC_TX_TL register and the transmission of the address/data from the internal shift register for the most recently popped command is completed. It is automatically cleared by hardware when the buffer level goes above the threshold. When IC_ENABLE[0] is set to 0, the TX FIFO is flushed and held in reset. There the TX FIFO looks like it has no data within it, so this bit is set to 1, provided there is activity in the master or slave state machines. When there is no longer any activity, then with ic_en=0, this bit is set to 0.

Reset value: 0x0.

Aliased Type

struct TX_EMPTY_R { /* private fields */ }

Implementations

impl TX_EMPTY_R

pub const fn variant(&self) -> TX_EMPTY_A

Get enumerated values variant

pub fn is_inactive(&self) -> bool

TX_EMPTY interrupt is inactive

pub fn is_active(&self) -> bool

TX_EMPTY interrupt is active