rp2040_pac::i2c0::ic_rx_tl

Type Alias RX_TL_W

pub type RX_TL_W<'a, REG> = FieldWriter<'a, REG, 8>;

Field RX_TL writer - Receive FIFO Threshold Level.

Controls the level of entries (or above) that triggers the RX_FULL interrupt (bit 2 in IC_RAW_INTR_STAT register). The valid range is 0-255, with the additional restriction that hardware does not allow this value to be set to a value larger than the depth of the buffer. If an attempt is made to do that, the actual value set will be the maximum depth of the buffer. A value of 0 sets the threshold for 1 entry, and a value of 255 sets the threshold for 256 entries.

Aliased Type

struct RX_TL_W<'a, REG> { /* private fields */ }