Type Alias rp2040_pac::i2c0::ic_slv_data_nack_only::NACK_R

pub type NACK_R = BitReader<NACK_A>;

Field NACK reader - Generate NACK. This NACK generation only occurs when DW_apb_i2c is a slave-receiver. If this register is set to a value of 1, it can only generate a NACK after a data byte is received; hence, the data transfer is aborted and the data received is not pushed to the receive buffer.

When the register is set to a value of 0, it generates NACK/ACK, depending on normal criteria. - 1: generate NACK after data byte received - 0: generate NACK/ACK normally Reset value: 0x0

Aliased Type

struct NACK_R { /* private fields */ }

Implementations

impl NACK_R

pub const fn variant(&self) -> NACK_A

Get enumerated values variant

pub fn is_disabled(&self) -> bool

Slave receiver generates NACK normally

pub fn is_enabled(&self) -> bool

Slave receiver generates NACK upon data reception only