rp2040_pac::i2c0::ic_slv_data_nack_only

Type Alias NACK_W

pub type NACK_W<'a, REG> = BitWriter<'a, REG, NACK_A>;

Field NACK writer - Generate NACK. This NACK generation only occurs when DW_apb_i2c is a slave-receiver. If this register is set to a value of 1, it can only generate a NACK after a data byte is received; hence, the data transfer is aborted and the data received is not pushed to the receive buffer.

When the register is set to a value of 0, it generates NACK/ACK, depending on normal criteria. - 1: generate NACK after data byte received - 0: generate NACK/ACK normally Reset value: 0x0

Aliased Type

struct NACK_W<'a, REG> { /* private fields */ }

Implementations

impl<'a, REG> NACK_W<'a, REG>
where REG: Writable + RegisterSpec,

pub fn disabled(self) -> &'a mut W<REG>

Slave receiver generates NACK normally

pub fn enabled(self) -> &'a mut W<REG>

Slave receiver generates NACK upon data reception only