rp2040_pac::i2c0::ic_tx_abrt_source

Type Alias ABRT_HS_NORSTRT_R

pub type ABRT_HS_NORSTRT_R = BitReader<ABRT_HS_NORSTRT_A>;

Field ABRT_HS_NORSTRT reader - This field indicates that the restart is disabled (IC_RESTART_EN bit (IC_CON[5]) =0) and the user is trying to use the master to transfer data in High Speed mode.

Reset value: 0x0

Role of DW_apb_i2c: Master-Transmitter or Master-Receiver

Aliased Type

struct ABRT_HS_NORSTRT_R { /* private fields */ }

Implementations

impl ABRT_HS_NORSTRT_R

pub const fn variant(&self) -> ABRT_HS_NORSTRT_A

Get enumerated values variant

pub fn is_abrt_hs_norstrt_void(&self) -> bool

User trying to switch Master to HS mode when RESTART disabled- scenario not present

pub fn is_abrt_hs_norstrt_generated(&self) -> bool

User trying to switch Master to HS mode when RESTART disabled