Struct rp2040_pac::i2c0::RegisterBlock

#[repr(C)]
pub struct RegisterBlock {

    pub ic_con: Reg<IC_CON_SPEC>,
    pub ic_tar: Reg<IC_TAR_SPEC>,
    pub ic_sar: Reg<IC_SAR_SPEC>,
    pub ic_data_cmd: Reg<IC_DATA_CMD_SPEC>,
    pub ic_ss_scl_hcnt: Reg<IC_SS_SCL_HCNT_SPEC>,
    pub ic_ss_scl_lcnt: Reg<IC_SS_SCL_LCNT_SPEC>,
    pub ic_fs_scl_hcnt: Reg<IC_FS_SCL_HCNT_SPEC>,
    pub ic_fs_scl_lcnt: Reg<IC_FS_SCL_LCNT_SPEC>,
    pub ic_intr_stat: Reg<IC_INTR_STAT_SPEC>,
    pub ic_intr_mask: Reg<IC_INTR_MASK_SPEC>,
    pub ic_raw_intr_stat: Reg<IC_RAW_INTR_STAT_SPEC>,
    pub ic_rx_tl: Reg<IC_RX_TL_SPEC>,
    pub ic_tx_tl: Reg<IC_TX_TL_SPEC>,
    pub ic_clr_intr: Reg<IC_CLR_INTR_SPEC>,
    pub ic_clr_rx_under: Reg<IC_CLR_RX_UNDER_SPEC>,
    pub ic_clr_rx_over: Reg<IC_CLR_RX_OVER_SPEC>,
    pub ic_clr_tx_over: Reg<IC_CLR_TX_OVER_SPEC>,
    pub ic_clr_rd_req: Reg<IC_CLR_RD_REQ_SPEC>,
    pub ic_clr_tx_abrt: Reg<IC_CLR_TX_ABRT_SPEC>,
    pub ic_clr_rx_done: Reg<IC_CLR_RX_DONE_SPEC>,
    pub ic_clr_activity: Reg<IC_CLR_ACTIVITY_SPEC>,
    pub ic_clr_stop_det: Reg<IC_CLR_STOP_DET_SPEC>,
    pub ic_clr_start_det: Reg<IC_CLR_START_DET_SPEC>,
    pub ic_clr_gen_call: Reg<IC_CLR_GEN_CALL_SPEC>,
    pub ic_enable: Reg<IC_ENABLE_SPEC>,
    pub ic_status: Reg<IC_STATUS_SPEC>,
    pub ic_txflr: Reg<IC_TXFLR_SPEC>,
    pub ic_rxflr: Reg<IC_RXFLR_SPEC>,
    pub ic_sda_hold: Reg<IC_SDA_HOLD_SPEC>,
    pub ic_tx_abrt_source: Reg<IC_TX_ABRT_SOURCE_SPEC>,
    pub ic_slv_data_nack_only: Reg<IC_SLV_DATA_NACK_ONLY_SPEC>,
    pub ic_dma_cr: Reg<IC_DMA_CR_SPEC>,
    pub ic_dma_tdlr: Reg<IC_DMA_TDLR_SPEC>,
    pub ic_dma_rdlr: Reg<IC_DMA_RDLR_SPEC>,
    pub ic_sda_setup: Reg<IC_SDA_SETUP_SPEC>,
    pub ic_ack_general_call: Reg<IC_ACK_GENERAL_CALL_SPEC>,
    pub ic_enable_status: Reg<IC_ENABLE_STATUS_SPEC>,
    pub ic_fs_spklen: Reg<IC_FS_SPKLEN_SPEC>,
    pub ic_clr_restart_det: Reg<IC_CLR_RESTART_DET_SPEC>,
    pub ic_comp_param_1: Reg<IC_COMP_PARAM_1_SPEC>,
    pub ic_comp_version: Reg<IC_COMP_VERSION_SPEC>,
    pub ic_comp_type: Reg<IC_COMP_TYPE_SPEC>,
    /* private fields */

}

Register block

Fields
ic_con: Reg<IC_CON_SPEC>

0x00 - I2C Control Register. This register can be written only when the DW_apb_i2c is disabled, which corresponds to the IC_ENABLE[0] register being set to 0. Writes at other times have no effect.

Read/Write Access: - bit 10 is read only. - bit 11 is read only - bit 16 is read only - bit 17 is read only - bits 18 and 19 are read only.

ic_tar: Reg<IC_TAR_SPEC>

0x04 - I2C Target Address Register

This register is 12 bits wide, and bits 31:12 are reserved. This register can be written to only when IC_ENABLE[0] is set to 0.

Note: If the software or application is aware that the DW_apb_i2c is not using the TAR address for the pending commands in the Tx FIFO, then it is possible to update the TAR address even while the Tx FIFO has entries (IC_STATUS[2]= 0). - It is not necessary to perform any write to this register if DW_apb_i2c is enabled as an I2C slave only.

ic_sar: Reg<IC_SAR_SPEC>

0x08 - I2C Slave Address Register

ic_data_cmd: Reg<IC_DATA_CMD_SPEC>

0x10 - I2C Rx/Tx Data Buffer and Command Register; this is the register the CPU writes to when filling the TX FIFO and the CPU reads from when retrieving bytes from RX FIFO.

The size of the register changes as follows:

Write: - 11 bits when IC_EMPTYFIFO_HOLD_MASTER_EN=1 - 9 bits when IC_EMPTYFIFO_HOLD_MASTER_EN=0 Read: - 12 bits when IC_FIRST_DATA_BYTE_STATUS = 1 - 8 bits when IC_FIRST_DATA_BYTE_STATUS = 0 Note: In order for the DW_apb_i2c to continue acknowledging reads, a read command should be written for every byte that is to be received; otherwise the DW_apb_i2c will stop acknowledging.

ic_ss_scl_hcnt: Reg<IC_SS_SCL_HCNT_SPEC>

0x14 - Standard Speed I2C Clock SCL High Count Register

ic_ss_scl_lcnt: Reg<IC_SS_SCL_LCNT_SPEC>

0x18 - Standard Speed I2C Clock SCL Low Count Register

ic_fs_scl_hcnt: Reg<IC_FS_SCL_HCNT_SPEC>

0x1c - Fast Mode or Fast Mode Plus I2C Clock SCL High Count Register

ic_fs_scl_lcnt: Reg<IC_FS_SCL_LCNT_SPEC>

0x20 - Fast Mode or Fast Mode Plus I2C Clock SCL Low Count Register

ic_intr_stat: Reg<IC_INTR_STAT_SPEC>

0x2c - I2C Interrupt Status Register

Each bit in this register has a corresponding mask bit in the IC_INTR_MASK register. These bits are cleared by reading the matching interrupt clear register. The unmasked raw versions of these bits are available in the IC_RAW_INTR_STAT register.

ic_intr_mask: Reg<IC_INTR_MASK_SPEC>

0x30 - I2C Interrupt Mask Register.

These bits mask their corresponding interrupt status bits. This register is active low; a value of 0 masks the interrupt, whereas a value of 1 unmasks the interrupt.

ic_raw_intr_stat: Reg<IC_RAW_INTR_STAT_SPEC>

0x34 - I2C Raw Interrupt Status Register

Unlike the IC_INTR_STAT register, these bits are not masked so they always show the true status of the DW_apb_i2c.

ic_rx_tl: Reg<IC_RX_TL_SPEC>

0x38 - I2C Receive FIFO Threshold Register

ic_tx_tl: Reg<IC_TX_TL_SPEC>

0x3c - I2C Transmit FIFO Threshold Register

ic_clr_intr: Reg<IC_CLR_INTR_SPEC>

0x40 - Clear Combined and Individual Interrupt Register

ic_clr_rx_under: Reg<IC_CLR_RX_UNDER_SPEC>

0x44 - Clear RX_UNDER Interrupt Register

ic_clr_rx_over: Reg<IC_CLR_RX_OVER_SPEC>

0x48 - Clear RX_OVER Interrupt Register

ic_clr_tx_over: Reg<IC_CLR_TX_OVER_SPEC>

0x4c - Clear TX_OVER Interrupt Register

ic_clr_rd_req: Reg<IC_CLR_RD_REQ_SPEC>

0x50 - Clear RD_REQ Interrupt Register

ic_clr_tx_abrt: Reg<IC_CLR_TX_ABRT_SPEC>

0x54 - Clear TX_ABRT Interrupt Register

ic_clr_rx_done: Reg<IC_CLR_RX_DONE_SPEC>

0x58 - Clear RX_DONE Interrupt Register

ic_clr_activity: Reg<IC_CLR_ACTIVITY_SPEC>

0x5c - Clear ACTIVITY Interrupt Register

ic_clr_stop_det: Reg<IC_CLR_STOP_DET_SPEC>

0x60 - Clear STOP_DET Interrupt Register

ic_clr_start_det: Reg<IC_CLR_START_DET_SPEC>

0x64 - Clear START_DET Interrupt Register

ic_clr_gen_call: Reg<IC_CLR_GEN_CALL_SPEC>

0x68 - Clear GEN_CALL Interrupt Register

ic_enable: Reg<IC_ENABLE_SPEC>

0x6c - I2C Enable Register

ic_status: Reg<IC_STATUS_SPEC>

0x70 - I2C Status Register

This is a read-only register used to indicate the current transfer status and FIFO status. The status register may be read at any time. None of the bits in this register request an interrupt.

When the I2C is disabled by writing 0 in bit 0 of the IC_ENABLE register: - Bits 1 and 2 are set to 1 - Bits 3 and 10 are set to 0 When the master or slave state machines goes to idle and ic_en=0: - Bits 5 and 6 are set to 0

ic_txflr: Reg<IC_TXFLR_SPEC>

0x74 - I2C Transmit FIFO Level Register This register contains the number of valid data entries in the transmit FIFO buffer. It is cleared whenever: - The I2C is disabled - There is a transmit abort - that is, TX_ABRT bit is set in the IC_RAW_INTR_STAT register - The slave bulk transmit mode is aborted The register increments whenever data is placed into the transmit FIFO and decrements when data is taken from the transmit FIFO.

ic_rxflr: Reg<IC_RXFLR_SPEC>

0x78 - I2C Receive FIFO Level Register This register contains the number of valid data entries in the receive FIFO buffer. It is cleared whenever: - The I2C is disabled - Whenever there is a transmit abort caused by any of the events tracked in IC_TX_ABRT_SOURCE The register increments whenever data is placed into the receive FIFO and decrements when data is taken from the receive FIFO.

ic_sda_hold: Reg<IC_SDA_HOLD_SPEC>

0x7c - I2C SDA Hold Time Length Register

The bits [15:0] of this register are used to control the hold time of SDA during transmit in both slave and master mode (after SCL goes from HIGH to LOW).

The bits [23:16] of this register are used to extend the SDA transition (if any) whenever SCL is HIGH in the receiver in either master or slave mode.

Writes to this register succeed only when IC_ENABLE[0]=0.

The values in this register are in units of ic_clk period. The value programmed in IC_SDA_TX_HOLD must be greater than the minimum hold time in each mode (one cycle in master mode, seven cycles in slave mode) for the value to be implemented.

The programmed SDA hold time during transmit (IC_SDA_TX_HOLD) cannot exceed at any time the duration of the low part of scl. Therefore the programmed value cannot be larger than N_SCL_LOW-2, where N_SCL_LOW is the duration of the low part of the scl period measured in ic_clk cycles.

ic_tx_abrt_source: Reg<IC_TX_ABRT_SOURCE_SPEC>

0x80 - I2C Transmit Abort Source Register

This register has 32 bits that indicate the source of the TX_ABRT bit. Except for Bit 9, this register is cleared whenever the IC_CLR_TX_ABRT register or the IC_CLR_INTR register is read. To clear Bit 9, the source of the ABRT_SBYTE_NORSTRT must be fixed first; RESTART must be enabled (IC_CON[5]=1), the SPECIAL bit must be cleared (IC_TAR[11]), or the GC_OR_START bit must be cleared (IC_TAR[10]).

Once the source of the ABRT_SBYTE_NORSTRT is fixed, then this bit can be cleared in the same manner as other bits in this register. If the source of the ABRT_SBYTE_NORSTRT is not fixed before attempting to clear this bit, Bit 9 clears for one cycle and is then re-asserted.

ic_slv_data_nack_only: Reg<IC_SLV_DATA_NACK_ONLY_SPEC>

0x84 - Generate Slave Data NACK Register

The register is used to generate a NACK for the data part of a transfer when DW_apb_i2c is acting as a slave-receiver. This register only exists when the IC_SLV_DATA_NACK_ONLY parameter is set to 1. When this parameter disabled, this register does not exist and writing to the register’s address has no effect.

A write can occur on this register if both of the following conditions are met: - DW_apb_i2c is disabled (IC_ENABLE[0] = 0) - Slave part is inactive (IC_STATUS[6] = 0) Note: The IC_STATUS[6] is a register read-back location for the internal slv_activity signal; the user should poll this before writing the ic_slv_data_nack_only bit.

ic_dma_cr: Reg<IC_DMA_CR_SPEC>

0x88 - DMA Control Register

The register is used to enable the DMA Controller interface operation. There is a separate bit for transmit and receive. This can be programmed regardless of the state of IC_ENABLE.

ic_dma_tdlr: Reg<IC_DMA_TDLR_SPEC>

0x8c - DMA Transmit Data Level Register

ic_dma_rdlr: Reg<IC_DMA_RDLR_SPEC>

0x90 - I2C Receive Data Level Register

ic_sda_setup: Reg<IC_SDA_SETUP_SPEC>

0x94 - I2C SDA Setup Register

This register controls the amount of time delay (in terms of number of ic_clk clock periods) introduced in the rising edge of SCL - relative to SDA changing - when DW_apb_i2c services a read request in a slave-transmitter operation. The relevant I2C requirement is tSU:DAT (note 4) as detailed in the I2C Bus Specification. This register must be programmed with a value equal to or greater than 2.

Writes to this register succeed only when IC_ENABLE[0] = 0.

Note: The length of setup time is calculated using [(IC_SDA_SETUP - 1) * (ic_clk_period)], so if the user requires 10 ic_clk periods of setup time, they should program a value of 11. The IC_SDA_SETUP register is only used by the DW_apb_i2c when operating as a slave transmitter.

ic_ack_general_call: Reg<IC_ACK_GENERAL_CALL_SPEC>

0x98 - I2C ACK General Call Register

The register controls whether DW_apb_i2c responds with a ACK or NACK when it receives an I2C General Call address.

This register is applicable only when the DW_apb_i2c is in slave mode.

ic_enable_status: Reg<IC_ENABLE_STATUS_SPEC>

0x9c - I2C Enable Status Register

The register is used to report the DW_apb_i2c hardware status when the IC_ENABLE[0] register is set from 1 to 0; that is, when DW_apb_i2c is disabled.

If IC_ENABLE[0] has been set to 1, bits 2:1 are forced to 0, and bit 0 is forced to 1.

If IC_ENABLE[0] has been set to 0, bits 2:1 is only be valid as soon as bit 0 is read as ‘0’.

Note: When IC_ENABLE[0] has been set to 0, a delay occurs for bit 0 to be read as 0 because disabling the DW_apb_i2c depends on I2C bus activities.

ic_fs_spklen: Reg<IC_FS_SPKLEN_SPEC>

0xa0 - I2C SS, FS or FM+ spike suppression limit

This register is used to store the duration, measured in ic_clk cycles, of the longest spike that is filtered out by the spike suppression logic when the component is operating in SS, FS or FM+ modes. The relevant I2C requirement is tSP (table 4) as detailed in the I2C Bus Specification. This register must be programmed with a minimum value of 1.

ic_clr_restart_det: Reg<IC_CLR_RESTART_DET_SPEC>

0xa8 - Clear RESTART_DET Interrupt Register

ic_comp_param_1: Reg<IC_COMP_PARAM_1_SPEC>

0xf4 - Component Parameter Register 1

Note This register is not implemented and therefore reads as 0. If it was implemented it would be a constant read-only register that contains encoded information about the component’s parameter settings. Fields shown below are the settings for those parameters

ic_comp_version: Reg<IC_COMP_VERSION_SPEC>

0xf8 - I2C Component Version Register

ic_comp_type: Reg<IC_COMP_TYPE_SPEC>

0xfc - I2C Component Type Register