rp2040_pac::i2c0

Type Alias IC_STATUS

pub type IC_STATUS = Reg<IC_STATUS_SPEC>;

IC_STATUS (r) register accessor: I2C Status Register

This is a read-only register used to indicate the current transfer status and FIFO status. The status register may be read at any time. None of the bits in this register request an interrupt.

When the I2C is disabled by writing 0 in bit 0 of the IC_ENABLE register: - Bits 1 and 2 are set to 1 - Bits 3 and 10 are set to 0 When the master or slave state machines goes to idle and ic_en=0: - Bits 5 and 6 are set to 0

You can read this register and get ic_status::R. See API.

For information about available fields see ic_status module

Aliased Type

struct IC_STATUS { /* private fields */ }