rp2040_pac::i2c0

Type Alias IC_TAR

pub type IC_TAR = Reg<IC_TAR_SPEC>;

IC_TAR (rw) register accessor: I2C Target Address Register

This register is 12 bits wide, and bits 31:12 are reserved. This register can be written to only when IC_ENABLE[0] is set to 0.

Note: If the software or application is aware that the DW_apb_i2c is not using the TAR address for the pending commands in the Tx FIFO, then it is possible to update the TAR address even while the Tx FIFO has entries (IC_STATUS[2]= 0). - It is not necessary to perform any write to this register if DW_apb_i2c is enabled as an I2C slave only.

You can read this register and get ic_tar::R. You can reset, write, write_with_zero this register using ic_tar::W. You can also modify this register. See API.

For information about available fields see ic_tar module

Aliased Type

pub struct IC_TAR { /* private fields */ }