rp2040_pac::io_bank0::gpio::gpio_ctrl

Type Alias OEOVER_R

pub type OEOVER_R = FieldReader<OEOVER_A>;

Field OEOVER reader -

Aliased Type

struct OEOVER_R { /* private fields */ }

Implementations

impl OEOVER_R

pub const fn variant(&self) -> OEOVER_A

Get enumerated values variant

pub fn is_normal(&self) -> bool

drive output enable from peripheral signal selected by funcsel

pub fn is_invert(&self) -> bool

drive output enable from inverse of peripheral signal selected by funcsel

pub fn is_disable(&self) -> bool

disable output

pub fn is_enable(&self) -> bool

enable output