Module rp2040_pac::io_bank0::gpio::gpio_status

GPIO status

Structs

GPIO_STATUS_SPEC

GPIO status

Type Aliases

INFROMPAD_R

Field INFROMPAD reader - input signal from pad, before override is applied

INTOPERI_R

Field INTOPERI reader - input signal to peripheral, after override is applied

IRQFROMPAD_R

Field IRQFROMPAD reader - interrupt from pad before override is applied

IRQTOPROC_R

Field IRQTOPROC reader - interrupt to processors, after override is applied

OEFROMPERI_R

Field OEFROMPERI reader - output enable from selected peripheral, before register override is applied

OETOPAD_R

Field OETOPAD reader - output enable to pad after register override is applied

OUTFROMPERI_R

Field OUTFROMPERI reader - output signal from selected peripheral, before register override is applied

OUTTOPAD_R

Field OUTTOPAD reader - output signal to pad after register override is applied

R

Register GPIO_STATUS reader