Type Definition rp2040_pac::io_bank0::gpio26_ctrl::INOVER_R

type INOVER_R = R<u8, INOVER_A>;

Reader of field INOVER

Implementations

impl INOVER_R

pub fn variant(&self) -> INOVER_A

Get enumerated values variant

pub fn is_normal(&self) -> bool

Checks if the value of the field is NORMAL

pub fn is_invert(&self) -> bool

Checks if the value of the field is INVERT

pub fn is_low(&self) -> bool

Checks if the value of the field is LOW

pub fn is_high(&self) -> bool

Checks if the value of the field is HIGH