rp2040_pac::io_qspi::gpio_qspi::gpio_ctrl

Type Alias OEOVER_W

pub type OEOVER_W<'a, REG> = FieldWriterSafe<'a, REG, 2, OEOVER_A>;

Field OEOVER writer -

Aliased Type

struct OEOVER_W<'a, REG> { /* private fields */ }

Implementations

impl<'a, REG> OEOVER_W<'a, REG>
where REG: Writable + RegisterSpec, REG::Ux: From<u8>,

pub fn normal(self) -> &'a mut W<REG>

drive output enable from peripheral signal selected by funcsel

pub fn invert(self) -> &'a mut W<REG>

drive output enable from inverse of peripheral signal selected by funcsel

pub fn disable(self) -> &'a mut W<REG>

disable output

pub fn enable(self) -> &'a mut W<REG>

enable output