Type Alias rp2040_pac::pio0::fdebug::W

pub type W = W<FDEBUG_SPEC>;

Register FDEBUG writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn rxstall(&mut self) -> RXSTALL_W<'_, FDEBUG_SPEC>

Bits 0:3 - State machine has stalled on full RX FIFO during a blocking PUSH, or an IN with autopush enabled. This flag is also set when a nonblocking PUSH to a full FIFO took place, in which case the state machine has dropped data. Write 1 to clear.

pub fn rxunder(&mut self) -> RXUNDER_W<'_, FDEBUG_SPEC>

Bits 8:11 - RX FIFO underflow (i.e. read-on-empty by the system) has occurred. Write 1 to clear. Note that read-on-empty does not perturb the state of the FIFO in any way, but the data returned by reading from an empty FIFO is undefined, so this flag generally only becomes set due to some kind of software error.

pub fn txover(&mut self) -> TXOVER_W<'_, FDEBUG_SPEC>

Bits 16:19 - TX FIFO overflow (i.e. write-on-full by the system) has occurred. Write 1 to clear. Note that write-on-full does not alter the state or contents of the FIFO in any way, but the data that the system attempted to write is dropped, so if this flag is set, your software has quite likely dropped some data on the floor.

pub fn txstall(&mut self) -> TXSTALL_W<'_, FDEBUG_SPEC>

Bits 24:27 - State machine has stalled on empty TX FIFO during a blocking PULL, or an OUT with autopull enabled. Write 1 to clear.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual