Module rp2040_pac::pio0::sm::sm_pinctrl

State machine pin control

Structs

SM_PINCTRL_SPEC

State machine pin control

Type Aliases

IN_BASE_R

Field IN_BASE reader - The pin which is mapped to the least-significant bit of a state machine’s IN data bus. Higher-numbered pins are mapped to consecutively more-significant data bits, with a modulo of 32 applied to pin number.

IN_BASE_W

Field IN_BASE writer - The pin which is mapped to the least-significant bit of a state machine’s IN data bus. Higher-numbered pins are mapped to consecutively more-significant data bits, with a modulo of 32 applied to pin number.

OUT_BASE_R

Field OUT_BASE reader - The lowest-numbered pin that will be affected by an OUT PINS, OUT PINDIRS or MOV PINS instruction. The data written to this pin will always be the least-significant bit of the OUT or MOV data.

OUT_BASE_W

Field OUT_BASE writer - The lowest-numbered pin that will be affected by an OUT PINS, OUT PINDIRS or MOV PINS instruction. The data written to this pin will always be the least-significant bit of the OUT or MOV data.

OUT_COUNT_R

Field OUT_COUNT reader - The number of pins asserted by an OUT PINS, OUT PINDIRS or MOV PINS instruction. In the range 0 to 32 inclusive.

OUT_COUNT_W

Field OUT_COUNT writer - The number of pins asserted by an OUT PINS, OUT PINDIRS or MOV PINS instruction. In the range 0 to 32 inclusive.

R

Register SM_PINCTRL reader

SET_BASE_R

Field SET_BASE reader - The lowest-numbered pin that will be affected by a SET PINS or SET PINDIRS instruction. The data written to this pin is the least-significant bit of the SET data.

SET_BASE_W

Field SET_BASE writer - The lowest-numbered pin that will be affected by a SET PINS or SET PINDIRS instruction. The data written to this pin is the least-significant bit of the SET data.

SET_COUNT_R

Field SET_COUNT reader - The number of pins asserted by a SET. In the range 0 to 5 inclusive.

SET_COUNT_W

Field SET_COUNT writer - The number of pins asserted by a SET. In the range 0 to 5 inclusive.

SIDESET_BASE_R

Field SIDESET_BASE reader - The lowest-numbered pin that will be affected by a side-set operation. The MSBs of an instruction’s side-set/delay field (up to 5, determined by SIDESET_COUNT) are used for side-set data, with the remaining LSBs used for delay. The least-significant bit of the side-set portion is the bit written to this pin, with more-significant bits written to higher-numbered pins.

SIDESET_BASE_W

Field SIDESET_BASE writer - The lowest-numbered pin that will be affected by a side-set operation. The MSBs of an instruction’s side-set/delay field (up to 5, determined by SIDESET_COUNT) are used for side-set data, with the remaining LSBs used for delay. The least-significant bit of the side-set portion is the bit written to this pin, with more-significant bits written to higher-numbered pins.

SIDESET_COUNT_R

Field SIDESET_COUNT reader - The number of MSBs of the Delay/Side-set instruction field which are used for side-set. Inclusive of the enable bit, if present. Minimum of 0 (all delay bits, no side-set) and maximum of 5 (all side-set, no delay).

SIDESET_COUNT_W

Field SIDESET_COUNT writer - The number of MSBs of the Delay/Side-set instruction field which are used for side-set. Inclusive of the enable bit, if present. Minimum of 0 (all delay bits, no side-set) and maximum of 5 (all side-set, no delay).

W

Register SM_PINCTRL writer