rp2040_pac::pio0::sm
Module sm_shiftctrl

Control behaviour of the input/output shift registers for state machine 0

Structs

SM_SHIFTCTRL_SPEC

Control behaviour of the input/output shift registers for state machine 0

Type Aliases

AUTOPULL_R

Field AUTOPULL reader - Pull automatically when the output shift register is emptied, i.e. on or following an OUT instruction which causes the output shift counter to reach or exceed PULL_THRESH.

AUTOPULL_W

Field AUTOPULL writer - Pull automatically when the output shift register is emptied, i.e. on or following an OUT instruction which causes the output shift counter to reach or exceed PULL_THRESH.

AUTOPUSH_R

Field AUTOPUSH reader - Push automatically when the input shift register is filled, i.e. on an IN instruction which causes the input shift counter to reach or exceed PUSH_THRESH.

AUTOPUSH_W

Field AUTOPUSH writer - Push automatically when the input shift register is filled, i.e. on an IN instruction which causes the input shift counter to reach or exceed PUSH_THRESH.

FJOIN_RX_R

Field FJOIN_RX reader - When 1, RX FIFO steals the TX FIFO’s storage, and becomes twice as deep.
TX FIFO is disabled as a result (always reads as both full and empty).
FIFOs are flushed when this bit is changed.

FJOIN_RX_W

Field FJOIN_RX writer - When 1, RX FIFO steals the TX FIFO’s storage, and becomes twice as deep.
TX FIFO is disabled as a result (always reads as both full and empty).
FIFOs are flushed when this bit is changed.

FJOIN_TX_R

Field FJOIN_TX reader - When 1, TX FIFO steals the RX FIFO’s storage, and becomes twice as deep.
RX FIFO is disabled as a result (always reads as both full and empty).
FIFOs are flushed when this bit is changed.

FJOIN_TX_W

Field FJOIN_TX writer - When 1, TX FIFO steals the RX FIFO’s storage, and becomes twice as deep.
RX FIFO is disabled as a result (always reads as both full and empty).
FIFOs are flushed when this bit is changed.

IN_SHIFTDIR_R

Field IN_SHIFTDIR reader - 1 = shift input shift register to right (data enters from left). 0 = to left.

IN_SHIFTDIR_W

Field IN_SHIFTDIR writer - 1 = shift input shift register to right (data enters from left). 0 = to left.

OUT_SHIFTDIR_R

Field OUT_SHIFTDIR reader - 1 = shift out of output shift register to right. 0 = to left.

OUT_SHIFTDIR_W

Field OUT_SHIFTDIR writer - 1 = shift out of output shift register to right. 0 = to left.

PULL_THRESH_R

Field PULL_THRESH reader - Number of bits shifted out of OSR before autopull, or conditional pull (PULL IFEMPTY), will take place.
Write 0 for value of 32.

PULL_THRESH_W

Field PULL_THRESH writer - Number of bits shifted out of OSR before autopull, or conditional pull (PULL IFEMPTY), will take place.
Write 0 for value of 32.

PUSH_THRESH_R

Field PUSH_THRESH reader - Number of bits shifted into ISR before autopush, or conditional push (PUSH IFFULL), will take place.
Write 0 for value of 32.

PUSH_THRESH_W

Field PUSH_THRESH writer - Number of bits shifted into ISR before autopush, or conditional push (PUSH IFFULL), will take place.
Write 0 for value of 32.

R

Register SM_SHIFTCTRL reader

W

Register SM_SHIFTCTRL writer