rp2040_pac::pio0::sm::sm_shiftctrl

Type Alias R

pub type R = R<SM_SHIFTCTRL_SPEC>;

Register SM_SHIFTCTRL reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn autopush(&self) -> AUTOPUSH_R

Bit 16 - Push automatically when the input shift register is filled, i.e. on an IN instruction which causes the input shift counter to reach or exceed PUSH_THRESH.

pub fn autopull(&self) -> AUTOPULL_R

Bit 17 - Pull automatically when the output shift register is emptied, i.e. on or following an OUT instruction which causes the output shift counter to reach or exceed PULL_THRESH.

pub fn in_shiftdir(&self) -> IN_SHIFTDIR_R

Bit 18 - 1 = shift input shift register to right (data enters from left). 0 = to left.

pub fn out_shiftdir(&self) -> OUT_SHIFTDIR_R

Bit 19 - 1 = shift out of output shift register to right. 0 = to left.

pub fn push_thresh(&self) -> PUSH_THRESH_R

Bits 20:24 - Number of bits shifted into ISR before autopush, or conditional push (PUSH IFFULL), will take place.
Write 0 for value of 32.

pub fn pull_thresh(&self) -> PULL_THRESH_R

Bits 25:29 - Number of bits shifted out of OSR before autopull, or conditional pull (PULL IFEMPTY), will take place.
Write 0 for value of 32.

pub fn fjoin_tx(&self) -> FJOIN_TX_R

Bit 30 - When 1, TX FIFO steals the RX FIFO’s storage, and becomes twice as deep.
RX FIFO is disabled as a result (always reads as both full and empty).
FIFOs are flushed when this bit is changed.

pub fn fjoin_rx(&self) -> FJOIN_RX_R

Bit 31 - When 1, RX FIFO steals the TX FIFO’s storage, and becomes twice as deep.
TX FIFO is disabled as a result (always reads as both full and empty).
FIFOs are flushed when this bit is changed.