Struct rp2040_pac::pio0::RegisterBlock

#[repr(C)]
pub struct RegisterBlock {

    pub ctrl: Reg<CTRL_SPEC>,
    pub fstat: Reg<FSTAT_SPEC>,
    pub fdebug: Reg<FDEBUG_SPEC>,
    pub flevel: Reg<FLEVEL_SPEC>,
    pub txf: [Reg<TXF_SPEC>; 4],
    pub rxf: [Reg<RXF_SPEC>; 4],
    pub irq: Reg<IRQ_SPEC>,
    pub irq_force: Reg<IRQ_FORCE_SPEC>,
    pub input_sync_bypass: Reg<INPUT_SYNC_BYPASS_SPEC>,
    pub dbg_padout: Reg<DBG_PADOUT_SPEC>,
    pub dbg_padoe: Reg<DBG_PADOE_SPEC>,
    pub dbg_cfginfo: Reg<DBG_CFGINFO_SPEC>,
    pub instr_mem: [Reg<INSTR_MEM_SPEC>; 32],
    pub sm: [SM; 4],
    pub intr: Reg<INTR_SPEC>,
    pub sm_irq: [SM_IRQ; 2],

}

Register block

Fields
ctrl: Reg<CTRL_SPEC>

0x00 - PIO control register

fstat: Reg<FSTAT_SPEC>

0x04 - FIFO status register

fdebug: Reg<FDEBUG_SPEC>

0x08 - FIFO debug register

flevel: Reg<FLEVEL_SPEC>

0x0c - FIFO levels

txf: [Reg<TXF_SPEC>; 4]

0x10..0x20 - Direct write access to the TX FIFO for this state machine. Each write pushes one word to the FIFO. Attempting to write to a full FIFO has no effect on the FIFO state or contents, and sets the sticky FDEBUG_TXOVER error flag for this FIFO.

rxf: [Reg<RXF_SPEC>; 4]

0x20..0x30 - Direct read access to the RX FIFO for this state machine. Each read pops one word from the FIFO. Attempting to read from an empty FIFO has no effect on the FIFO state, and sets the sticky FDEBUG_RXUNDER error flag for this FIFO. The data returned to the system on a read from an empty FIFO is undefined.

irq: Reg<IRQ_SPEC>

0x30 - State machine IRQ flags register. Write 1 to clear. There are 8 state machine IRQ flags, which can be set, cleared, and waited on by the state machines. There’s no fixed association between flags and state machines – any state machine can use any flag.

Any of the 8 flags can be used for timing synchronisation between state machines, using IRQ and WAIT instructions. The lower four of these flags are also routed out to system-level interrupt requests, alongside FIFO status interrupts – see e.g. IRQ0_INTE.

irq_force: Reg<IRQ_FORCE_SPEC>

0x34 - Writing a 1 to each of these bits will forcibly assert the corresponding IRQ. Note this is different to the INTF register: writing here affects PIO internal state. INTF just asserts the processor-facing IRQ signal for testing ISRs, and is not visible to the state machines.

input_sync_bypass: Reg<INPUT_SYNC_BYPASS_SPEC>

0x38 - There is a 2-flipflop synchronizer on each GPIO input, which protects PIO logic from metastabilities. This increases input delay, and for fast synchronous IO (e.g. SPI) these synchronizers may need to be bypassed. Each bit in this register corresponds to one GPIO.
0 -> input is synchronized (default)
1 -> synchronizer is bypassed
If in doubt, leave this register as all zeroes.

dbg_padout: Reg<DBG_PADOUT_SPEC>

0x3c - Read to sample the pad output values PIO is currently driving to the GPIOs. On RP2040 there are 30 GPIOs, so the two most significant bits are hardwired to 0.

dbg_padoe: Reg<DBG_PADOE_SPEC>

0x40 - Read to sample the pad output enables (direction) PIO is currently driving to the GPIOs. On RP2040 there are 30 GPIOs, so the two most significant bits are hardwired to 0.

dbg_cfginfo: Reg<DBG_CFGINFO_SPEC>

0x44 - The PIO hardware has some free parameters that may vary between chip products.
These should be provided in the chip datasheet, but are also exposed here.

instr_mem: [Reg<INSTR_MEM_SPEC>; 32]

0x48..0xc8 - Write-only access to instruction memory location %s

sm: [SM; 4]

0xc8..0x128 - Cluster SM%s, containing SM*_CLKDIV, SM*_EXECCTRL, SM*_SHIFTCTRL, SM*_ADDR, SM*_INSTR, SM*_PINCTRL

intr: Reg<INTR_SPEC>

0x128 - Raw Interrupts

sm_irq: [SM_IRQ; 2]

0x12c..0x144 - Cluster SM_IRQ%s, containing IRQ*_INTE, IRQ*_INTF, IRQ*_INTS