Module rp2040_pac::ppb::aircr

Use the Application Interrupt and Reset Control Register to: determine data endianness, clear all active state information from debug halt mode, request a system reset.

Structs

AIRCR_SPEC

Use the Application Interrupt and Reset Control Register to: determine data endianness, clear all active state information from debug halt mode, request a system reset.

ENDIANESS_R

Field ENDIANESS reader - Data endianness implemented:
0 = Little-endian.

R

Register AIRCR reader

SYSRESETREQ_R

Field SYSRESETREQ reader - Writing 1 to this bit causes the SYSRESETREQ signal to the outer system to be asserted to request a reset. The intention is to force a large system reset of all major components except for debug. The C_HALT bit in the DHCSR is cleared as a result of the system reset requested. The debugger does not lose contact with the device.

SYSRESETREQ_W

Field SYSRESETREQ writer - Writing 1 to this bit causes the SYSRESETREQ signal to the outer system to be asserted to request a reset. The intention is to force a large system reset of all major components except for debug. The C_HALT bit in the DHCSR is cleared as a result of the system reset requested. The debugger does not lose contact with the device.

VECTCLRACTIVE_R

Field VECTCLRACTIVE reader - Clears all active state information for fixed and configurable exceptions. This bit: is self-clearing, can only be set by the DAP when the core is halted. When set: clears all active exception status of the processor, forces a return to Thread mode, forces an IPSR of 0. A debugger must re-initialize the stack.

VECTCLRACTIVE_W

Field VECTCLRACTIVE writer - Clears all active state information for fixed and configurable exceptions. This bit: is self-clearing, can only be set by the DAP when the core is halted. When set: clears all active exception status of the processor, forces a return to Thread mode, forces an IPSR of 0. A debugger must re-initialize the stack.

VECTKEY_R

Field VECTKEY reader - Register key:
Reads as Unknown
On writes, write 0x05FA to VECTKEY, otherwise the write is ignored.

VECTKEY_W

Field VECTKEY writer - Register key:
Reads as Unknown
On writes, write 0x05FA to VECTKEY, otherwise the write is ignored.

W

Register AIRCR writer