rp2040_pac::ppb
Module cpuid

Read the CPU ID Base Register to determine: the ID number of the processor core, the version number of the processor core, the implementation details of the processor core.

Structs

CPUID_SPEC

Read the CPU ID Base Register to determine: the ID number of the processor core, the version number of the processor core, the implementation details of the processor core.

Type Aliases

ARCHITECTURE_R

Field ARCHITECTURE reader - Constant that defines the architecture of the processor:
0xC = ARMv6-M architecture.

IMPLEMENTER_R

Field IMPLEMENTER reader - Implementor code: 0x41 = ARM

PARTNO_R

Field PARTNO reader - Number of processor within family: 0xC60 = Cortex-M0+

R

Register CPUID reader

REVISION_R

Field REVISION reader - Minor revision number m in the rnpm revision status:
0x1 = Patch 1.

VARIANT_R

Field VARIANT reader - Major revision number n in the rnpm revision status:
0x0 = Revision 0.