Type Alias rp2040_pac::ppb::cpuid::R

pub type R = R<CPUID_SPEC>;

Register CPUID reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn revision(&self) -> REVISION_R

Bits 0:3 - Minor revision number m in the rnpm revision status:
0x1 = Patch 1.

pub fn partno(&self) -> PARTNO_R

Bits 4:15 - Number of processor within family: 0xC60 = Cortex-M0+

pub fn architecture(&self) -> ARCHITECTURE_R

Bits 16:19 - Constant that defines the architecture of the processor:
0xC = ARMv6-M architecture.

pub fn variant(&self) -> VARIANT_R

Bits 20:23 - Major revision number n in the rnpm revision status:
0x0 = Revision 0.

pub fn implementer(&self) -> IMPLEMENTER_R

Bits 24:31 - Implementor code: 0x41 = ARM