Struct rp2040_pac::ppb::mpu_rbar::MPU_RBAR_SPEC

pub struct MPU_RBAR_SPEC;

Read the MPU Region Base Address Register to determine the base address of the region identified by MPU_RNR. Write to update the base address of said region or that of a specified region, with whose number MPU_RNR will also be updated.

This register you can read, write_with_zero, reset, write, modify. See API.

For information about available fields see mpu_rbar module