rp2040_pac::ppb::mpu_rbar

Type Alias R

pub type R = R<MPU_RBAR_SPEC>;

Register MPU_RBAR reader

Aliased Type

pub struct R { /* private fields */ }

Implementations

impl R

pub fn region(&self) -> REGION_R

Bits 0:3 - On writes, specifies the number of the region whose base address to update provided VALID is set written as 1. On reads, returns bits [3:0] of MPU_RNR.

pub fn valid(&self) -> VALID_R

Bit 4 - On writes, indicates whether the write must update the base address of the region identified by the REGION field, updating the MPU_RNR to indicate this new region.
Write:
0 = MPU_RNR not changed, and the processor:
Updates the base address for the region specified in the MPU_RNR.
Ignores the value of the REGION field.
1 = The processor:
Updates the value of the MPU_RNR to the value of the REGION field.
Updates the base address for the region specified in the REGION field.
Always reads as zero.

pub fn addr(&self) -> ADDR_R

Bits 8:31 - Base address of the region.