rp2040_pac::ppb::mpu_rbar

Type Alias VALID_W

pub type VALID_W<'a, REG> = BitWriter<'a, REG>;

Field VALID writer - On writes, indicates whether the write must update the base address of the region identified by the REGION field, updating the MPU_RNR to indicate this new region.
Write:
0 = MPU_RNR not changed, and the processor:
Updates the base address for the region specified in the MPU_RNR.
Ignores the value of the REGION field.
1 = The processor:
Updates the value of the MPU_RNR to the value of the REGION field.
Updates the base address for the region specified in the REGION field.
Always reads as zero.

Aliased Type

struct VALID_W<'a, REG> { /* private fields */ }