Module rp2040_pac::ppb::mpu_type

Read the MPU Type Register to determine if the processor implements an MPU, and how many regions the MPU supports.

Structs

DREGION_R

Field DREGION reader - Number of regions supported by the MPU.

IREGION_R

Field IREGION reader - Instruction region. Reads as zero as ARMv6-M only supports a unified MPU.

MPU_TYPE_SPEC

Read the MPU Type Register to determine if the processor implements an MPU, and how many regions the MPU supports.

R

Register MPU_TYPE reader

SEPARATE_R

Field SEPARATE reader - Indicates support for separate instruction and data address maps. Reads as 0 as ARMv6-M only supports a unified MPU.