Module rp2040_pac::ppb::nvic_ipr2

Use the Interrupt Priority Registers to assign a priority from 0 to 3 to each of the available interrupts. 0 is the highest priority, and 3 is the lowest.

Structs

NVIC_IPR2_SPEC

Use the Interrupt Priority Registers to assign a priority from 0 to 3 to each of the available interrupts. 0 is the highest priority, and 3 is the lowest.

Type Aliases

IP_8_R

Field IP_8 reader - Priority of interrupt 8

IP_8_W

Field IP_8 writer - Priority of interrupt 8

IP_9_R

Field IP_9 reader - Priority of interrupt 9

IP_9_W

Field IP_9 writer - Priority of interrupt 9

IP_10_R

Field IP_10 reader - Priority of interrupt 10

IP_10_W

Field IP_10 writer - Priority of interrupt 10

IP_11_R

Field IP_11 reader - Priority of interrupt 11

IP_11_W

Field IP_11 writer - Priority of interrupt 11

R

Register NVIC_IPR2 reader

W

Register NVIC_IPR2 writer