Module rp2040_pac::ppb::nvic_ipr3

Use the Interrupt Priority Registers to assign a priority from 0 to 3 to each of the available interrupts. 0 is the highest priority, and 3 is the lowest.

Structs

NVIC_IPR3_SPEC

Use the Interrupt Priority Registers to assign a priority from 0 to 3 to each of the available interrupts. 0 is the highest priority, and 3 is the lowest.

Type Aliases

IP_12_R

Field IP_12 reader - Priority of interrupt 12

IP_12_W

Field IP_12 writer - Priority of interrupt 12

IP_13_R

Field IP_13 reader - Priority of interrupt 13

IP_13_W

Field IP_13 writer - Priority of interrupt 13

IP_14_R

Field IP_14 reader - Priority of interrupt 14

IP_14_W

Field IP_14 writer - Priority of interrupt 14

IP_15_R

Field IP_15 reader - Priority of interrupt 15

IP_15_W

Field IP_15 writer - Priority of interrupt 15

R

Register NVIC_IPR3 reader

W

Register NVIC_IPR3 writer