rp2040_pac::ppb

Module nvic_ipr5

Use the Interrupt Priority Registers to assign a priority from 0 to 3 to each of the available interrupts. 0 is the highest priority, and 3 is the lowest.

Structs

NVIC_IPR5_SPEC

Use the Interrupt Priority Registers to assign a priority from 0 to 3 to each of the available interrupts. 0 is the highest priority, and 3 is the lowest.

Type Aliases

IP_20_R

Field IP_20 reader - Priority of interrupt 20

IP_20_W

Field IP_20 writer - Priority of interrupt 20

IP_21_R

Field IP_21 reader - Priority of interrupt 21

IP_21_W

Field IP_21 writer - Priority of interrupt 21

IP_22_R

Field IP_22 reader - Priority of interrupt 22

IP_22_W

Field IP_22 writer - Priority of interrupt 22

IP_23_R

Field IP_23 reader - Priority of interrupt 23

IP_23_W

Field IP_23 writer - Priority of interrupt 23

R

Register NVIC_IPR5 reader

W

Register NVIC_IPR5 writer