rp2040_pac::ppb
Module nvic_ipr6

Use the Interrupt Priority Registers to assign a priority from 0 to 3 to each of the available interrupts. 0 is the highest priority, and 3 is the lowest.

Structs

NVIC_IPR6_SPEC

Use the Interrupt Priority Registers to assign a priority from 0 to 3 to each of the available interrupts. 0 is the highest priority, and 3 is the lowest.

Type Aliases

IP_24_R

Field IP_24 reader - Priority of interrupt 24

IP_24_W

Field IP_24 writer - Priority of interrupt 24

IP_25_R

Field IP_25 reader - Priority of interrupt 25

IP_25_W

Field IP_25 writer - Priority of interrupt 25

IP_26_R

Field IP_26 reader - Priority of interrupt 26

IP_26_W

Field IP_26 writer - Priority of interrupt 26

IP_27_R

Field IP_27 reader - Priority of interrupt 27

IP_27_W

Field IP_27 writer - Priority of interrupt 27

R

Register NVIC_IPR6 reader

W

Register NVIC_IPR6 writer