Module rp2040_pac::ppb::shpr3

System handlers are a special class of exception handler that can have their priority set to any of the priority levels. Use the System Handler Priority Register 3 to set the priority of PendSV and SysTick.

Structs

SHPR3_SPEC

System handlers are a special class of exception handler that can have their priority set to any of the priority levels. Use the System Handler Priority Register 3 to set the priority of PendSV and SysTick.

Type Aliases

PRI_14_R

Field PRI_14 reader - Priority of system handler 14, PendSV

PRI_14_W

Field PRI_14 writer - Priority of system handler 14, PendSV

PRI_15_R

Field PRI_15 reader - Priority of system handler 15, SysTick

PRI_15_W

Field PRI_15 writer - Priority of system handler 15, SysTick

R

Register SHPR3 reader

W

Register SHPR3 writer