Struct rp2040_pac::ppb::RegisterBlock

#[repr(C)]
pub struct RegisterBlock {

    pub syst_csr: Reg<SYST_CSR_SPEC>,
    pub syst_rvr: Reg<SYST_RVR_SPEC>,
    pub syst_cvr: Reg<SYST_CVR_SPEC>,
    pub syst_calib: Reg<SYST_CALIB_SPEC>,
    pub nvic_iser: Reg<NVIC_ISER_SPEC>,
    pub nvic_icer: Reg<NVIC_ICER_SPEC>,
    pub nvic_ispr: Reg<NVIC_ISPR_SPEC>,
    pub nvic_icpr: Reg<NVIC_ICPR_SPEC>,
    pub nvic_ipr0: Reg<NVIC_IPR0_SPEC>,
    pub nvic_ipr1: Reg<NVIC_IPR1_SPEC>,
    pub nvic_ipr2: Reg<NVIC_IPR2_SPEC>,
    pub nvic_ipr3: Reg<NVIC_IPR3_SPEC>,
    pub nvic_ipr4: Reg<NVIC_IPR4_SPEC>,
    pub nvic_ipr5: Reg<NVIC_IPR5_SPEC>,
    pub nvic_ipr6: Reg<NVIC_IPR6_SPEC>,
    pub nvic_ipr7: Reg<NVIC_IPR7_SPEC>,
    pub cpuid: Reg<CPUID_SPEC>,
    pub icsr: Reg<ICSR_SPEC>,
    pub vtor: Reg<VTOR_SPEC>,
    pub aircr: Reg<AIRCR_SPEC>,
    pub scr: Reg<SCR_SPEC>,
    pub ccr: Reg<CCR_SPEC>,
    pub shpr2: Reg<SHPR2_SPEC>,
    pub shpr3: Reg<SHPR3_SPEC>,
    pub shcsr: Reg<SHCSR_SPEC>,
    pub mpu_type: Reg<MPU_TYPE_SPEC>,
    pub mpu_ctrl: Reg<MPU_CTRL_SPEC>,
    pub mpu_rnr: Reg<MPU_RNR_SPEC>,
    pub mpu_rbar: Reg<MPU_RBAR_SPEC>,
    pub mpu_rasr: Reg<MPU_RASR_SPEC>,
    /* private fields */

}

Register block

Fields
syst_csr: Reg<SYST_CSR_SPEC>

0xe010 - Use the SysTick Control and Status Register to enable the SysTick features.

syst_rvr: Reg<SYST_RVR_SPEC>

0xe014 - Use the SysTick Reload Value Register to specify the start value to load into the current value register when the counter reaches 0. It can be any value between 0 and 0x00FFFFFF. A start value of 0 is possible, but has no effect because the SysTick interrupt and COUNTFLAG are activated when counting from 1 to 0. The reset value of this register is UNKNOWN.
To generate a multi-shot timer with a period of N processor clock cycles, use a RELOAD value of N-1. For example, if the SysTick interrupt is required every 100 clock pulses, set RELOAD to 99.

syst_cvr: Reg<SYST_CVR_SPEC>

0xe018 - Use the SysTick Current Value Register to find the current value in the register. The reset value of this register is UNKNOWN.

syst_calib: Reg<SYST_CALIB_SPEC>

0xe01c - Use the SysTick Calibration Value Register to enable software to scale to any required speed using divide and multiply.

nvic_iser: Reg<NVIC_ISER_SPEC>

0xe100 - Use the Interrupt Set-Enable Register to enable interrupts and determine which interrupts are currently enabled.
If a pending interrupt is enabled, the NVIC activates the interrupt based on its priority. If an interrupt is not enabled, asserting its interrupt signal changes the interrupt state to pending, but the NVIC never activates the interrupt, regardless of its priority.

nvic_icer: Reg<NVIC_ICER_SPEC>

0xe180 - Use the Interrupt Clear-Enable Registers to disable interrupts and determine which interrupts are currently enabled.

nvic_ispr: Reg<NVIC_ISPR_SPEC>

0xe200 - The NVIC_ISPR forces interrupts into the pending state, and shows which interrupts are pending.

nvic_icpr: Reg<NVIC_ICPR_SPEC>

0xe280 - Use the Interrupt Clear-Pending Register to clear pending interrupts and determine which interrupts are currently pending.

nvic_ipr0: Reg<NVIC_IPR0_SPEC>

0xe400 - Use the Interrupt Priority Registers to assign a priority from 0 to 3 to each of the available interrupts. 0 is the highest priority, and 3 is the lowest.
Note: Writing 1 to an NVIC_ICPR bit does not affect the active state of the corresponding interrupt.
These registers are only word-accessible

nvic_ipr1: Reg<NVIC_IPR1_SPEC>

0xe404 - Use the Interrupt Priority Registers to assign a priority from 0 to 3 to each of the available interrupts. 0 is the highest priority, and 3 is the lowest.

nvic_ipr2: Reg<NVIC_IPR2_SPEC>

0xe408 - Use the Interrupt Priority Registers to assign a priority from 0 to 3 to each of the available interrupts. 0 is the highest priority, and 3 is the lowest.

nvic_ipr3: Reg<NVIC_IPR3_SPEC>

0xe40c - Use the Interrupt Priority Registers to assign a priority from 0 to 3 to each of the available interrupts. 0 is the highest priority, and 3 is the lowest.

nvic_ipr4: Reg<NVIC_IPR4_SPEC>

0xe410 - Use the Interrupt Priority Registers to assign a priority from 0 to 3 to each of the available interrupts. 0 is the highest priority, and 3 is the lowest.

nvic_ipr5: Reg<NVIC_IPR5_SPEC>

0xe414 - Use the Interrupt Priority Registers to assign a priority from 0 to 3 to each of the available interrupts. 0 is the highest priority, and 3 is the lowest.

nvic_ipr6: Reg<NVIC_IPR6_SPEC>

0xe418 - Use the Interrupt Priority Registers to assign a priority from 0 to 3 to each of the available interrupts. 0 is the highest priority, and 3 is the lowest.

nvic_ipr7: Reg<NVIC_IPR7_SPEC>

0xe41c - Use the Interrupt Priority Registers to assign a priority from 0 to 3 to each of the available interrupts. 0 is the highest priority, and 3 is the lowest.

cpuid: Reg<CPUID_SPEC>

0xed00 - Read the CPU ID Base Register to determine: the ID number of the processor core, the version number of the processor core, the implementation details of the processor core.

icsr: Reg<ICSR_SPEC>

0xed04 - Use the Interrupt Control State Register to set a pending Non-Maskable Interrupt (NMI), set or clear a pending PendSV, set or clear a pending SysTick, check for pending exceptions, check the vector number of the highest priority pended exception, check the vector number of the active exception.

vtor: Reg<VTOR_SPEC>

0xed08 - The VTOR holds the vector table offset address.

aircr: Reg<AIRCR_SPEC>

0xed0c - Use the Application Interrupt and Reset Control Register to: determine data endianness, clear all active state information from debug halt mode, request a system reset.

scr: Reg<SCR_SPEC>

0xed10 - System Control Register. Use the System Control Register for power-management functions: signal to the system when the processor can enter a low power state, control how the processor enters and exits low power states.

ccr: Reg<CCR_SPEC>

0xed14 - The Configuration and Control Register permanently enables stack alignment and causes unaligned accesses to result in a Hard Fault.

shpr2: Reg<SHPR2_SPEC>

0xed1c - System handlers are a special class of exception handler that can have their priority set to any of the priority levels. Use the System Handler Priority Register 2 to set the priority of SVCall.

shpr3: Reg<SHPR3_SPEC>

0xed20 - System handlers are a special class of exception handler that can have their priority set to any of the priority levels. Use the System Handler Priority Register 3 to set the priority of PendSV and SysTick.

shcsr: Reg<SHCSR_SPEC>

0xed24 - Use the System Handler Control and State Register to determine or clear the pending status of SVCall.

mpu_type: Reg<MPU_TYPE_SPEC>

0xed90 - Read the MPU Type Register to determine if the processor implements an MPU, and how many regions the MPU supports.

mpu_ctrl: Reg<MPU_CTRL_SPEC>

0xed94 - Use the MPU Control Register to enable and disable the MPU, and to control whether the default memory map is enabled as a background region for privileged accesses, and whether the MPU is enabled for HardFaults and NMIs.

mpu_rnr: Reg<MPU_RNR_SPEC>

0xed98 - Use the MPU Region Number Register to select the region currently accessed by MPU_RBAR and MPU_RASR.

mpu_rbar: Reg<MPU_RBAR_SPEC>

0xed9c - Read the MPU Region Base Address Register to determine the base address of the region identified by MPU_RNR. Write to update the base address of said region or that of a specified region, with whose number MPU_RNR will also be updated.

mpu_rasr: Reg<MPU_RASR_SPEC>

0xeda0 - Use the MPU Region Attribute and Size Register to define the size, access behaviour and memory type of the region identified by MPU_RNR, and enable that region.