Module rp2040_pac::ppb::syst_cvr

Use the SysTick Current Value Register to find the current value in the register. The reset value of this register is UNKNOWN.

Structs

SYST_CVR_SPEC

Use the SysTick Current Value Register to find the current value in the register. The reset value of this register is UNKNOWN.

Type Aliases

CURRENT_R

Field CURRENT reader - Reads return the current value of the SysTick counter. This register is write-clear. Writing to it with any value clears the register to 0. Clearing this register also clears the COUNTFLAG bit of the SysTick Control and Status Register.

CURRENT_W

Field CURRENT writer - Reads return the current value of the SysTick counter. This register is write-clear. Writing to it with any value clears the register to 0. Clearing this register also clears the COUNTFLAG bit of the SysTick Control and Status Register.

R

Register SYST_CVR reader

W

Register SYST_CVR writer