rp2040_pac::ppb

Type Alias NVIC_ISER

pub type NVIC_ISER = Reg<NVIC_ISER_SPEC>;

NVIC_ISER (rw) register accessor: Use the Interrupt Set-Enable Register to enable interrupts and determine which interrupts are currently enabled.
If a pending interrupt is enabled, the NVIC activates the interrupt based on its priority. If an interrupt is not enabled, asserting its interrupt signal changes the interrupt state to pending, but the NVIC never activates the interrupt, regardless of its priority.

You can read this register and get nvic_iser::R. You can reset, write, write_with_zero this register using nvic_iser::W. You can also modify this register. See API.

For information about available fields see nvic_iser module

Aliased Type

struct NVIC_ISER { /* private fields */ }