Type Alias rp2040_pac::ppb::SCR

pub type SCR = Reg<SCR_SPEC>;

SCR (rw) register accessor: System Control Register. Use the System Control Register for power-management functions: signal to the system when the processor can enter a low power state, control how the processor enters and exits low power states.

You can read this register and get scr::R. You can reset, write, write_with_zero this register using scr::W. You can also modify this register. See API.

For information about available fields see scr module

Aliased Type

struct SCR { /* private fields */ }