Module rp2040_pac::psm

PSM

Modules

done

Indicates the peripheral’s registers are ready to access.

frce_off

Force into reset (i.e. power it off)

frce_on

Force block out of reset (i.e. power it on)

wdsel

Set to 1 if this peripheral should be reset when the watchdog fires.

Structs

RegisterBlock

Register block

Type Aliases

DONE

DONE (r) register accessor: Indicates the peripheral’s registers are ready to access.

FRCE_OFF

FRCE_OFF (rw) register accessor: Force into reset (i.e. power it off)

FRCE_ON

FRCE_ON (rw) register accessor: Force block out of reset (i.e. power it on)

WDSEL

WDSEL (rw) register accessor: Set to 1 if this peripheral should be reset when the watchdog fires.