Module rp2040_pac::pwm

Simple PWM

Modules

ch

Register block Cluster CH%s, containing CH*_CC, CH*_CSR, CH*_CTR, CH*_DIV, CH*_TOP

en

This register aliases the CSR_EN bits for all channels.
Writing to this register allows multiple channels to be enabled
or disabled simultaneously, so they can run in perfect sync.
For each channel, there is only one physical EN register bit,
which can be accessed through here or CHx_CSR.

inte

Interrupt Enable

intf

Interrupt Force

intr

Raw Interrupts

ints

Interrupt status after masking & forcing

Structs

CH

Register block

RegisterBlock

Register block

Type Definitions

EN

EN register accessor: an alias for Reg<EN_SPEC>

INTE

INTE register accessor: an alias for Reg<INTE_SPEC>

INTF

INTF register accessor: an alias for Reg<INTF_SPEC>

INTR

INTR register accessor: an alias for Reg<INTR_SPEC>

INTS

INTS register accessor: an alias for Reg<INTS_SPEC>