rp2040_pac::pwm::ch::csr

Type Alias DIVMODE_R

pub type DIVMODE_R = FieldReader<DIVMODE_A>;

Field DIVMODE reader -

Aliased Type

struct DIVMODE_R { /* private fields */ }

Implementations

impl DIVMODE_R

pub const fn variant(&self) -> DIVMODE_A

Get enumerated values variant

pub fn is_div(&self) -> bool

Free-running counting at rate dictated by fractional divider

pub fn is_level(&self) -> bool

Fractional divider operation is gated by the PWM B pin.

pub fn is_rise(&self) -> bool

Counter advances with each rising edge of the PWM B pin.

pub fn is_fall(&self) -> bool

Counter advances with each falling edge of the PWM B pin.