Type Alias rp2040_pac::pwm::ch::csr::DIVMODE_W

pub type DIVMODE_W<'a, REG> = FieldWriterSafe<'a, REG, 2, DIVMODE_A>;

Field DIVMODE writer -

Aliased Type

struct DIVMODE_W<'a, REG> { /* private fields */ }

Implementations

impl<'a, REG> DIVMODE_W<'a, REG>
where REG: Writable + RegisterSpec, REG::Ux: From<u8>,

pub fn div(self) -> &'a mut W<REG>

Free-running counting at rate dictated by fractional divider

pub fn level(self) -> &'a mut W<REG>

Fractional divider operation is gated by the PWM B pin.

pub fn rise(self) -> &'a mut W<REG>

Counter advances with each rising edge of the PWM B pin.

pub fn fall(self) -> &'a mut W<REG>

Counter advances with each falling edge of the PWM B pin.