rp2040_pac::pwm

Module en

This register aliases the CSR_EN bits for all channels.
Writing to this register allows multiple channels to be enabled
or disabled simultaneously, so they can run in perfect sync.
For each channel, there is only one physical EN register bit,
which can be accessed through here or CHx_CSR.

Structs

EN_SPEC

This register aliases the CSR_EN bits for all channels.
Writing to this register allows multiple channels to be enabled
or disabled simultaneously, so they can run in perfect sync.
For each channel, there is only one physical EN register bit,
which can be accessed through here or CHx_CSR.

Type Aliases

CH0_R

Field CH0 reader -

CH0_W

Field CH0 writer -

CH1_R

Field CH1 reader -

CH1_W

Field CH1 writer -

CH2_R

Field CH2 reader -

CH2_W

Field CH2 writer -

CH3_R

Field CH3 reader -

CH3_W

Field CH3 writer -

CH4_R

Field CH4 reader -

CH4_W

Field CH4 writer -

CH5_R

Field CH5 reader -

CH5_W

Field CH5 writer -

CH6_R

Field CH6 reader -

CH6_W

Field CH6 writer -

CH7_R

Field CH7 reader -

CH7_W

Field CH7 writer -

R

Register EN reader

W

Register EN writer