rp2040_pac::rosc::ctrl

Type Alias FREQ_RANGE_R

pub type FREQ_RANGE_R = FieldReader<FREQ_RANGE_A>;

Field FREQ_RANGE reader - Controls the number of delay stages in the ROSC ring
LOW uses stages 0 to 7
MEDIUM uses stages 0 to 5
HIGH uses stages 0 to 3
TOOHIGH uses stages 0 to 1 and should not be used because its frequency exceeds design specifications
The clock output will not glitch when changing the range up one step at a time
The clock output will glitch when changing the range down
Note: the values here are gray coded which is why HIGH comes before TOOHIGH

Aliased Type

struct FREQ_RANGE_R { /* private fields */ }

Implementations

impl FREQ_RANGE_R

pub const fn variant(&self) -> Option<FREQ_RANGE_A>

Get enumerated values variant

pub fn is_low(&self) -> bool

111110100100

pub fn is_medium(&self) -> bool

111110100101

pub fn is_high(&self) -> bool

111110100111

pub fn is_toohigh(&self) -> bool

111110100110