rp2040_pac::rosc::ctrl

Type Alias FREQ_RANGE_W

pub type FREQ_RANGE_W<'a, REG> = FieldWriter<'a, REG, 12, FREQ_RANGE_A>;

Field FREQ_RANGE writer - Controls the number of delay stages in the ROSC ring
LOW uses stages 0 to 7
MEDIUM uses stages 0 to 5
HIGH uses stages 0 to 3
TOOHIGH uses stages 0 to 1 and should not be used because its frequency exceeds design specifications
The clock output will not glitch when changing the range up one step at a time
The clock output will glitch when changing the range down
Note: the values here are gray coded which is why HIGH comes before TOOHIGH

Aliased Type

struct FREQ_RANGE_W<'a, REG> { /* private fields */ }

Implementations

impl<'a, REG> FREQ_RANGE_W<'a, REG>
where REG: Writable + RegisterSpec, REG::Ux: From<u16>,

pub fn low(self) -> &'a mut W<REG>

111110100100

pub fn medium(self) -> &'a mut W<REG>

111110100101

pub fn high(self) -> &'a mut W<REG>

111110100111

pub fn toohigh(self) -> &'a mut W<REG>

111110100110