Module rp2040_pac::rosc::freqb

For a detailed description see freqa register

Structs

DS4_R

Field DS4 reader - Stage 4 drive strength

DS4_W

Field DS4 writer - Stage 4 drive strength

DS5_R

Field DS5 reader - Stage 5 drive strength

DS5_W

Field DS5 writer - Stage 5 drive strength

DS6_R

Field DS6 reader - Stage 6 drive strength

DS6_W

Field DS6 writer - Stage 6 drive strength

DS7_R

Field DS7 reader - Stage 7 drive strength

DS7_W

Field DS7 writer - Stage 7 drive strength

FREQB_SPEC

For a detailed description see freqa register

PASSWD_R

Field PASSWD reader - Set to 0x9696 to apply the settings
Any other value in this field will set all drive strengths to 0

PASSWD_W

Field PASSWD writer - Set to 0x9696 to apply the settings
Any other value in this field will set all drive strengths to 0

R

Register FREQB reader

W

Register FREQB writer

Enums

PASSWD_A

Set to 0x9696 to apply the settings
Any other value in this field will set all drive strengths to 0