Module rp2040_pac::rtc::irq_setup_0

Interrupt setup register 0

Structs

DAY_ENA_R

Field DAY_ENA reader - Enable day matching

DAY_ENA_W

Field DAY_ENA writer - Enable day matching

DAY_R

Field DAY reader - Day of the month (1..31)

DAY_W

Field DAY writer - Day of the month (1..31)

IRQ_SETUP_0_SPEC

Interrupt setup register 0

MATCH_ACTIVE_R

Field MATCH_ACTIVE reader -

MATCH_ENA_R

Field MATCH_ENA reader - Global match enable. Don’t change any other value while this one is enabled

MATCH_ENA_W

Field MATCH_ENA writer - Global match enable. Don’t change any other value while this one is enabled

MONTH_ENA_R

Field MONTH_ENA reader - Enable month matching

MONTH_ENA_W

Field MONTH_ENA writer - Enable month matching

MONTH_R

Field MONTH reader - Month (1..12)

MONTH_W

Field MONTH writer - Month (1..12)

R

Register IRQ_SETUP_0 reader

W

Register IRQ_SETUP_0 writer

YEAR_ENA_R

Field YEAR_ENA reader - Enable year matching

YEAR_ENA_W

Field YEAR_ENA writer - Enable year matching

YEAR_R

Field YEAR reader - Year

YEAR_W

Field YEAR writer - Year