rp2040_pac::sio

Module div_csr

Control and status register for divider.

Structs

DIV_CSR_SPEC

Control and status register for divider.

Type Aliases

DIRTY_R

Field DIRTY reader - Changes to 1 when any register is written, and back to 0 when QUOTIENT is read.
Software can use this flag to make save/restore more efficient (skip if not DIRTY).
If the flag is used in this way, it’s recommended to either read QUOTIENT only,
or REMAINDER and then QUOTIENT, to prevent data loss on context switch.

R

Register DIV_CSR reader

READY_R

Field READY reader - Reads as 0 when a calculation is in progress, 1 otherwise.
Writing an operand (xDIVIDEND, xDIVISOR) will immediately start a new calculation, no
matter if one is already in progress.
Writing to a result register will immediately terminate any in-progress calculation
and set the READY and DIRTY flags.