rp2040_pac::sio::div_csr

Type Alias R

pub type R = R<DIV_CSR_SPEC>;

Register DIV_CSR reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn ready(&self) -> READY_R

Bit 0 - Reads as 0 when a calculation is in progress, 1 otherwise.
Writing an operand (xDIVIDEND, xDIVISOR) will immediately start a new calculation, no
matter if one is already in progress.
Writing to a result register will immediately terminate any in-progress calculation
and set the READY and DIRTY flags.

pub fn dirty(&self) -> DIRTY_R

Bit 1 - Changes to 1 when any register is written, and back to 0 when QUOTIENT is read.
Software can use this flag to make save/restore more efficient (skip if not DIRTY).
If the flag is used in this way, it’s recommended to either read QUOTIENT only,
or REMAINDER and then QUOTIENT, to prevent data loss on context switch.