rp2040_pac::sio
Module gpio_hi_out

QSPI output value

Structs

GPIO_HI_OUT_SPEC

QSPI output value

Type Aliases

GPIO_HI_OUT_R

Field GPIO_HI_OUT reader - Set output level (1/0 -> high/low) for QSPI IO0…5.
Reading back gives the last value written, NOT the input value from the pins.
If core 0 and core 1 both write to GPIO_HI_OUT simultaneously (or to a SET/CLR/XOR alias),
the result is as though the write from core 0 took place first,
and the write from core 1 was then applied to that intermediate result.

GPIO_HI_OUT_W

Field GPIO_HI_OUT writer - Set output level (1/0 -> high/low) for QSPI IO0…5.
Reading back gives the last value written, NOT the input value from the pins.
If core 0 and core 1 both write to GPIO_HI_OUT simultaneously (or to a SET/CLR/XOR alias),
the result is as though the write from core 0 took place first,
and the write from core 1 was then applied to that intermediate result.

R

Register GPIO_HI_OUT reader

W

Register GPIO_HI_OUT writer