Module rp2040_pac::sio::gpio_oe

GPIO output enable

Structs

GPIO_OE_SPEC

GPIO output enable

Type Aliases

GPIO_OE_R

Field GPIO_OE reader - Set output enable (1/0 -> output/input) for GPIO0…29.
Reading back gives the last value written.
If core 0 and core 1 both write to GPIO_OE simultaneously (or to a SET/CLR/XOR alias),
the result is as though the write from core 0 took place first,
and the write from core 1 was then applied to that intermediate result.

GPIO_OE_W

Field GPIO_OE writer - Set output enable (1/0 -> output/input) for GPIO0…29.
Reading back gives the last value written.
If core 0 and core 1 both write to GPIO_OE simultaneously (or to a SET/CLR/XOR alias),
the result is as though the write from core 0 took place first,
and the write from core 1 was then applied to that intermediate result.

R

Register GPIO_OE reader

W

Register GPIO_OE writer