rp2040_pac::sio

Module gpio_out

GPIO output value

Structs

GPIO_OUT_SPEC

GPIO output value

Type Aliases

GPIO_OUT_R

Field GPIO_OUT reader - Set output level (1/0 -> high/low) for GPIO0…29.
Reading back gives the last value written, NOT the input value from the pins.
If core 0 and core 1 both write to GPIO_OUT simultaneously (or to a SET/CLR/XOR alias),
the result is as though the write from core 0 took place first,
and the write from core 1 was then applied to that intermediate result.

GPIO_OUT_W

Field GPIO_OUT writer - Set output level (1/0 -> high/low) for GPIO0…29.
Reading back gives the last value written, NOT the input value from the pins.
If core 0 and core 1 both write to GPIO_OUT simultaneously (or to a SET/CLR/XOR alias),
the result is as though the write from core 0 took place first,
and the write from core 1 was then applied to that intermediate result.

R

Register GPIO_OUT reader

W

Register GPIO_OUT writer