Struct rp2040_pac::sio::RegisterBlock

#[repr(C)]
pub struct RegisterBlock {

    pub cpuid: Reg<CPUID_SPEC>,
    pub gpio_in: Reg<GPIO_IN_SPEC>,
    pub gpio_hi_in: Reg<GPIO_HI_IN_SPEC>,
    pub gpio_out: Reg<GPIO_OUT_SPEC>,
    pub gpio_out_set: Reg<GPIO_OUT_SET_SPEC>,
    pub gpio_out_clr: Reg<GPIO_OUT_CLR_SPEC>,
    pub gpio_out_xor: Reg<GPIO_OUT_XOR_SPEC>,
    pub gpio_oe: Reg<GPIO_OE_SPEC>,
    pub gpio_oe_set: Reg<GPIO_OE_SET_SPEC>,
    pub gpio_oe_clr: Reg<GPIO_OE_CLR_SPEC>,
    pub gpio_oe_xor: Reg<GPIO_OE_XOR_SPEC>,
    pub gpio_hi_out: Reg<GPIO_HI_OUT_SPEC>,
    pub gpio_hi_out_set: Reg<GPIO_HI_OUT_SET_SPEC>,
    pub gpio_hi_out_clr: Reg<GPIO_HI_OUT_CLR_SPEC>,
    pub gpio_hi_out_xor: Reg<GPIO_HI_OUT_XOR_SPEC>,
    pub gpio_hi_oe: Reg<GPIO_HI_OE_SPEC>,
    pub gpio_hi_oe_set: Reg<GPIO_HI_OE_SET_SPEC>,
    pub gpio_hi_oe_clr: Reg<GPIO_HI_OE_CLR_SPEC>,
    pub gpio_hi_oe_xor: Reg<GPIO_HI_OE_XOR_SPEC>,
    pub fifo_st: Reg<FIFO_ST_SPEC>,
    pub fifo_wr: Reg<FIFO_WR_SPEC>,
    pub fifo_rd: Reg<FIFO_RD_SPEC>,
    pub spinlock_st: Reg<SPINLOCK_ST_SPEC>,
    pub div_udividend: Reg<DIV_UDIVIDEND_SPEC>,
    pub div_udivisor: Reg<DIV_UDIVISOR_SPEC>,
    pub div_sdividend: Reg<DIV_SDIVIDEND_SPEC>,
    pub div_sdivisor: Reg<DIV_SDIVISOR_SPEC>,
    pub div_quotient: Reg<DIV_QUOTIENT_SPEC>,
    pub div_remainder: Reg<DIV_REMAINDER_SPEC>,
    pub div_csr: Reg<DIV_CSR_SPEC>,
    pub interp0_accum0: Reg<INTERP0_ACCUM0_SPEC>,
    pub interp0_accum1: Reg<INTERP0_ACCUM1_SPEC>,
    pub interp0_base0: Reg<INTERP0_BASE0_SPEC>,
    pub interp0_base1: Reg<INTERP0_BASE1_SPEC>,
    pub interp0_base2: Reg<INTERP0_BASE2_SPEC>,
    pub interp0_pop_lane0: Reg<INTERP0_POP_LANE0_SPEC>,
    pub interp0_pop_lane1: Reg<INTERP0_POP_LANE1_SPEC>,
    pub interp0_pop_full: Reg<INTERP0_POP_FULL_SPEC>,
    pub interp0_peek_lane0: Reg<INTERP0_PEEK_LANE0_SPEC>,
    pub interp0_peek_lane1: Reg<INTERP0_PEEK_LANE1_SPEC>,
    pub interp0_peek_full: Reg<INTERP0_PEEK_FULL_SPEC>,
    pub interp0_ctrl_lane0: Reg<INTERP0_CTRL_LANE0_SPEC>,
    pub interp0_ctrl_lane1: Reg<INTERP0_CTRL_LANE1_SPEC>,
    pub interp0_accum0_add: Reg<INTERP0_ACCUM0_ADD_SPEC>,
    pub interp0_accum1_add: Reg<INTERP0_ACCUM1_ADD_SPEC>,
    pub interp0_base_1and0: Reg<INTERP0_BASE_1AND0_SPEC>,
    pub interp1_accum0: Reg<INTERP1_ACCUM0_SPEC>,
    pub interp1_accum1: Reg<INTERP1_ACCUM1_SPEC>,
    pub interp1_base0: Reg<INTERP1_BASE0_SPEC>,
    pub interp1_base1: Reg<INTERP1_BASE1_SPEC>,
    pub interp1_base2: Reg<INTERP1_BASE2_SPEC>,
    pub interp1_pop_lane0: Reg<INTERP1_POP_LANE0_SPEC>,
    pub interp1_pop_lane1: Reg<INTERP1_POP_LANE1_SPEC>,
    pub interp1_pop_full: Reg<INTERP1_POP_FULL_SPEC>,
    pub interp1_peek_lane0: Reg<INTERP1_PEEK_LANE0_SPEC>,
    pub interp1_peek_lane1: Reg<INTERP1_PEEK_LANE1_SPEC>,
    pub interp1_peek_full: Reg<INTERP1_PEEK_FULL_SPEC>,
    pub interp1_ctrl_lane0: Reg<INTERP1_CTRL_LANE0_SPEC>,
    pub interp1_ctrl_lane1: Reg<INTERP1_CTRL_LANE1_SPEC>,
    pub interp1_accum0_add: Reg<INTERP1_ACCUM0_ADD_SPEC>,
    pub interp1_accum1_add: Reg<INTERP1_ACCUM1_ADD_SPEC>,
    pub interp1_base_1and0: Reg<INTERP1_BASE_1AND0_SPEC>,
    pub spinlock: [Reg<SPINLOCK_SPEC>; 32],
    /* private fields */

}

Register block

Fields
cpuid: Reg<CPUID_SPEC>

0x00 - Processor core identifier
Value is 0 when read from processor core 0, and 1 when read from processor core 1.

gpio_in: Reg<GPIO_IN_SPEC>

0x04 - Input value for GPIO pins

gpio_hi_in: Reg<GPIO_HI_IN_SPEC>

0x08 - Input value for QSPI pins

gpio_out: Reg<GPIO_OUT_SPEC>

0x10 - GPIO output value

gpio_out_set: Reg<GPIO_OUT_SET_SPEC>

0x14 - GPIO output value set

gpio_out_clr: Reg<GPIO_OUT_CLR_SPEC>

0x18 - GPIO output value clear

gpio_out_xor: Reg<GPIO_OUT_XOR_SPEC>

0x1c - GPIO output value XOR

gpio_oe: Reg<GPIO_OE_SPEC>

0x20 - GPIO output enable

gpio_oe_set: Reg<GPIO_OE_SET_SPEC>

0x24 - GPIO output enable set

gpio_oe_clr: Reg<GPIO_OE_CLR_SPEC>

0x28 - GPIO output enable clear

gpio_oe_xor: Reg<GPIO_OE_XOR_SPEC>

0x2c - GPIO output enable XOR

gpio_hi_out: Reg<GPIO_HI_OUT_SPEC>

0x30 - QSPI output value

gpio_hi_out_set: Reg<GPIO_HI_OUT_SET_SPEC>

0x34 - QSPI output value set

gpio_hi_out_clr: Reg<GPIO_HI_OUT_CLR_SPEC>

0x38 - QSPI output value clear

gpio_hi_out_xor: Reg<GPIO_HI_OUT_XOR_SPEC>

0x3c - QSPI output value XOR

gpio_hi_oe: Reg<GPIO_HI_OE_SPEC>

0x40 - QSPI output enable

gpio_hi_oe_set: Reg<GPIO_HI_OE_SET_SPEC>

0x44 - QSPI output enable set

gpio_hi_oe_clr: Reg<GPIO_HI_OE_CLR_SPEC>

0x48 - QSPI output enable clear

gpio_hi_oe_xor: Reg<GPIO_HI_OE_XOR_SPEC>

0x4c - QSPI output enable XOR

fifo_st: Reg<FIFO_ST_SPEC>

0x50 - Status register for inter-core FIFOs (mailboxes).
There is one FIFO in the core 0 -> core 1 direction, and one core 1 -> core 0. Both are 32 bits wide and 8 words deep.
Core 0 can see the read side of the 1->0 FIFO (RX), and the write side of 0->1 FIFO (TX).
Core 1 can see the read side of the 0->1 FIFO (RX), and the write side of 1->0 FIFO (TX).
The SIO IRQ for each core is the logical OR of the VLD, WOF and ROE fields of its FIFO_ST register.

fifo_wr: Reg<FIFO_WR_SPEC>

0x54 - Write access to this core’s TX FIFO

fifo_rd: Reg<FIFO_RD_SPEC>

0x58 - Read access to this core’s RX FIFO

spinlock_st: Reg<SPINLOCK_ST_SPEC>

0x5c - Spinlock state
A bitmap containing the state of all 32 spinlocks (1=locked).
Mainly intended for debugging.

div_udividend: Reg<DIV_UDIVIDEND_SPEC>

0x60 - Divider unsigned dividend
Write to the DIVIDEND operand of the divider, i.e. the p in p / q.
Any operand write starts a new calculation. The results appear in QUOTIENT, REMAINDER.
UDIVIDEND/SDIVIDEND are aliases of the same internal register. The U alias starts an
unsigned calculation, and the S alias starts a signed calculation.

div_udivisor: Reg<DIV_UDIVISOR_SPEC>

0x64 - Divider unsigned divisor
Write to the DIVISOR operand of the divider, i.e. the q in p / q.
Any operand write starts a new calculation. The results appear in QUOTIENT, REMAINDER.
UDIVISOR/SDIVISOR are aliases of the same internal register. The U alias starts an
unsigned calculation, and the S alias starts a signed calculation.

div_sdividend: Reg<DIV_SDIVIDEND_SPEC>

0x68 - Divider signed dividend
The same as UDIVIDEND, but starts a signed calculation, rather than unsigned.

div_sdivisor: Reg<DIV_SDIVISOR_SPEC>

0x6c - Divider signed divisor
The same as UDIVISOR, but starts a signed calculation, rather than unsigned.

div_quotient: Reg<DIV_QUOTIENT_SPEC>

0x70 - Divider result quotient
The result of DIVIDEND / DIVISOR (division). Contents undefined while CSR_READY is low.
For signed calculations, QUOTIENT is negative when the signs of DIVIDEND and DIVISOR differ.
This register can be written to directly, for context save/restore purposes. This halts any
in-progress calculation and sets the CSR_READY and CSR_DIRTY flags.
Reading from QUOTIENT clears the CSR_DIRTY flag, so should read results in the order
REMAINDER, QUOTIENT if CSR_DIRTY is used.

div_remainder: Reg<DIV_REMAINDER_SPEC>

0x74 - Divider result remainder
The result of DIVIDEND % DIVISOR (modulo). Contents undefined while CSR_READY is low.
For signed calculations, REMAINDER is negative only when DIVIDEND is negative.
This register can be written to directly, for context save/restore purposes. This halts any
in-progress calculation and sets the CSR_READY and CSR_DIRTY flags.

div_csr: Reg<DIV_CSR_SPEC>

0x78 - Control and status register for divider.

interp0_accum0: Reg<INTERP0_ACCUM0_SPEC>

0x80 - Read/write access to accumulator 0

interp0_accum1: Reg<INTERP0_ACCUM1_SPEC>

0x84 - Read/write access to accumulator 1

interp0_base0: Reg<INTERP0_BASE0_SPEC>

0x88 - Read/write access to BASE0 register.

interp0_base1: Reg<INTERP0_BASE1_SPEC>

0x8c - Read/write access to BASE1 register.

interp0_base2: Reg<INTERP0_BASE2_SPEC>

0x90 - Read/write access to BASE2 register.

interp0_pop_lane0: Reg<INTERP0_POP_LANE0_SPEC>

0x94 - Read LANE0 result, and simultaneously write lane results to both accumulators (POP).

interp0_pop_lane1: Reg<INTERP0_POP_LANE1_SPEC>

0x98 - Read LANE1 result, and simultaneously write lane results to both accumulators (POP).

interp0_pop_full: Reg<INTERP0_POP_FULL_SPEC>

0x9c - Read FULL result, and simultaneously write lane results to both accumulators (POP).

interp0_peek_lane0: Reg<INTERP0_PEEK_LANE0_SPEC>

0xa0 - Read LANE0 result, without altering any internal state (PEEK).

interp0_peek_lane1: Reg<INTERP0_PEEK_LANE1_SPEC>

0xa4 - Read LANE1 result, without altering any internal state (PEEK).

interp0_peek_full: Reg<INTERP0_PEEK_FULL_SPEC>

0xa8 - Read FULL result, without altering any internal state (PEEK).

interp0_ctrl_lane0: Reg<INTERP0_CTRL_LANE0_SPEC>

0xac - Control register for lane 0

interp0_ctrl_lane1: Reg<INTERP0_CTRL_LANE1_SPEC>

0xb0 - Control register for lane 1

interp0_accum0_add: Reg<INTERP0_ACCUM0_ADD_SPEC>

0xb4 - Values written here are atomically added to ACCUM0
Reading yields lane 0’s raw shift and mask value (BASE0 not added).

interp0_accum1_add: Reg<INTERP0_ACCUM1_ADD_SPEC>

0xb8 - Values written here are atomically added to ACCUM1
Reading yields lane 1’s raw shift and mask value (BASE1 not added).

interp0_base_1and0: Reg<INTERP0_BASE_1AND0_SPEC>

0xbc - On write, the lower 16 bits go to BASE0, upper bits to BASE1 simultaneously.
Each half is sign-extended to 32 bits if that lane’s SIGNED flag is set.

interp1_accum0: Reg<INTERP1_ACCUM0_SPEC>

0xc0 - Read/write access to accumulator 0

interp1_accum1: Reg<INTERP1_ACCUM1_SPEC>

0xc4 - Read/write access to accumulator 1

interp1_base0: Reg<INTERP1_BASE0_SPEC>

0xc8 - Read/write access to BASE0 register.

interp1_base1: Reg<INTERP1_BASE1_SPEC>

0xcc - Read/write access to BASE1 register.

interp1_base2: Reg<INTERP1_BASE2_SPEC>

0xd0 - Read/write access to BASE2 register.

interp1_pop_lane0: Reg<INTERP1_POP_LANE0_SPEC>

0xd4 - Read LANE0 result, and simultaneously write lane results to both accumulators (POP).

interp1_pop_lane1: Reg<INTERP1_POP_LANE1_SPEC>

0xd8 - Read LANE1 result, and simultaneously write lane results to both accumulators (POP).

interp1_pop_full: Reg<INTERP1_POP_FULL_SPEC>

0xdc - Read FULL result, and simultaneously write lane results to both accumulators (POP).

interp1_peek_lane0: Reg<INTERP1_PEEK_LANE0_SPEC>

0xe0 - Read LANE0 result, without altering any internal state (PEEK).

interp1_peek_lane1: Reg<INTERP1_PEEK_LANE1_SPEC>

0xe4 - Read LANE1 result, without altering any internal state (PEEK).

interp1_peek_full: Reg<INTERP1_PEEK_FULL_SPEC>

0xe8 - Read FULL result, without altering any internal state (PEEK).

interp1_ctrl_lane0: Reg<INTERP1_CTRL_LANE0_SPEC>

0xec - Control register for lane 0

interp1_ctrl_lane1: Reg<INTERP1_CTRL_LANE1_SPEC>

0xf0 - Control register for lane 1

interp1_accum0_add: Reg<INTERP1_ACCUM0_ADD_SPEC>

0xf4 - Values written here are atomically added to ACCUM0
Reading yields lane 0’s raw shift and mask value (BASE0 not added).

interp1_accum1_add: Reg<INTERP1_ACCUM1_ADD_SPEC>

0xf8 - Values written here are atomically added to ACCUM1
Reading yields lane 1’s raw shift and mask value (BASE1 not added).

interp1_base_1and0: Reg<INTERP1_BASE_1AND0_SPEC>

0xfc - On write, the lower 16 bits go to BASE0, upper bits to BASE1 simultaneously.
Each half is sign-extended to 32 bits if that lane’s SIGNED flag is set.

spinlock: [Reg<SPINLOCK_SPEC>; 32]

0x100..0x180 - Reading from a spinlock address will:

Return 0 if lock is already locked

Otherwise return nonzero, and simultaneously claim the lock

Writing (any value) releases the lock.
If core 0 and core 1 attempt to claim the same lock simultaneously, core 0 wins.
The value returned on success is 0x1 << lock number.