Type Alias rp2040_pac::spi0::sspimsc::R

pub type R = R<SSPIMSC_SPEC>;

Register SSPIMSC reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn rorim(&self) -> RORIM_R

Bit 0 - Receive overrun interrupt mask: 0 Receive FIFO written to while full condition interrupt is masked. 1 Receive FIFO written to while full condition interrupt is not masked.

pub fn rtim(&self) -> RTIM_R

Bit 1 - Receive timeout interrupt mask: 0 Receive FIFO not empty and no read prior to timeout period interrupt is masked. 1 Receive FIFO not empty and no read prior to timeout period interrupt is not masked.

pub fn rxim(&self) -> RXIM_R

Bit 2 - Receive FIFO interrupt mask: 0 Receive FIFO half full or less condition interrupt is masked. 1 Receive FIFO half full or less condition interrupt is not masked.

pub fn txim(&self) -> TXIM_R

Bit 3 - Transmit FIFO interrupt mask: 0 Transmit FIFO half empty or less condition interrupt is masked. 1 Transmit FIFO half empty or less condition interrupt is not masked.