Type Alias rp2040_pac::spi0::sspimsc::W

pub type W = W<SSPIMSC_SPEC>;

Register SSPIMSC writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn rorim(&mut self) -> RORIM_W<'_, SSPIMSC_SPEC>

Bit 0 - Receive overrun interrupt mask: 0 Receive FIFO written to while full condition interrupt is masked. 1 Receive FIFO written to while full condition interrupt is not masked.

pub fn rtim(&mut self) -> RTIM_W<'_, SSPIMSC_SPEC>

Bit 1 - Receive timeout interrupt mask: 0 Receive FIFO not empty and no read prior to timeout period interrupt is masked. 1 Receive FIFO not empty and no read prior to timeout period interrupt is not masked.

pub fn rxim(&mut self) -> RXIM_W<'_, SSPIMSC_SPEC>

Bit 2 - Receive FIFO interrupt mask: 0 Receive FIFO half full or less condition interrupt is masked. 1 Receive FIFO half full or less condition interrupt is not masked.

pub fn txim(&mut self) -> TXIM_W<'_, SSPIMSC_SPEC>

Bit 3 - Transmit FIFO interrupt mask: 0 Transmit FIFO half empty or less condition interrupt is masked. 1 Transmit FIFO half empty or less condition interrupt is not masked.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual