Type Alias rp2040_pac::syscfg::dbgforce::W

pub type W = W<DBGFORCE_SPEC>;

Register DBGFORCE writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn proc0_swdi(&mut self) -> PROC0_SWDI_W<'_, DBGFORCE_SPEC>

Bit 1 - Directly drive processor 0 SWDIO input, if PROC0_ATTACH is set

pub fn proc0_swclk(&mut self) -> PROC0_SWCLK_W<'_, DBGFORCE_SPEC>

Bit 2 - Directly drive processor 0 SWCLK, if PROC0_ATTACH is set

pub fn proc0_attach(&mut self) -> PROC0_ATTACH_W<'_, DBGFORCE_SPEC>

Bit 3 - Attach processor 0 debug port to syscfg controls, and disconnect it from external SWD pads.

pub fn proc1_swdi(&mut self) -> PROC1_SWDI_W<'_, DBGFORCE_SPEC>

Bit 5 - Directly drive processor 1 SWDIO input, if PROC1_ATTACH is set

pub fn proc1_swclk(&mut self) -> PROC1_SWCLK_W<'_, DBGFORCE_SPEC>

Bit 6 - Directly drive processor 1 SWCLK, if PROC1_ATTACH is set

pub fn proc1_attach(&mut self) -> PROC1_ATTACH_W<'_, DBGFORCE_SPEC>

Bit 7 - Attach processor 1 debug port to syscfg controls, and disconnect it from external SWD pads.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual