Type Alias rp2040_pac::syscfg::proc_config::R

pub type R = R<PROC_CONFIG_SPEC>;

Register PROC_CONFIG reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn proc0_halted(&self) -> PROC0_HALTED_R

Bit 0 - Indication that proc0 has halted

pub fn proc1_halted(&self) -> PROC1_HALTED_R

Bit 1 - Indication that proc1 has halted

pub fn proc0_dap_instid(&self) -> PROC0_DAP_INSTID_R

Bits 24:27 - Configure proc0 DAP instance ID.
Recommend that this is NOT changed until you require debug access in multi-chip environment
WARNING: do not set to 15 as this is reserved for RescueDP

pub fn proc1_dap_instid(&self) -> PROC1_DAP_INSTID_R

Bits 28:31 - Configure proc1 DAP instance ID.
Recommend that this is NOT changed until you require debug access in multi-chip environment
WARNING: do not set to 15 as this is reserved for RescueDP