Module rp2040_pac::syscfg::proc_in_sync_bypass

For each bit, if 1, bypass the input synchronizer between that GPIO
and the GPIO input register in the SIO. The input synchronizers should
generally be unbypassed, to avoid injecting metastabilities into processors.
If you're feeling brave, you can bypass to save two cycles of input
latency. This register applies to GPIO 0...29.

Structs

PROC_IN_SYNC_BYPASS_W
Write proxy for field PROC_IN_SYNC_BYPASS

Type Definitions

PROC_IN_SYNC_BYPASS_R
Reader of field PROC_IN_SYNC_BYPASS

R
Reader of register PROC_IN_SYNC_BYPASS

W
Writer for register PROC_IN_SYNC_BYPASS