rp2040_pac

Module sysinfo

SYSINFO

Modules

chip_id

JEDEC JEP-106 compliant chip identifier.

gitref_rp2040

Git hash of the chip source. Used to identify chip version.

platform

Platform register. Allows software to know what environment it is running in.

Structs

RegisterBlock

Register block

Type Aliases

CHIP_ID

CHIP_ID (r) register accessor: JEDEC JEP-106 compliant chip identifier.

GITREF_RP2040

GITREF_RP2040 (r) register accessor: Git hash of the chip source. Used to identify chip version.

PLATFORM

PLATFORM (r) register accessor: Platform register. Allows software to know what environment it is running in.