rp2040_pac::timer

Struct RegisterBlock

#[repr(C)]
pub struct RegisterBlock { /* private fields */ }

Register block

Implementations

impl RegisterBlock

pub const fn timehw(&self) -> &TIMEHW

0x00 - Write to bits 63:32 of time
always write timelw before timehw

pub const fn timelw(&self) -> &TIMELW

0x04 - Write to bits 31:0 of time
writes do not get copied to time until timehw is written

pub const fn timehr(&self) -> &TIMEHR

0x08 - Read from bits 63:32 of time
always read timelr before timehr

pub const fn timelr(&self) -> &TIMELR

0x0c - Read from bits 31:0 of time

pub const fn alarm0(&self) -> &ALARM0

0x10 - Arm alarm 0, and configure the time it will fire.
Once armed, the alarm fires when TIMER_ALARM0 == TIMELR.
The alarm will disarm itself once it fires, and can
be disarmed early using the ARMED status register.

pub const fn alarm1(&self) -> &ALARM1

0x14 - Arm alarm 1, and configure the time it will fire.
Once armed, the alarm fires when TIMER_ALARM1 == TIMELR.
The alarm will disarm itself once it fires, and can
be disarmed early using the ARMED status register.

pub const fn alarm2(&self) -> &ALARM2

0x18 - Arm alarm 2, and configure the time it will fire.
Once armed, the alarm fires when TIMER_ALARM2 == TIMELR.
The alarm will disarm itself once it fires, and can
be disarmed early using the ARMED status register.

pub const fn alarm3(&self) -> &ALARM3

0x1c - Arm alarm 3, and configure the time it will fire.
Once armed, the alarm fires when TIMER_ALARM3 == TIMELR.
The alarm will disarm itself once it fires, and can
be disarmed early using the ARMED status register.

pub const fn armed(&self) -> &ARMED

0x20 - Indicates the armed/disarmed status of each alarm.
A write to the corresponding ALARMx register arms the alarm.
Alarms automatically disarm upon firing, but writing ones here
will disarm immediately without waiting to fire.

pub const fn timerawh(&self) -> &TIMERAWH

0x24 - Raw read from bits 63:32 of time (no side effects)

pub const fn timerawl(&self) -> &TIMERAWL

0x28 - Raw read from bits 31:0 of time (no side effects)

pub const fn dbgpause(&self) -> &DBGPAUSE

0x2c - Set bits high to enable pause when the corresponding debug ports are active

pub const fn pause(&self) -> &PAUSE

0x30 - Set high to pause the timer

pub const fn intr(&self) -> &INTR

0x34 - Raw Interrupts

pub const fn inte(&self) -> &INTE

0x38 - Interrupt Enable

pub const fn intf(&self) -> &INTF

0x3c - Interrupt Force

pub const fn ints(&self) -> &INTS

0x40 - Interrupt status after masking & forcing