rp2040_pac::uart0::uartimsc

Type Alias W

pub type W = W<UARTIMSC_SPEC>;

Register UARTIMSC writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn rimim(&mut self) -> RIMIM_W<'_, UARTIMSC_SPEC>

Bit 0 - nUARTRI modem interrupt mask. A read returns the current mask for the UARTRIINTR interrupt. On a write of 1, the mask of the UARTRIINTR interrupt is set. A write of 0 clears the mask.

pub fn ctsmim(&mut self) -> CTSMIM_W<'_, UARTIMSC_SPEC>

Bit 1 - nUARTCTS modem interrupt mask. A read returns the current mask for the UARTCTSINTR interrupt. On a write of 1, the mask of the UARTCTSINTR interrupt is set. A write of 0 clears the mask.

pub fn dcdmim(&mut self) -> DCDMIM_W<'_, UARTIMSC_SPEC>

Bit 2 - nUARTDCD modem interrupt mask. A read returns the current mask for the UARTDCDINTR interrupt. On a write of 1, the mask of the UARTDCDINTR interrupt is set. A write of 0 clears the mask.

pub fn dsrmim(&mut self) -> DSRMIM_W<'_, UARTIMSC_SPEC>

Bit 3 - nUARTDSR modem interrupt mask. A read returns the current mask for the UARTDSRINTR interrupt. On a write of 1, the mask of the UARTDSRINTR interrupt is set. A write of 0 clears the mask.

pub fn rxim(&mut self) -> RXIM_W<'_, UARTIMSC_SPEC>

Bit 4 - Receive interrupt mask. A read returns the current mask for the UARTRXINTR interrupt. On a write of 1, the mask of the UARTRXINTR interrupt is set. A write of 0 clears the mask.

pub fn txim(&mut self) -> TXIM_W<'_, UARTIMSC_SPEC>

Bit 5 - Transmit interrupt mask. A read returns the current mask for the UARTTXINTR interrupt. On a write of 1, the mask of the UARTTXINTR interrupt is set. A write of 0 clears the mask.

pub fn rtim(&mut self) -> RTIM_W<'_, UARTIMSC_SPEC>

Bit 6 - Receive timeout interrupt mask. A read returns the current mask for the UARTRTINTR interrupt. On a write of 1, the mask of the UARTRTINTR interrupt is set. A write of 0 clears the mask.

pub fn feim(&mut self) -> FEIM_W<'_, UARTIMSC_SPEC>

Bit 7 - Framing error interrupt mask. A read returns the current mask for the UARTFEINTR interrupt. On a write of 1, the mask of the UARTFEINTR interrupt is set. A write of 0 clears the mask.

pub fn peim(&mut self) -> PEIM_W<'_, UARTIMSC_SPEC>

Bit 8 - Parity error interrupt mask. A read returns the current mask for the UARTPEINTR interrupt. On a write of 1, the mask of the UARTPEINTR interrupt is set. A write of 0 clears the mask.

pub fn beim(&mut self) -> BEIM_W<'_, UARTIMSC_SPEC>

Bit 9 - Break error interrupt mask. A read returns the current mask for the UARTBEINTR interrupt. On a write of 1, the mask of the UARTBEINTR interrupt is set. A write of 0 clears the mask.

pub fn oeim(&mut self) -> OEIM_W<'_, UARTIMSC_SPEC>

Bit 10 - Overrun error interrupt mask. A read returns the current mask for the UARTOEINTR interrupt. On a write of 1, the mask of the UARTOEINTR interrupt is set. A write of 0 clears the mask.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual