Module rp2040_pac::usbctrl_dpram::ep_buffer_control

Structs

EP_BUFFER_CONTROL_SPEC

Enums

DOUBLE_BUFFER_ISO_OFFSET_A

The number of bytes buffer 1 is offset from buffer 0 in Isochronous mode. Only valid in double buffered mode for an Isochronous endpoint.
For a non Isochronous endpoint the offset is always 64 bytes.

Type Aliases

AVAILABLE_0_R

Field AVAILABLE_0 reader - Buffer 0 is available. This bit is set to indicate the buffer can be used by the controller. The controller clears the available bit when writing the status back.

AVAILABLE_0_W

Field AVAILABLE_0 writer - Buffer 0 is available. This bit is set to indicate the buffer can be used by the controller. The controller clears the available bit when writing the status back.

AVAILABLE_1_R

Field AVAILABLE_1 reader - Buffer 1 is available. This bit is set to indicate the buffer can be used by the controller. The controller clears the available bit when writing the status back.

AVAILABLE_1_W

Field AVAILABLE_1 writer - Buffer 1 is available. This bit is set to indicate the buffer can be used by the controller. The controller clears the available bit when writing the status back.

DOUBLE_BUFFER_ISO_OFFSET_R

Field DOUBLE_BUFFER_ISO_OFFSET reader - The number of bytes buffer 1 is offset from buffer 0 in Isochronous mode. Only valid in double buffered mode for an Isochronous endpoint.
For a non Isochronous endpoint the offset is always 64 bytes.

DOUBLE_BUFFER_ISO_OFFSET_W

Field DOUBLE_BUFFER_ISO_OFFSET writer - The number of bytes buffer 1 is offset from buffer 0 in Isochronous mode. Only valid in double buffered mode for an Isochronous endpoint.
For a non Isochronous endpoint the offset is always 64 bytes.

FULL_0_R

Field FULL_0 reader - Buffer 0 is full. For an IN transfer (TX to the host) the bit is set to indicate the data is valid. For an OUT transfer (RX from the host) this bit should be left as a 0. The host will set it when it has filled the buffer with data.

FULL_0_W

Field FULL_0 writer - Buffer 0 is full. For an IN transfer (TX to the host) the bit is set to indicate the data is valid. For an OUT transfer (RX from the host) this bit should be left as a 0. The host will set it when it has filled the buffer with data.

FULL_1_R

Field FULL_1 reader - Buffer 1 is full. For an IN transfer (TX to the host) the bit is set to indicate the data is valid. For an OUT transfer (RX from the host) this bit should be left as a 0. The host will set it when it has filled the buffer with data.

FULL_1_W

Field FULL_1 writer - Buffer 1 is full. For an IN transfer (TX to the host) the bit is set to indicate the data is valid. For an OUT transfer (RX from the host) this bit should be left as a 0. The host will set it when it has filled the buffer with data.

LAST_0_R

Field LAST_0 reader - Buffer 0 is the last buffer of the transfer.

LAST_0_W

Field LAST_0 writer - Buffer 0 is the last buffer of the transfer.

LAST_1_R

Field LAST_1 reader - Buffer 1 is the last buffer of the transfer.

LAST_1_W

Field LAST_1 writer - Buffer 1 is the last buffer of the transfer.

LENGTH_0_R

Field LENGTH_0 reader - The length of the data in buffer 0.

LENGTH_0_W

Field LENGTH_0 writer - The length of the data in buffer 0.

LENGTH_1_R

Field LENGTH_1 reader - The length of the data in buffer 1.

LENGTH_1_W

Field LENGTH_1 writer - The length of the data in buffer 1.

PID_0_R

Field PID_0 reader - The data pid of buffer 0.

PID_0_W

Field PID_0 writer - The data pid of buffer 0.

PID_1_R

Field PID_1 reader - The data pid of buffer 1.

PID_1_W

Field PID_1 writer - The data pid of buffer 1.

R

Register EP_BUFFER_CONTROL%s reader

RESET_R

Field RESET reader - Reset the buffer selector to buffer 0.

RESET_W

Field RESET writer - Reset the buffer selector to buffer 0.

STALL_R

Field STALL reader - Reply with a stall (valid for both buffers).

STALL_W

Field STALL writer - Reply with a stall (valid for both buffers).

W

Register EP_BUFFER_CONTROL%s writer