Type Alias rp2040_pac::usbctrl_dpram::ep_buffer_control::DOUBLE_BUFFER_ISO_OFFSET_R

pub type DOUBLE_BUFFER_ISO_OFFSET_R = FieldReader<DOUBLE_BUFFER_ISO_OFFSET_A>;

Field DOUBLE_BUFFER_ISO_OFFSET reader - The number of bytes buffer 1 is offset from buffer 0 in Isochronous mode. Only valid in double buffered mode for an Isochronous endpoint.
For a non Isochronous endpoint the offset is always 64 bytes.

Aliased Type

struct DOUBLE_BUFFER_ISO_OFFSET_R { /* private fields */ }

Implementations

impl DOUBLE_BUFFER_ISO_OFFSET_R

pub const fn variant(&self) -> DOUBLE_BUFFER_ISO_OFFSET_A

Get enumerated values variant

pub fn is_128(&self) -> bool

0

pub fn is_256(&self) -> bool

1

pub fn is_512(&self) -> bool

10

pub fn is_1024(&self) -> bool

11