rp2040_pac::usbctrl_dpram::ep_buffer_control

Type Alias R

pub type R = R<EP_BUFFER_CONTROL_SPEC>;

Register EP_BUFFER_CONTROL%s reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn length_0(&self) -> LENGTH_0_R

Bits 0:9 - The length of the data in buffer 0.

pub fn available_0(&self) -> AVAILABLE_0_R

Bit 10 - Buffer 0 is available. This bit is set to indicate the buffer can be used by the controller. The controller clears the available bit when writing the status back.

pub fn stall(&self) -> STALL_R

Bit 11 - Reply with a stall (valid for both buffers).

pub fn reset(&self) -> RESET_R

Bit 12 - Reset the buffer selector to buffer 0.

pub fn pid_0(&self) -> PID_0_R

Bit 13 - The data pid of buffer 0.

pub fn last_0(&self) -> LAST_0_R

Bit 14 - Buffer 0 is the last buffer of the transfer.

pub fn full_0(&self) -> FULL_0_R

Bit 15 - Buffer 0 is full. For an IN transfer (TX to the host) the bit is set to indicate the data is valid. For an OUT transfer (RX from the host) this bit should be left as a 0. The host will set it when it has filled the buffer with data.

pub fn length_1(&self) -> LENGTH_1_R

Bits 16:25 - The length of the data in buffer 1.

pub fn available_1(&self) -> AVAILABLE_1_R

Bit 26 - Buffer 1 is available. This bit is set to indicate the buffer can be used by the controller. The controller clears the available bit when writing the status back.

pub fn double_buffer_iso_offset(&self) -> DOUBLE_BUFFER_ISO_OFFSET_R

Bits 27:28 - The number of bytes buffer 1 is offset from buffer 0 in Isochronous mode. Only valid in double buffered mode for an Isochronous endpoint.
For a non Isochronous endpoint the offset is always 64 bytes.

pub fn pid_1(&self) -> PID_1_R

Bit 29 - The data pid of buffer 1.

pub fn last_1(&self) -> LAST_1_R

Bit 30 - Buffer 1 is the last buffer of the transfer.

pub fn full_1(&self) -> FULL_1_R

Bit 31 - Buffer 1 is full. For an IN transfer (TX to the host) the bit is set to indicate the data is valid. For an OUT transfer (RX from the host) this bit should be left as a 0. The host will set it when it has filled the buffer with data.