rp2040_pac::usbctrl_dpram::ep_control

Type Alias R

pub type R = R<EP_CONTROL_SPEC>;

Register EP_CONTROL%s reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn buffer_address(&self) -> BUFFER_ADDRESS_R

Bits 0:15 - 64 byte aligned buffer address for this EP (bits 0-5 are ignored). Relative to the start of the DPRAM.

pub fn interrupt_on_nak(&self) -> INTERRUPT_ON_NAK_R

Bit 16 - Trigger an interrupt if a NAK is sent. Intended for debug only.

pub fn host_poll_interval(&self) -> HOST_POLL_INTERVAL_R

Bits 16:25 - The interval the host controller should poll this endpoint. Only applicable for interrupt endpoints. Specified in ms - 1. For example: a value of 9 would poll the endpoint every 10ms.

pub fn interrupt_on_stall(&self) -> INTERRUPT_ON_STALL_R

Bit 17 - Trigger an interrupt if a STALL is sent. Intended for debug only.

pub fn endpoint_type(&self) -> ENDPOINT_TYPE_R

Bits 26:27

pub fn interrupt_per_double_buff(&self) -> INTERRUPT_PER_DOUBLE_BUFF_R

Bit 28 - Trigger an interrupt each time both buffers are done. Only valid in double buffered mode.

pub fn interrupt_per_buff(&self) -> INTERRUPT_PER_BUFF_R

Bit 29 - Trigger an interrupt each time a buffer is done.

pub fn double_buffered(&self) -> DOUBLE_BUFFERED_R

Bit 30 - This endpoint is double buffered.

pub fn enable(&self) -> ENABLE_R

Bit 31 - Enable this endpoint. The device will not reply to any packets for this endpoint if this bit is not set.