Module rp2040_pac::usbctrl_dpram::epx_control

EPx Control (Host-mode only!)

Structs

EPX_CONTROL_SPEC

EPx Control (Host-mode only!)

Enums

ENDPOINT_TYPE_A

Value on reset: 0

Type Aliases

BUFFER_ADDRESS_R

Field BUFFER_ADDRESS reader - 64 byte aligned buffer address for this EP (bits 0-5 are ignored). Relative to the start of the DPRAM.

BUFFER_ADDRESS_W

Field BUFFER_ADDRESS writer - 64 byte aligned buffer address for this EP (bits 0-5 are ignored). Relative to the start of the DPRAM.

DOUBLE_BUFFERED_R

Field DOUBLE_BUFFERED reader - This endpoint is double buffered.

DOUBLE_BUFFERED_W

Field DOUBLE_BUFFERED writer - This endpoint is double buffered.

ENABLE_R

Field ENABLE reader - Enable this endpoint. The device will not reply to any packets for this endpoint if this bit is not set.

ENABLE_W

Field ENABLE writer - Enable this endpoint. The device will not reply to any packets for this endpoint if this bit is not set.

ENDPOINT_TYPE_R

Field ENDPOINT_TYPE reader -

ENDPOINT_TYPE_W

Field ENDPOINT_TYPE writer -

INTERRUPT_ON_NAK_R

Field INTERRUPT_ON_NAK reader - Trigger an interrupt if a NAK is sent. Intended for debug only.

INTERRUPT_ON_NAK_W

Field INTERRUPT_ON_NAK writer - Trigger an interrupt if a NAK is sent. Intended for debug only.

INTERRUPT_ON_STALL_R

Field INTERRUPT_ON_STALL reader - Trigger an interrupt if a STALL is sent. Intended for debug only.

INTERRUPT_ON_STALL_W

Field INTERRUPT_ON_STALL writer - Trigger an interrupt if a STALL is sent. Intended for debug only.

INTERRUPT_PER_BUFF_R

Field INTERRUPT_PER_BUFF reader - Trigger an interrupt each time a buffer is done.

INTERRUPT_PER_BUFF_W

Field INTERRUPT_PER_BUFF writer - Trigger an interrupt each time a buffer is done.

INTERRUPT_PER_DOUBLE_BUFF_R

Field INTERRUPT_PER_DOUBLE_BUFF reader - Trigger an interrupt each time both buffers are done. Only valid in double buffered mode.

INTERRUPT_PER_DOUBLE_BUFF_W

Field INTERRUPT_PER_DOUBLE_BUFF writer - Trigger an interrupt each time both buffers are done. Only valid in double buffered mode.

R

Register EPX_CONTROL reader

W

Register EPX_CONTROL writer