Module rp2040_pac::usbctrl_regs::ep_stall_arm

Device: this bit must be set in conjunction with the STALL bit in the buffer control register to send a STALL on EP0. The device controller clears these bits when a SETUP packet is received because the USB spec requires that a STALL condition is cleared when a SETUP packet is received.

Structs

EP0_IN_R

Field EP0_IN reader -

EP0_IN_W

Field EP0_IN writer -

EP0_OUT_R

Field EP0_OUT reader -

EP0_OUT_W

Field EP0_OUT writer -

EP_STALL_ARM_SPEC

Device: this bit must be set in conjunction with the STALL bit in the buffer control register to send a STALL on EP0. The device controller clears these bits when a SETUP packet is received because the USB spec requires that a STALL condition is cleared when a SETUP packet is received.

R

Register EP_STALL_ARM reader

W

Register EP_STALL_ARM writer