rp2040_pac::usbctrl_regs
Module host_addr_endp

Interrupt endpoints. Only valid in HOST mode.

Structs

HOST_ADDR_ENDP_SPEC

Interrupt endpoints. Only valid in HOST mode.

Type Aliases

ADDRESS_R

Field ADDRESS reader - Device address

ADDRESS_W

Field ADDRESS writer - Device address

ENDPOINT_R

Field ENDPOINT reader - Endpoint number of the interrupt endpoint

ENDPOINT_W

Field ENDPOINT writer - Endpoint number of the interrupt endpoint

INTEP_DIR_R

Field INTEP_DIR reader - Direction of the interrupt endpoint. In=0, Out=1

INTEP_DIR_W

Field INTEP_DIR writer - Direction of the interrupt endpoint. In=0, Out=1

INTEP_PREAMBLE_R

Field INTEP_PREAMBLE reader - Interrupt EP requires preamble (is a low speed device on a full speed hub)

INTEP_PREAMBLE_W

Field INTEP_PREAMBLE writer - Interrupt EP requires preamble (is a low speed device on a full speed hub)

R

Register HOST_ADDR_ENDP%s reader

W

Register HOST_ADDR_ENDP%s writer