Module rp2040_pac::usbctrl_regs::inte

Interrupt Enable

Structs

ABORT_DONE_R

Field ABORT_DONE reader - Raised when any bit in ABORT_DONE is set. Clear by clearing all bits in ABORT_DONE.

ABORT_DONE_W

Field ABORT_DONE writer - Raised when any bit in ABORT_DONE is set. Clear by clearing all bits in ABORT_DONE.

BUFF_STATUS_R

Field BUFF_STATUS reader - Raised when any bit in BUFF_STATUS is set. Clear by clearing all bits in BUFF_STATUS.

BUFF_STATUS_W

Field BUFF_STATUS writer - Raised when any bit in BUFF_STATUS is set. Clear by clearing all bits in BUFF_STATUS.

BUS_RESET_R

Field BUS_RESET reader - Source: SIE_STATUS.BUS_RESET

BUS_RESET_W

Field BUS_RESET writer - Source: SIE_STATUS.BUS_RESET

DEV_CONN_DIS_R

Field DEV_CONN_DIS reader - Set when the device connection state changes. Cleared by writing to SIE_STATUS.CONNECTED

DEV_CONN_DIS_W

Field DEV_CONN_DIS writer - Set when the device connection state changes. Cleared by writing to SIE_STATUS.CONNECTED

DEV_RESUME_FROM_HOST_R

Field DEV_RESUME_FROM_HOST reader - Set when the device receives a resume from the host. Cleared by writing to SIE_STATUS.RESUME

DEV_RESUME_FROM_HOST_W

Field DEV_RESUME_FROM_HOST writer - Set when the device receives a resume from the host. Cleared by writing to SIE_STATUS.RESUME

DEV_SOF_R

Field DEV_SOF reader - Set every time the device receives a SOF (Start of Frame) packet. Cleared by reading SOF_RD

DEV_SOF_W

Field DEV_SOF writer - Set every time the device receives a SOF (Start of Frame) packet. Cleared by reading SOF_RD

DEV_SUSPEND_R

Field DEV_SUSPEND reader - Set when the device suspend state changes. Cleared by writing to SIE_STATUS.SUSPENDED

DEV_SUSPEND_W

Field DEV_SUSPEND writer - Set when the device suspend state changes. Cleared by writing to SIE_STATUS.SUSPENDED

EP_STALL_NAK_R

Field EP_STALL_NAK reader - Raised when any bit in EP_STATUS_STALL_NAK is set. Clear by clearing all bits in EP_STATUS_STALL_NAK.

EP_STALL_NAK_W

Field EP_STALL_NAK writer - Raised when any bit in EP_STATUS_STALL_NAK is set. Clear by clearing all bits in EP_STATUS_STALL_NAK.

ERROR_BIT_STUFF_R

Field ERROR_BIT_STUFF reader - Source: SIE_STATUS.BIT_STUFF_ERROR

ERROR_BIT_STUFF_W

Field ERROR_BIT_STUFF writer - Source: SIE_STATUS.BIT_STUFF_ERROR

ERROR_CRC_R

Field ERROR_CRC reader - Source: SIE_STATUS.CRC_ERROR

ERROR_CRC_W

Field ERROR_CRC writer - Source: SIE_STATUS.CRC_ERROR

ERROR_DATA_SEQ_R

Field ERROR_DATA_SEQ reader - Source: SIE_STATUS.DATA_SEQ_ERROR

ERROR_DATA_SEQ_W

Field ERROR_DATA_SEQ writer - Source: SIE_STATUS.DATA_SEQ_ERROR

ERROR_RX_OVERFLOW_R

Field ERROR_RX_OVERFLOW reader - Source: SIE_STATUS.RX_OVERFLOW

ERROR_RX_OVERFLOW_W

Field ERROR_RX_OVERFLOW writer - Source: SIE_STATUS.RX_OVERFLOW

ERROR_RX_TIMEOUT_R

Field ERROR_RX_TIMEOUT reader - Source: SIE_STATUS.RX_TIMEOUT

ERROR_RX_TIMEOUT_W

Field ERROR_RX_TIMEOUT writer - Source: SIE_STATUS.RX_TIMEOUT

HOST_CONN_DIS_R

Field HOST_CONN_DIS reader - Host: raised when a device is connected or disconnected (i.e. when SIE_STATUS.SPEED changes). Cleared by writing to SIE_STATUS.SPEED

HOST_CONN_DIS_W

Field HOST_CONN_DIS writer - Host: raised when a device is connected or disconnected (i.e. when SIE_STATUS.SPEED changes). Cleared by writing to SIE_STATUS.SPEED

HOST_RESUME_R

Field HOST_RESUME reader - Host: raised when a device wakes up the host. Cleared by writing to SIE_STATUS.RESUME

HOST_RESUME_W

Field HOST_RESUME writer - Host: raised when a device wakes up the host. Cleared by writing to SIE_STATUS.RESUME

HOST_SOF_R

Field HOST_SOF reader - Host: raised every time the host sends a SOF (Start of Frame). Cleared by reading SOF_RD

HOST_SOF_W

Field HOST_SOF writer - Host: raised every time the host sends a SOF (Start of Frame). Cleared by reading SOF_RD

INTE_SPEC

Interrupt Enable

R

Register INTE reader

SETUP_REQ_R

Field SETUP_REQ reader - Device. Source: SIE_STATUS.SETUP_REC

SETUP_REQ_W

Field SETUP_REQ writer - Device. Source: SIE_STATUS.SETUP_REC

STALL_R

Field STALL reader - Source: SIE_STATUS.STALL_REC

STALL_W

Field STALL writer - Source: SIE_STATUS.STALL_REC

TRANS_COMPLETE_R

Field TRANS_COMPLETE reader - Raised every time SIE_STATUS.TRANS_COMPLETE is set. Clear by writing to this bit.

TRANS_COMPLETE_W

Field TRANS_COMPLETE writer - Raised every time SIE_STATUS.TRANS_COMPLETE is set. Clear by writing to this bit.

VBUS_DETECT_R

Field VBUS_DETECT reader - Source: SIE_STATUS.VBUS_DETECTED

VBUS_DETECT_W

Field VBUS_DETECT writer - Source: SIE_STATUS.VBUS_DETECTED

W

Register INTE writer