rp2040_pac::usbctrl_regs::inte

Type Alias W

pub type W = W<INTE_SPEC>;

Register INTE writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn host_conn_dis(&mut self) -> HOST_CONN_DIS_W<'_, INTE_SPEC>

Bit 0 - Host: raised when a device is connected or disconnected (i.e. when SIE_STATUS.SPEED changes). Cleared by writing to SIE_STATUS.SPEED

pub fn host_resume(&mut self) -> HOST_RESUME_W<'_, INTE_SPEC>

Bit 1 - Host: raised when a device wakes up the host. Cleared by writing to SIE_STATUS.RESUME

pub fn host_sof(&mut self) -> HOST_SOF_W<'_, INTE_SPEC>

Bit 2 - Host: raised every time the host sends a SOF (Start of Frame). Cleared by reading SOF_RD

pub fn trans_complete(&mut self) -> TRANS_COMPLETE_W<'_, INTE_SPEC>

Bit 3 - Raised every time SIE_STATUS.TRANS_COMPLETE is set. Clear by writing to this bit.

pub fn buff_status(&mut self) -> BUFF_STATUS_W<'_, INTE_SPEC>

Bit 4 - Raised when any bit in BUFF_STATUS is set. Clear by clearing all bits in BUFF_STATUS.

pub fn error_data_seq(&mut self) -> ERROR_DATA_SEQ_W<'_, INTE_SPEC>

Bit 5 - Source: SIE_STATUS.DATA_SEQ_ERROR

pub fn error_rx_timeout(&mut self) -> ERROR_RX_TIMEOUT_W<'_, INTE_SPEC>

Bit 6 - Source: SIE_STATUS.RX_TIMEOUT

pub fn error_rx_overflow(&mut self) -> ERROR_RX_OVERFLOW_W<'_, INTE_SPEC>

Bit 7 - Source: SIE_STATUS.RX_OVERFLOW

pub fn error_bit_stuff(&mut self) -> ERROR_BIT_STUFF_W<'_, INTE_SPEC>

Bit 8 - Source: SIE_STATUS.BIT_STUFF_ERROR

pub fn error_crc(&mut self) -> ERROR_CRC_W<'_, INTE_SPEC>

Bit 9 - Source: SIE_STATUS.CRC_ERROR

pub fn stall(&mut self) -> STALL_W<'_, INTE_SPEC>

Bit 10 - Source: SIE_STATUS.STALL_REC

pub fn vbus_detect(&mut self) -> VBUS_DETECT_W<'_, INTE_SPEC>

Bit 11 - Source: SIE_STATUS.VBUS_DETECTED

pub fn bus_reset(&mut self) -> BUS_RESET_W<'_, INTE_SPEC>

Bit 12 - Source: SIE_STATUS.BUS_RESET

pub fn dev_conn_dis(&mut self) -> DEV_CONN_DIS_W<'_, INTE_SPEC>

Bit 13 - Set when the device connection state changes. Cleared by writing to SIE_STATUS.CONNECTED

pub fn dev_suspend(&mut self) -> DEV_SUSPEND_W<'_, INTE_SPEC>

Bit 14 - Set when the device suspend state changes. Cleared by writing to SIE_STATUS.SUSPENDED

pub fn dev_resume_from_host(&mut self) -> DEV_RESUME_FROM_HOST_W<'_, INTE_SPEC>

Bit 15 - Set when the device receives a resume from the host. Cleared by writing to SIE_STATUS.RESUME

pub fn setup_req(&mut self) -> SETUP_REQ_W<'_, INTE_SPEC>

Bit 16 - Device. Source: SIE_STATUS.SETUP_REC

pub fn dev_sof(&mut self) -> DEV_SOF_W<'_, INTE_SPEC>

Bit 17 - Set every time the device receives a SOF (Start of Frame) packet. Cleared by reading SOF_RD

pub fn abort_done(&mut self) -> ABORT_DONE_W<'_, INTE_SPEC>

Bit 18 - Raised when any bit in ABORT_DONE is set. Clear by clearing all bits in ABORT_DONE.

pub fn ep_stall_nak(&mut self) -> EP_STALL_NAK_W<'_, INTE_SPEC>

Bit 19 - Raised when any bit in EP_STATUS_STALL_NAK is set. Clear by clearing all bits in EP_STATUS_STALL_NAK.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual