rp2040_pac::usbctrl_regs::intr

Type Alias R

pub type R = R<INTR_SPEC>;

Register INTR reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn host_conn_dis(&self) -> HOST_CONN_DIS_R

Bit 0 - Host: raised when a device is connected or disconnected (i.e. when SIE_STATUS.SPEED changes). Cleared by writing to SIE_STATUS.SPEED

pub fn host_resume(&self) -> HOST_RESUME_R

Bit 1 - Host: raised when a device wakes up the host. Cleared by writing to SIE_STATUS.RESUME

pub fn host_sof(&self) -> HOST_SOF_R

Bit 2 - Host: raised every time the host sends a SOF (Start of Frame). Cleared by reading SOF_RD

pub fn trans_complete(&self) -> TRANS_COMPLETE_R

Bit 3 - Raised every time SIE_STATUS.TRANS_COMPLETE is set. Clear by writing to this bit.

pub fn buff_status(&self) -> BUFF_STATUS_R

Bit 4 - Raised when any bit in BUFF_STATUS is set. Clear by clearing all bits in BUFF_STATUS.

pub fn error_data_seq(&self) -> ERROR_DATA_SEQ_R

Bit 5 - Source: SIE_STATUS.DATA_SEQ_ERROR

pub fn error_rx_timeout(&self) -> ERROR_RX_TIMEOUT_R

Bit 6 - Source: SIE_STATUS.RX_TIMEOUT

pub fn error_rx_overflow(&self) -> ERROR_RX_OVERFLOW_R

Bit 7 - Source: SIE_STATUS.RX_OVERFLOW

pub fn error_bit_stuff(&self) -> ERROR_BIT_STUFF_R

Bit 8 - Source: SIE_STATUS.BIT_STUFF_ERROR

pub fn error_crc(&self) -> ERROR_CRC_R

Bit 9 - Source: SIE_STATUS.CRC_ERROR

pub fn stall(&self) -> STALL_R

Bit 10 - Source: SIE_STATUS.STALL_REC

pub fn vbus_detect(&self) -> VBUS_DETECT_R

Bit 11 - Source: SIE_STATUS.VBUS_DETECTED

pub fn bus_reset(&self) -> BUS_RESET_R

Bit 12 - Source: SIE_STATUS.BUS_RESET

pub fn dev_conn_dis(&self) -> DEV_CONN_DIS_R

Bit 13 - Set when the device connection state changes. Cleared by writing to SIE_STATUS.CONNECTED

pub fn dev_suspend(&self) -> DEV_SUSPEND_R

Bit 14 - Set when the device suspend state changes. Cleared by writing to SIE_STATUS.SUSPENDED

pub fn dev_resume_from_host(&self) -> DEV_RESUME_FROM_HOST_R

Bit 15 - Set when the device receives a resume from the host. Cleared by writing to SIE_STATUS.RESUME

pub fn setup_req(&self) -> SETUP_REQ_R

Bit 16 - Device. Source: SIE_STATUS.SETUP_REC

pub fn dev_sof(&self) -> DEV_SOF_R

Bit 17 - Set every time the device receives a SOF (Start of Frame) packet. Cleared by reading SOF_RD

pub fn abort_done(&self) -> ABORT_DONE_R

Bit 18 - Raised when any bit in ABORT_DONE is set. Clear by clearing all bits in ABORT_DONE.

pub fn ep_stall_nak(&self) -> EP_STALL_NAK_R

Bit 19 - Raised when any bit in EP_STATUS_STALL_NAK is set. Clear by clearing all bits in EP_STATUS_STALL_NAK.