Struct rp2040_pac::usbctrl_regs::RegisterBlock

#[repr(C)]
pub struct RegisterBlock {

    pub addr_endp: Reg<ADDR_ENDP_SPEC>,
    pub addr_endp1: Reg<ADDR_ENDP1_SPEC>,
    pub addr_endp2: Reg<ADDR_ENDP2_SPEC>,
    pub addr_endp3: Reg<ADDR_ENDP3_SPEC>,
    pub addr_endp4: Reg<ADDR_ENDP4_SPEC>,
    pub addr_endp5: Reg<ADDR_ENDP5_SPEC>,
    pub addr_endp6: Reg<ADDR_ENDP6_SPEC>,
    pub addr_endp7: Reg<ADDR_ENDP7_SPEC>,
    pub addr_endp8: Reg<ADDR_ENDP8_SPEC>,
    pub addr_endp9: Reg<ADDR_ENDP9_SPEC>,
    pub addr_endp10: Reg<ADDR_ENDP10_SPEC>,
    pub addr_endp11: Reg<ADDR_ENDP11_SPEC>,
    pub addr_endp12: Reg<ADDR_ENDP12_SPEC>,
    pub addr_endp13: Reg<ADDR_ENDP13_SPEC>,
    pub addr_endp14: Reg<ADDR_ENDP14_SPEC>,
    pub addr_endp15: Reg<ADDR_ENDP15_SPEC>,
    pub main_ctrl: Reg<MAIN_CTRL_SPEC>,
    pub sof_wr: Reg<SOF_WR_SPEC>,
    pub sof_rd: Reg<SOF_RD_SPEC>,
    pub sie_ctrl: Reg<SIE_CTRL_SPEC>,
    pub sie_status: Reg<SIE_STATUS_SPEC>,
    pub int_ep_ctrl: Reg<INT_EP_CTRL_SPEC>,
    pub buff_status: Reg<BUFF_STATUS_SPEC>,
    pub buff_cpu_should_handle: Reg<BUFF_CPU_SHOULD_HANDLE_SPEC>,
    pub ep_abort: Reg<EP_ABORT_SPEC>,
    pub ep_abort_done: Reg<EP_ABORT_DONE_SPEC>,
    pub ep_stall_arm: Reg<EP_STALL_ARM_SPEC>,
    pub nak_poll: Reg<NAK_POLL_SPEC>,
    pub ep_status_stall_nak: Reg<EP_STATUS_STALL_NAK_SPEC>,
    pub usb_muxing: Reg<USB_MUXING_SPEC>,
    pub usb_pwr: Reg<USB_PWR_SPEC>,
    pub usbphy_direct: Reg<USBPHY_DIRECT_SPEC>,
    pub usbphy_direct_override: Reg<USBPHY_DIRECT_OVERRIDE_SPEC>,
    pub usbphy_trim: Reg<USBPHY_TRIM_SPEC>,
    pub intr: Reg<INTR_SPEC>,
    pub inte: Reg<INTE_SPEC>,
    pub intf: Reg<INTF_SPEC>,
    pub ints: Reg<INTS_SPEC>,
    /* private fields */

}

Register block

Fields
addr_endp: Reg<ADDR_ENDP_SPEC>

0x00 - Device address and endpoint control

addr_endp1: Reg<ADDR_ENDP1_SPEC>

0x04 - Interrupt endpoint 1. Only valid for HOST mode.

addr_endp2: Reg<ADDR_ENDP2_SPEC>

0x08 - Interrupt endpoint 2. Only valid for HOST mode.

addr_endp3: Reg<ADDR_ENDP3_SPEC>

0x0c - Interrupt endpoint 3. Only valid for HOST mode.

addr_endp4: Reg<ADDR_ENDP4_SPEC>

0x10 - Interrupt endpoint 4. Only valid for HOST mode.

addr_endp5: Reg<ADDR_ENDP5_SPEC>

0x14 - Interrupt endpoint 5. Only valid for HOST mode.

addr_endp6: Reg<ADDR_ENDP6_SPEC>

0x18 - Interrupt endpoint 6. Only valid for HOST mode.

addr_endp7: Reg<ADDR_ENDP7_SPEC>

0x1c - Interrupt endpoint 7. Only valid for HOST mode.

addr_endp8: Reg<ADDR_ENDP8_SPEC>

0x20 - Interrupt endpoint 8. Only valid for HOST mode.

addr_endp9: Reg<ADDR_ENDP9_SPEC>

0x24 - Interrupt endpoint 9. Only valid for HOST mode.

addr_endp10: Reg<ADDR_ENDP10_SPEC>

0x28 - Interrupt endpoint 10. Only valid for HOST mode.

addr_endp11: Reg<ADDR_ENDP11_SPEC>

0x2c - Interrupt endpoint 11. Only valid for HOST mode.

addr_endp12: Reg<ADDR_ENDP12_SPEC>

0x30 - Interrupt endpoint 12. Only valid for HOST mode.

addr_endp13: Reg<ADDR_ENDP13_SPEC>

0x34 - Interrupt endpoint 13. Only valid for HOST mode.

addr_endp14: Reg<ADDR_ENDP14_SPEC>

0x38 - Interrupt endpoint 14. Only valid for HOST mode.

addr_endp15: Reg<ADDR_ENDP15_SPEC>

0x3c - Interrupt endpoint 15. Only valid for HOST mode.

main_ctrl: Reg<MAIN_CTRL_SPEC>

0x40 - Main control register

sof_wr: Reg<SOF_WR_SPEC>

0x44 - Set the SOF (Start of Frame) frame number in the host controller. The SOF packet is sent every 1ms and the host will increment the frame number by 1 each time.

sof_rd: Reg<SOF_RD_SPEC>

0x48 - Read the last SOF (Start of Frame) frame number seen. In device mode the last SOF received from the host. In host mode the last SOF sent by the host.

sie_ctrl: Reg<SIE_CTRL_SPEC>

0x4c - SIE control register

sie_status: Reg<SIE_STATUS_SPEC>

0x50 - SIE status register

int_ep_ctrl: Reg<INT_EP_CTRL_SPEC>

0x54 - interrupt endpoint control register

buff_status: Reg<BUFF_STATUS_SPEC>

0x58 - Buffer status register. A bit set here indicates that a buffer has completed on the endpoint (if the buffer interrupt is enabled). It is possible for 2 buffers to be completed, so clearing the buffer status bit may instantly re set it on the next clock cycle.

buff_cpu_should_handle: Reg<BUFF_CPU_SHOULD_HANDLE_SPEC>

0x5c - Which of the double buffers should be handled. Only valid if using an interrupt per buffer (i.e. not per 2 buffers). Not valid for host interrupt endpoint polling because they are only single buffered.

ep_abort: Reg<EP_ABORT_SPEC>

0x60 - Device only: Can be set to ignore the buffer control register for this endpoint in case you would like to revoke a buffer. A NAK will be sent for every access to the endpoint until this bit is cleared. A corresponding bit in EP_ABORT_DONE is set when it is safe to modify the buffer control register.

ep_abort_done: Reg<EP_ABORT_DONE_SPEC>

0x64 - Device only: Used in conjunction with EP_ABORT. Set once an endpoint is idle so the programmer knows it is safe to modify the buffer control register.

ep_stall_arm: Reg<EP_STALL_ARM_SPEC>

0x68 - Device: this bit must be set in conjunction with the STALL bit in the buffer control register to send a STALL on EP0. The device controller clears these bits when a SETUP packet is received because the USB spec requires that a STALL condition is cleared when a SETUP packet is received.

nak_poll: Reg<NAK_POLL_SPEC>

0x6c - Used by the host controller. Sets the wait time in microseconds before trying again if the device replies with a NAK.

ep_status_stall_nak: Reg<EP_STATUS_STALL_NAK_SPEC>

0x70 - Device: bits are set when the IRQ_ON_NAK or IRQ_ON_STALL bits are set. For EP0 this comes from SIE_CTRL. For all other endpoints it comes from the endpoint control register.

usb_muxing: Reg<USB_MUXING_SPEC>

0x74 - Where to connect the USB controller. Should be to_phy by default.

usb_pwr: Reg<USB_PWR_SPEC>

0x78 - Overrides for the power signals in the event that the VBUS signals are not hooked up to GPIO. Set the value of the override and then the override enable to switch over to the override value.

usbphy_direct: Reg<USBPHY_DIRECT_SPEC>

0x7c - This register allows for direct control of the USB phy. Use in conjunction with usbphy_direct_override register to enable each override bit.

usbphy_direct_override: Reg<USBPHY_DIRECT_OVERRIDE_SPEC>

0x80 - Override enable for each control in usbphy_direct

usbphy_trim: Reg<USBPHY_TRIM_SPEC>

0x84 - Used to adjust trim values of USB phy pull down resistors.

intr: Reg<INTR_SPEC>

0x8c - Raw Interrupts

inte: Reg<INTE_SPEC>

0x90 - Interrupt Enable

intf: Reg<INTF_SPEC>

0x94 - Interrupt Force

ints: Reg<INTS_SPEC>

0x98 - Interrupt status after masking & forcing