Module rp2040_pac::usbctrl_regs::usb_pwr

Overrides for the power signals in the event that the VBUS signals are not hooked up to GPIO. Set the value of the override and then the override enable to switch over to the override value.

Structs

OVERCURR_DETECT_EN_R

Field OVERCURR_DETECT_EN reader -

OVERCURR_DETECT_EN_W

Field OVERCURR_DETECT_EN writer -

OVERCURR_DETECT_R

Field OVERCURR_DETECT reader -

OVERCURR_DETECT_W

Field OVERCURR_DETECT writer -

R

Register USB_PWR reader

USB_PWR_SPEC

Overrides for the power signals in the event that the VBUS signals are not hooked up to GPIO. Set the value of the override and then the override enable to switch over to the override value.

VBUS_DETECT_OVERRIDE_EN_R

Field VBUS_DETECT_OVERRIDE_EN reader -

VBUS_DETECT_OVERRIDE_EN_W

Field VBUS_DETECT_OVERRIDE_EN writer -

VBUS_DETECT_R

Field VBUS_DETECT reader -

VBUS_DETECT_W

Field VBUS_DETECT writer -

VBUS_EN_OVERRIDE_EN_R

Field VBUS_EN_OVERRIDE_EN reader -

VBUS_EN_OVERRIDE_EN_W

Field VBUS_EN_OVERRIDE_EN writer -

VBUS_EN_R

Field VBUS_EN reader -

VBUS_EN_W

Field VBUS_EN writer -

W

Register USB_PWR writer