rp2040_pac::usbctrl_regs
Module usbphy_trim

Used to adjust trim values of USB phy pull down resistors.

Structs

USBPHY_TRIM_SPEC

Used to adjust trim values of USB phy pull down resistors.

Type Aliases

DM_PULLDN_TRIM_R

Field DM_PULLDN_TRIM reader - Value to drive to USB PHY
DM pulldown resistor trim control
Experimental data suggests that the reset value will work, but this register allows adjustment if required

DM_PULLDN_TRIM_W

Field DM_PULLDN_TRIM writer - Value to drive to USB PHY
DM pulldown resistor trim control
Experimental data suggests that the reset value will work, but this register allows adjustment if required

DP_PULLDN_TRIM_R

Field DP_PULLDN_TRIM reader - Value to drive to USB PHY
DP pulldown resistor trim control
Experimental data suggests that the reset value will work, but this register allows adjustment if required

DP_PULLDN_TRIM_W

Field DP_PULLDN_TRIM writer - Value to drive to USB PHY
DP pulldown resistor trim control
Experimental data suggests that the reset value will work, but this register allows adjustment if required

R

Register USBPHY_TRIM reader

W

Register USBPHY_TRIM writer