Module rp2040_pac::xosc

Controls the crystal oscillator

Modules

ctrl

Crystal Oscillator Control

dormant

Crystal Oscillator pause control
This is used to save power by pausing the XOSC
On power-up this field is initialised to WAKE
An invalid write will also select WAKE
WARNING: stop the PLLs before selecting dormant mode
WARNING: setup the irq before selecting dormant mode

startup

Controls the startup delay

status

Crystal Oscillator Status

Structs

RegisterBlock

Register block

Type Aliases

CTRL

CTRL (rw) register accessor: Crystal Oscillator Control

DORMANT

DORMANT (rw) register accessor: Crystal Oscillator pause control
This is used to save power by pausing the XOSC
On power-up this field is initialised to WAKE
An invalid write will also select WAKE
WARNING: stop the PLLs before selecting dormant mode
WARNING: setup the irq before selecting dormant mode

STARTUP

STARTUP (rw) register accessor: Controls the startup delay

STATUS

STATUS (rw) register accessor: Crystal Oscillator Status