Module rp2040_pac::busctrl

Register block for busfabric control signals and performance counters

Modules

bus_priority

Set the priority of each master for bus arbitration.

bus_priority_ack

Bus priority acknowledge

perfctr0

Bus fabric performance counter 0

perfctr1

Bus fabric performance counter 1

perfctr2

Bus fabric performance counter 2

perfctr3

Bus fabric performance counter 3

perfsel0

Bus fabric performance event select for PERFCTR0

perfsel1

Bus fabric performance event select for PERFCTR1

perfsel2

Bus fabric performance event select for PERFCTR2

perfsel3

Bus fabric performance event select for PERFCTR3

Structs

RegisterBlock

Register block

Type Definitions

BUS_PRIORITY

BUS_PRIORITY register accessor: an alias for Reg<BUS_PRIORITY_SPEC>

BUS_PRIORITY_ACK

BUS_PRIORITY_ACK register accessor: an alias for Reg<BUS_PRIORITY_ACK_SPEC>

PERFCTR0

PERFCTR0 register accessor: an alias for Reg<PERFCTR0_SPEC>

PERFCTR1

PERFCTR1 register accessor: an alias for Reg<PERFCTR1_SPEC>

PERFCTR2

PERFCTR2 register accessor: an alias for Reg<PERFCTR2_SPEC>

PERFCTR3

PERFCTR3 register accessor: an alias for Reg<PERFCTR3_SPEC>

PERFSEL0

PERFSEL0 register accessor: an alias for Reg<PERFSEL0_SPEC>

PERFSEL1

PERFSEL1 register accessor: an alias for Reg<PERFSEL1_SPEC>

PERFSEL2

PERFSEL2 register accessor: an alias for Reg<PERFSEL2_SPEC>

PERFSEL3

PERFSEL3 register accessor: an alias for Reg<PERFSEL3_SPEC>