rp2040_pac

Module busctrl

Register block for busfabric control signals and performance counters

Modules

bus_priority

Set the priority of each master for bus arbitration.

bus_priority_ack

Bus priority acknowledge

perfctr0

Bus fabric performance counter 0

perfctr1

Bus fabric performance counter 1

perfctr2

Bus fabric performance counter 2

perfctr3

Bus fabric performance counter 3

perfsel0

Bus fabric performance event select for PERFCTR0

perfsel1

Bus fabric performance event select for PERFCTR1

perfsel2

Bus fabric performance event select for PERFCTR2

perfsel3

Bus fabric performance event select for PERFCTR3

Structs

RegisterBlock

Register block

Type Aliases

BUS_PRIORITY

BUS_PRIORITY (rw) register accessor: Set the priority of each master for bus arbitration.

BUS_PRIORITY_ACK

BUS_PRIORITY_ACK (r) register accessor: Bus priority acknowledge

PERFCTR0

PERFCTR0 (rw) register accessor: Bus fabric performance counter 0

PERFCTR1

PERFCTR1 (rw) register accessor: Bus fabric performance counter 1

PERFCTR2

PERFCTR2 (rw) register accessor: Bus fabric performance counter 2

PERFCTR3

PERFCTR3 (rw) register accessor: Bus fabric performance counter 3

PERFSEL0

PERFSEL0 (rw) register accessor: Bus fabric performance event select for PERFCTR0

PERFSEL1

PERFSEL1 (rw) register accessor: Bus fabric performance event select for PERFCTR1

PERFSEL2

PERFSEL2 (rw) register accessor: Bus fabric performance event select for PERFCTR2

PERFSEL3

PERFSEL3 (rw) register accessor: Bus fabric performance event select for PERFCTR3