Module rp2040_pac::clocks::clk_sys_resus_ctrl

Structs

CLEAR_R

Field CLEAR reader - For clearing the resus after the fault that triggered it has been corrected

CLEAR_W

Field CLEAR writer - For clearing the resus after the fault that triggered it has been corrected

CLK_SYS_RESUS_CTRL_SPEC

This register you can read, write_with_zero, reset, write, modify. See API.

ENABLE_R

Field ENABLE reader - Enable resus

ENABLE_W

Field ENABLE writer - Enable resus

FRCE_R

Field FRCE reader - Force a resus, for test purposes only

FRCE_W

Field FRCE writer - Force a resus, for test purposes only

R

Register CLK_SYS_RESUS_CTRL reader

TIMEOUT_R

Field TIMEOUT reader - This is expressed as a number of clk_ref cycles
and must be >= 2x clk_ref_freq/min_clk_tst_freq

TIMEOUT_W

Field TIMEOUT writer - This is expressed as a number of clk_ref cycles
and must be >= 2x clk_ref_freq/min_clk_tst_freq

W

Register CLK_SYS_RESUS_CTRL writer