Module rp2040_pac::dma::sniff_ctrl

Sniffer Control

Structs

BSWAP_R

Field BSWAP reader - Locally perform a byte reverse on the sniffed data, before feeding into checksum.

BSWAP_W

Field BSWAP writer - Locally perform a byte reverse on the sniffed data, before feeding into checksum.

CALC_R

Field CALC reader -

CALC_W

Field CALC writer -

DMACH_R

Field DMACH reader - DMA channel for Sniffer to observe

DMACH_W

Field DMACH writer - DMA channel for Sniffer to observe

EN_R

Field EN reader - Enable sniffer

EN_W

Field EN writer - Enable sniffer

OUT_INV_R

Field OUT_INV reader - If set, the result appears inverted (bitwise complement) when read. This does not affect the way the checksum is calculated; the result is transformed on-the-fly between the result register and the bus.

OUT_INV_W

Field OUT_INV writer - If set, the result appears inverted (bitwise complement) when read. This does not affect the way the checksum is calculated; the result is transformed on-the-fly between the result register and the bus.

OUT_REV_R

Field OUT_REV reader - If set, the result appears bit-reversed when read. This does not affect the way the checksum is calculated; the result is transformed on-the-fly between the result register and the bus.

OUT_REV_W

Field OUT_REV writer - If set, the result appears bit-reversed when read. This does not affect the way the checksum is calculated; the result is transformed on-the-fly between the result register and the bus.

R

Register SNIFF_CTRL reader

SNIFF_CTRL_SPEC

Sniffer Control

W

Register SNIFF_CTRL writer

Enums

CALC_A

Value on reset: 0