Module rp2040_pac::i2c0

DW_apb_i2c address block

Modules

ic_ack_general_call
I2C ACK General Call Register

ic_clr_activity
Clear ACTIVITY Interrupt Register

ic_clr_gen_call
Clear GEN_CALL Interrupt Register

ic_clr_intr
Clear Combined and Individual Interrupt Register

ic_clr_rd_req
Clear RD_REQ Interrupt Register

ic_clr_restart_det
Clear RESTART_DET Interrupt Register

ic_clr_rx_done
Clear RX_DONE Interrupt Register

ic_clr_rx_over
Clear RX_OVER Interrupt Register

ic_clr_rx_under
Clear RX_UNDER Interrupt Register

ic_clr_start_det
Clear START_DET Interrupt Register

ic_clr_stop_det
Clear STOP_DET Interrupt Register

ic_clr_tx_abrt
Clear TX_ABRT Interrupt Register

ic_clr_tx_over
Clear TX_OVER Interrupt Register

ic_comp_param_1
Component Parameter Register 1

ic_comp_type
I2C Component Type Register

ic_comp_version
I2C Component Version Register

ic_con
I2C Control Register. This register can be written only when the DW_apb_i2c is disabled, which corresponds to the IC_ENABLE[0] register being set to 0. Writes at other times have no effect.

ic_data_cmd
I2C Rx/Tx Data Buffer and Command Register; this is the register the CPU writes to when filling the TX FIFO and the CPU reads from when retrieving bytes from RX FIFO.

ic_dma_cr
DMA Control Register

ic_dma_rdlr
I2C Receive Data Level Register

ic_dma_tdlr
DMA Transmit Data Level Register

ic_enable
I2C Enable Register

ic_enable_status
I2C Enable Status Register

ic_fs_scl_hcnt
Fast Mode or Fast Mode Plus I2C Clock SCL High Count Register

ic_fs_scl_lcnt
Fast Mode or Fast Mode Plus I2C Clock SCL Low Count Register

ic_fs_spklen
I2C SS, FS or FM+ spike suppression limit

ic_intr_mask
I2C Interrupt Mask Register.

ic_intr_stat
I2C Interrupt Status Register

ic_raw_intr_stat
I2C Raw Interrupt Status Register

ic_rx_tl
I2C Receive FIFO Threshold Register

ic_rxflr
I2C Receive FIFO Level Register This register contains the number of valid data entries in the receive FIFO buffer. It is cleared whenever: - The I2C is disabled - Whenever there is a transmit abort caused by any of the events tracked in IC_TX_ABRT_SOURCE The register increments whenever data is placed into the receive FIFO and decrements when data is taken from the receive FIFO.

ic_sar
I2C Slave Address Register

ic_sda_hold
I2C SDA Hold Time Length Register

ic_sda_setup
I2C SDA Setup Register

ic_slv_data_nack_only
Generate Slave Data NACK Register

ic_ss_scl_hcnt
Standard Speed I2C Clock SCL High Count Register

ic_ss_scl_lcnt
Standard Speed I2C Clock SCL Low Count Register

ic_status
I2C Status Register

ic_tar
I2C Target Address Register

ic_tx_abrt_source
I2C Transmit Abort Source Register

ic_tx_tl
I2C Transmit FIFO Threshold Register

ic_txflr
I2C Transmit FIFO Level Register This register contains the number of valid data entries in the transmit FIFO buffer. It is cleared whenever: - The I2C is disabled - There is a transmit abort - that is, TX_ABRT bit is set in the IC_RAW_INTR_STAT register - The slave bulk transmit mode is aborted The register increments whenever data is placed into the transmit FIFO and decrements when data is taken from the transmit FIFO.

Structs

RegisterBlock
Register block

Type Definitions

IC_ACK_GENERAL_CALL
I2C ACK General Call Register

IC_CLR_ACTIVITY
Clear ACTIVITY Interrupt Register

IC_CLR_GEN_CALL
Clear GEN_CALL Interrupt Register

IC_CLR_INTR
Clear Combined and Individual Interrupt Register

IC_CLR_RD_REQ
Clear RD_REQ Interrupt Register

IC_CLR_RESTART_DET
Clear RESTART_DET Interrupt Register

IC_CLR_RX_DONE
Clear RX_DONE Interrupt Register

IC_CLR_RX_OVER
Clear RX_OVER Interrupt Register

IC_CLR_RX_UNDER
Clear RX_UNDER Interrupt Register

IC_CLR_START_DET
Clear START_DET Interrupt Register

IC_CLR_STOP_DET
Clear STOP_DET Interrupt Register

IC_CLR_TX_ABRT
Clear TX_ABRT Interrupt Register

IC_CLR_TX_OVER
Clear TX_OVER Interrupt Register

IC_COMP_PARAM_1
Component Parameter Register 1

IC_COMP_TYPE
I2C Component Type Register

IC_COMP_VERSION
I2C Component Version Register

IC_CON
I2C Control Register. This register can be written only when the DW_apb_i2c is disabled, which corresponds to the IC_ENABLE[0] register being set to 0. Writes at other times have no effect.

IC_DATA_CMD
I2C Rx/Tx Data Buffer and Command Register; this is the register the CPU writes to when filling the TX FIFO and the CPU reads from when retrieving bytes from RX FIFO.

IC_DMA_CR
DMA Control Register

IC_DMA_RDLR
I2C Receive Data Level Register

IC_DMA_TDLR
DMA Transmit Data Level Register

IC_ENABLE
I2C Enable Register

IC_ENABLE_STATUS
I2C Enable Status Register

IC_FS_SCL_HCNT
Fast Mode or Fast Mode Plus I2C Clock SCL High Count Register

IC_FS_SCL_LCNT
Fast Mode or Fast Mode Plus I2C Clock SCL Low Count Register

IC_FS_SPKLEN
I2C SS, FS or FM+ spike suppression limit

IC_INTR_MASK
I2C Interrupt Mask Register.

IC_INTR_STAT
I2C Interrupt Status Register

IC_RAW_INTR_STAT
I2C Raw Interrupt Status Register

IC_RXFLR
I2C Receive FIFO Level Register This register contains the number of valid data entries in the receive FIFO buffer. It is cleared whenever: - The I2C is disabled - Whenever there is a transmit abort caused by any of the events tracked in IC_TX_ABRT_SOURCE The register increments whenever data is placed into the receive FIFO and decrements when data is taken from the receive FIFO.

IC_RX_TL
I2C Receive FIFO Threshold Register

IC_SAR
I2C Slave Address Register

IC_SDA_HOLD
I2C SDA Hold Time Length Register

IC_SDA_SETUP
I2C SDA Setup Register

IC_SLV_DATA_NACK_ONLY
Generate Slave Data NACK Register

IC_SS_SCL_HCNT
Standard Speed I2C Clock SCL High Count Register

IC_SS_SCL_LCNT
Standard Speed I2C Clock SCL Low Count Register

IC_STATUS
I2C Status Register

IC_TAR
I2C Target Address Register

IC_TXFLR
I2C Transmit FIFO Level Register This register contains the number of valid data entries in the transmit FIFO buffer. It is cleared whenever: - The I2C is disabled - There is a transmit abort - that is, TX_ABRT bit is set in the IC_RAW_INTR_STAT register - The slave bulk transmit mode is aborted The register increments whenever data is placed into the transmit FIFO and decrements when data is taken from the transmit FIFO.

IC_TX_ABRT_SOURCE
I2C Transmit Abort Source Register

IC_TX_TL
I2C Transmit FIFO Threshold Register