rp2040_pac::i2c0
Module ic_clr_activity

Clear ACTIVITY Interrupt Register

Structs

IC_CLR_ACTIVITY_SPEC

Clear ACTIVITY Interrupt Register

Type Aliases

CLR_ACTIVITY_R

Field CLR_ACTIVITY reader - Reading this register clears the ACTIVITY interrupt if the I2C is not active anymore. If the I2C module is still active on the bus, the ACTIVITY interrupt bit continues to be set. It is automatically cleared by hardware if the module is disabled and if there is no further activity on the bus. The value read from this register to get status of the ACTIVITY interrupt (bit 8) of the IC_RAW_INTR_STAT register.

R

Register IC_CLR_ACTIVITY reader