rp2040_pac::i2c0

Module ic_comp_param_1

Component Parameter Register 1

Note This register is not implemented and therefore reads as 0. If it was implemented it would be a constant read-only register that contains encoded information about the component’s parameter settings. Fields shown below are the settings for those parameters

Structs

IC_COMP_PARAM_1_SPEC

Component Parameter Register 1

Type Aliases

ADD_ENCODED_PARAMS_R

Field ADD_ENCODED_PARAMS reader - Encoded parameters not visible

APB_DATA_WIDTH_R

Field APB_DATA_WIDTH reader - APB data bus width is 32 bits

HAS_DMA_R

Field HAS_DMA reader - DMA handshaking signals are enabled

HC_COUNT_VALUES_R

Field HC_COUNT_VALUES reader - Programmable count values for each mode.

INTR_IO_R

Field INTR_IO reader - COMBINED Interrupt outputs

MAX_SPEED_MODE_R

Field MAX_SPEED_MODE reader - MAX SPEED MODE = FAST MODE

R

Register IC_COMP_PARAM_1 reader

RX_BUFFER_DEPTH_R

Field RX_BUFFER_DEPTH reader - RX Buffer Depth = 16

TX_BUFFER_DEPTH_R

Field TX_BUFFER_DEPTH reader - TX Buffer Depth = 16