Module rp2040_pac::i2c0::ic_data_cmd

I2C Rx/Tx Data Buffer and Command Register; this is the register the CPU writes to when filling the TX FIFO and the CPU reads from when retrieving bytes from RX FIFO.

The size of the register changes as follows:

Write: - 11 bits when IC_EMPTYFIFO_HOLD_MASTER_EN=1 - 9 bits when IC_EMPTYFIFO_HOLD_MASTER_EN=0 Read: - 12 bits when IC_FIRST_DATA_BYTE_STATUS = 1 - 8 bits when IC_FIRST_DATA_BYTE_STATUS = 0 Note: In order for the DW_apb_i2c to continue acknowledging reads, a read command should be written for every byte that is to be received; otherwise the DW_apb_i2c will stop acknowledging.

Structs

CMD_W
Write proxy for field CMD

DAT_W
Write proxy for field DAT

RESTART_W
Write proxy for field RESTART

STOP_W
Write proxy for field STOP

Enums

CMD_A
This bit controls whether a read or a write is performed. This bit does not control the direction when the DW_apb_i2con acts as a slave. It controls only the direction when it acts as a master.

FIRST_DATA_BYTE_A
Indicates the first data byte received after the address phase for receive transfer in Master receiver or Slave receiver mode.

RESTART_A
This bit controls whether a RESTART is issued before the byte is sent or received.

STOP_A
This bit controls whether a STOP is issued after the byte is sent or received.

Type Definitions

CMD_R
Reader of field CMD

DAT_R
Reader of field DAT

FIRST_DATA_BYTE_R
Reader of field FIRST_DATA_BYTE

R
Reader of register IC_DATA_CMD

RESTART_R
Reader of field RESTART

STOP_R
Reader of field STOP

W
Writer for register IC_DATA_CMD