rp2040_pac::i2c0::ic_enable

Type Alias TX_CMD_BLOCK_W

pub type TX_CMD_BLOCK_W<'a, REG> = BitWriter<'a, REG, TX_CMD_BLOCK_A>;

Field TX_CMD_BLOCK writer - In Master mode: - 1’b1: Blocks the transmission of data on I2C bus even if Tx FIFO has data to transmit. - 1’b0: The transmission of data starts on I2C bus automatically, as soon as the first data is available in the Tx FIFO. Note: To block the execution of Master commands, set the TX_CMD_BLOCK bit only when Tx FIFO is empty (IC_STATUS[2]==1) and Master is in Idle state (IC_STATUS[5] == 0). Any further commands put in the Tx FIFO are not executed until TX_CMD_BLOCK bit is unset. Reset value: IC_TX_CMD_BLOCK_DEFAULT

Aliased Type

struct TX_CMD_BLOCK_W<'a, REG> { /* private fields */ }

Implementations

impl<'a, REG> TX_CMD_BLOCK_W<'a, REG>
where REG: Writable + RegisterSpec,

pub fn not_blocked(self) -> &'a mut W<REG>

Tx Command execution not blocked

pub fn blocked(self) -> &'a mut W<REG>

Tx Command execution blocked