Module rp2040_pac::i2c0::ic_intr_stat

I2C Interrupt Status Register

Each bit in this register has a corresponding mask bit in the IC_INTR_MASK register. These bits are cleared by reading the matching interrupt clear register. The unmasked raw versions of these bits are available in the IC_RAW_INTR_STAT register.

Structs

IC_INTR_STAT_SPEC

I2C Interrupt Status Register

Enums

R_ACTIVITY_A

See IC_RAW_INTR_STAT for a detailed description of R_ACTIVITY bit.

R_GEN_CALL_A

See IC_RAW_INTR_STAT for a detailed description of R_GEN_CALL bit.

R_RD_REQ_A

See IC_RAW_INTR_STAT for a detailed description of R_RD_REQ bit.

R_RESTART_DET_A

See IC_RAW_INTR_STAT for a detailed description of R_RESTART_DET bit.

R_RX_DONE_A

See IC_RAW_INTR_STAT for a detailed description of R_RX_DONE bit.

R_RX_FULL_A

See IC_RAW_INTR_STAT for a detailed description of R_RX_FULL bit.

R_RX_OVER_A

See IC_RAW_INTR_STAT for a detailed description of R_RX_OVER bit.

R_RX_UNDER_A

See IC_RAW_INTR_STAT for a detailed description of R_RX_UNDER bit.

R_START_DET_A

See IC_RAW_INTR_STAT for a detailed description of R_START_DET bit.

R_STOP_DET_A

See IC_RAW_INTR_STAT for a detailed description of R_STOP_DET bit.

R_TX_ABRT_A

See IC_RAW_INTR_STAT for a detailed description of R_TX_ABRT bit.

R_TX_EMPTY_A

See IC_RAW_INTR_STAT for a detailed description of R_TX_EMPTY bit.

R_TX_OVER_A

See IC_RAW_INTR_STAT for a detailed description of R_TX_OVER bit.

Type Aliases

R

Register IC_INTR_STAT reader

R_ACTIVITY_R

Field R_ACTIVITY reader - See IC_RAW_INTR_STAT for a detailed description of R_ACTIVITY bit.

R_GEN_CALL_R

Field R_GEN_CALL reader - See IC_RAW_INTR_STAT for a detailed description of R_GEN_CALL bit.

R_RD_REQ_R

Field R_RD_REQ reader - See IC_RAW_INTR_STAT for a detailed description of R_RD_REQ bit.

R_RESTART_DET_R

Field R_RESTART_DET reader - See IC_RAW_INTR_STAT for a detailed description of R_RESTART_DET bit.

R_RX_DONE_R

Field R_RX_DONE reader - See IC_RAW_INTR_STAT for a detailed description of R_RX_DONE bit.

R_RX_FULL_R

Field R_RX_FULL reader - See IC_RAW_INTR_STAT for a detailed description of R_RX_FULL bit.

R_RX_OVER_R

Field R_RX_OVER reader - See IC_RAW_INTR_STAT for a detailed description of R_RX_OVER bit.

R_RX_UNDER_R

Field R_RX_UNDER reader - See IC_RAW_INTR_STAT for a detailed description of R_RX_UNDER bit.

R_START_DET_R

Field R_START_DET reader - See IC_RAW_INTR_STAT for a detailed description of R_START_DET bit.

R_STOP_DET_R

Field R_STOP_DET reader - See IC_RAW_INTR_STAT for a detailed description of R_STOP_DET bit.

R_TX_ABRT_R

Field R_TX_ABRT reader - See IC_RAW_INTR_STAT for a detailed description of R_TX_ABRT bit.

R_TX_EMPTY_R

Field R_TX_EMPTY reader - See IC_RAW_INTR_STAT for a detailed description of R_TX_EMPTY bit.

R_TX_OVER_R

Field R_TX_OVER reader - See IC_RAW_INTR_STAT for a detailed description of R_TX_OVER bit.