rp2040_pac::i2c0

Module ic_sda_hold

I2C SDA Hold Time Length Register

The bits [15:0] of this register are used to control the hold time of SDA during transmit in both slave and master mode (after SCL goes from HIGH to LOW).

The bits [23:16] of this register are used to extend the SDA transition (if any) whenever SCL is HIGH in the receiver in either master or slave mode.

Writes to this register succeed only when IC_ENABLE[0]=0.

The values in this register are in units of ic_clk period. The value programmed in IC_SDA_TX_HOLD must be greater than the minimum hold time in each mode (one cycle in master mode, seven cycles in slave mode) for the value to be implemented.

The programmed SDA hold time during transmit (IC_SDA_TX_HOLD) cannot exceed at any time the duration of the low part of scl. Therefore the programmed value cannot be larger than N_SCL_LOW-2, where N_SCL_LOW is the duration of the low part of the scl period measured in ic_clk cycles.

Structs

IC_SDA_HOLD_SPEC

I2C SDA Hold Time Length Register

Type Aliases

IC_SDA_RX_HOLD_R

Field IC_SDA_RX_HOLD reader - Sets the required SDA hold time in units of ic_clk period, when DW_apb_i2c acts as a receiver.

IC_SDA_RX_HOLD_W

Field IC_SDA_RX_HOLD writer - Sets the required SDA hold time in units of ic_clk period, when DW_apb_i2c acts as a receiver.

IC_SDA_TX_HOLD_R

Field IC_SDA_TX_HOLD reader - Sets the required SDA hold time in units of ic_clk period, when DW_apb_i2c acts as a transmitter.

IC_SDA_TX_HOLD_W

Field IC_SDA_TX_HOLD writer - Sets the required SDA hold time in units of ic_clk period, when DW_apb_i2c acts as a transmitter.

R

Register IC_SDA_HOLD reader

W

Register IC_SDA_HOLD writer