rp2040_pac::pio0

Module ctrl

PIO control register

Structs

CTRL_SPEC

PIO control register

Type Aliases

CLKDIV_RESTART_R

Field CLKDIV_RESTART reader - Restart a state machine’s clock divider from an initial phase of 0. Clock dividers are free-running, so once started, their output (including fractional jitter) is completely determined by the integer/fractional divisor configured in SMx_CLKDIV. This means that, if multiple clock dividers with the same divisor are restarted simultaneously, by writing multiple 1 bits to this field, the execution clocks of those state machines will run in precise lockstep.

CLKDIV_RESTART_W

Field CLKDIV_RESTART writer - Restart a state machine’s clock divider from an initial phase of 0. Clock dividers are free-running, so once started, their output (including fractional jitter) is completely determined by the integer/fractional divisor configured in SMx_CLKDIV. This means that, if multiple clock dividers with the same divisor are restarted simultaneously, by writing multiple 1 bits to this field, the execution clocks of those state machines will run in precise lockstep.

R

Register CTRL reader

SM_ENABLE_R

Field SM_ENABLE reader - Enable/disable each of the four state machines by writing 1/0 to each of these four bits. When disabled, a state machine will cease executing instructions, except those written directly to SMx_INSTR by the system. Multiple bits can be set/cleared at once to run/halt multiple state machines simultaneously.

SM_ENABLE_W

Field SM_ENABLE writer - Enable/disable each of the four state machines by writing 1/0 to each of these four bits. When disabled, a state machine will cease executing instructions, except those written directly to SMx_INSTR by the system. Multiple bits can be set/cleared at once to run/halt multiple state machines simultaneously.

SM_RESTART_R

Field SM_RESTART reader - Write 1 to instantly clear internal SM state which may be otherwise difficult to access and will affect future execution.

SM_RESTART_W

Field SM_RESTART writer - Write 1 to instantly clear internal SM state which may be otherwise difficult to access and will affect future execution.

W

Register CTRL writer