rp2040_pac::pio0

Module sm

Cluster Cluster SM%s, containing SM*_CLKDIV, SM*_EXECCTRL, SM*_SHIFTCTRL, SM*_ADDR, SM*_INSTR, SM*_PINCTRL

Modules

sm_addr

Current instruction address of state machine 0

sm_clkdiv

Clock divisor register for state machine 0
Frequency = clock freq / (CLKDIV_INT + CLKDIV_FRAC / 256)

sm_execctrl

Execution/behavioural settings for state machine 0

sm_instr

Read to see the instruction currently addressed by state machine 0’s program counter
Write to execute an instruction immediately (including jumps) and then resume execution.

sm_pinctrl

State machine pin control

sm_shiftctrl

Control behaviour of the input/output shift registers for state machine 0

Structs

SM

Register block

Type Aliases

SM_ADDR

SM_ADDR (r) register accessor: Current instruction address of state machine 0

SM_CLKDIV

SM_CLKDIV (rw) register accessor: Clock divisor register for state machine 0
Frequency = clock freq / (CLKDIV_INT + CLKDIV_FRAC / 256)

SM_EXECCTRL

SM_EXECCTRL (rw) register accessor: Execution/behavioural settings for state machine 0

SM_INSTR

SM_INSTR (rw) register accessor: Read to see the instruction currently addressed by state machine 0’s program counter
Write to execute an instruction immediately (including jumps) and then resume execution.

SM_PINCTRL

SM_PINCTRL (rw) register accessor: State machine pin control

SM_SHIFTCTRL

SM_SHIFTCTRL (rw) register accessor: Control behaviour of the input/output shift registers for state machine 0