Type Alias rp2040_pac::pio0::sm::sm_pinctrl::R

pub type R = R<SM_PINCTRL_SPEC>;

Register SM_PINCTRL reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn out_base(&self) -> OUT_BASE_R

Bits 0:4 - The lowest-numbered pin that will be affected by an OUT PINS, OUT PINDIRS or MOV PINS instruction. The data written to this pin will always be the least-significant bit of the OUT or MOV data.

pub fn set_base(&self) -> SET_BASE_R

Bits 5:9 - The lowest-numbered pin that will be affected by a SET PINS or SET PINDIRS instruction. The data written to this pin is the least-significant bit of the SET data.

pub fn sideset_base(&self) -> SIDESET_BASE_R

Bits 10:14 - The lowest-numbered pin that will be affected by a side-set operation. The MSBs of an instruction’s side-set/delay field (up to 5, determined by SIDESET_COUNT) are used for side-set data, with the remaining LSBs used for delay. The least-significant bit of the side-set portion is the bit written to this pin, with more-significant bits written to higher-numbered pins.

pub fn in_base(&self) -> IN_BASE_R

Bits 15:19 - The pin which is mapped to the least-significant bit of a state machine’s IN data bus. Higher-numbered pins are mapped to consecutively more-significant data bits, with a modulo of 32 applied to pin number.

pub fn out_count(&self) -> OUT_COUNT_R

Bits 20:25 - The number of pins asserted by an OUT PINS, OUT PINDIRS or MOV PINS instruction. In the range 0 to 32 inclusive.

pub fn set_count(&self) -> SET_COUNT_R

Bits 26:28 - The number of pins asserted by a SET. In the range 0 to 5 inclusive.

pub fn sideset_count(&self) -> SIDESET_COUNT_R

Bits 29:31 - The number of MSBs of the Delay/Side-set instruction field which are used for side-set. Inclusive of the enable bit, if present. Minimum of 0 (all delay bits, no side-set) and maximum of 5 (all side-set, no delay).