rp2040_pac::ppb

Module mpu_rnr

Use the MPU Region Number Register to select the region currently accessed by MPU_RBAR and MPU_RASR.

Structs

MPU_RNR_SPEC

Use the MPU Region Number Register to select the region currently accessed by MPU_RBAR and MPU_RASR.

Type Aliases

R

Register MPU_RNR reader

REGION_R

Field REGION reader - Indicates the MPU region referenced by the MPU_RBAR and MPU_RASR registers.
The MPU supports 8 memory regions, so the permitted values of this field are 0-7.

REGION_W

Field REGION writer - Indicates the MPU region referenced by the MPU_RBAR and MPU_RASR registers.
The MPU supports 8 memory regions, so the permitted values of this field are 0-7.

W

Register MPU_RNR writer