rp2040_pac::ppb

Module nvic_ipr0

Use the Interrupt Priority Registers to assign a priority from 0 to 3 to each of the available interrupts. 0 is the highest priority, and 3 is the lowest.
Note: Writing 1 to an NVIC_ICPR bit does not affect the active state of the corresponding interrupt.
These registers are only word-accessible

Structs

NVIC_IPR0_SPEC

Use the Interrupt Priority Registers to assign a priority from 0 to 3 to each of the available interrupts. 0 is the highest priority, and 3 is the lowest.
Note: Writing 1 to an NVIC_ICPR bit does not affect the active state of the corresponding interrupt.
These registers are only word-accessible

Type Aliases

IP_0_R

Field IP_0 reader - Priority of interrupt 0

IP_0_W

Field IP_0 writer - Priority of interrupt 0

IP_1_R

Field IP_1 reader - Priority of interrupt 1

IP_1_W

Field IP_1 writer - Priority of interrupt 1

IP_2_R

Field IP_2 reader - Priority of interrupt 2

IP_2_W

Field IP_2 writer - Priority of interrupt 2

IP_3_R

Field IP_3 reader - Priority of interrupt 3

IP_3_W

Field IP_3 writer - Priority of interrupt 3

R

Register NVIC_IPR0 reader

W

Register NVIC_IPR0 writer