Module rp2040_pac::ppb::nvic_ipr4

Use the Interrupt Priority Registers to assign a priority from 0 to 3 to each of the available interrupts. 0 is the highest priority, and 3 is the lowest.

Structs

IP_16_R

Field IP_16 reader - Priority of interrupt 16

IP_16_W

Field IP_16 writer - Priority of interrupt 16

IP_17_R

Field IP_17 reader - Priority of interrupt 17

IP_17_W

Field IP_17 writer - Priority of interrupt 17

IP_18_R

Field IP_18 reader - Priority of interrupt 18

IP_18_W

Field IP_18 writer - Priority of interrupt 18

IP_19_R

Field IP_19 reader - Priority of interrupt 19

IP_19_W

Field IP_19 writer - Priority of interrupt 19

NVIC_IPR4_SPEC

Use the Interrupt Priority Registers to assign a priority from 0 to 3 to each of the available interrupts. 0 is the highest priority, and 3 is the lowest.

R

Register NVIC_IPR4 reader

W

Register NVIC_IPR4 writer