rp2040_pac::ppb

Module nvic_ipr7

Use the Interrupt Priority Registers to assign a priority from 0 to 3 to each of the available interrupts. 0 is the highest priority, and 3 is the lowest.

Structs

NVIC_IPR7_SPEC

Use the Interrupt Priority Registers to assign a priority from 0 to 3 to each of the available interrupts. 0 is the highest priority, and 3 is the lowest.

Type Aliases

IP_28_R

Field IP_28 reader - Priority of interrupt 28

IP_28_W

Field IP_28 writer - Priority of interrupt 28

IP_29_R

Field IP_29 reader - Priority of interrupt 29

IP_29_W

Field IP_29 writer - Priority of interrupt 29

IP_30_R

Field IP_30 reader - Priority of interrupt 30

IP_30_W

Field IP_30 writer - Priority of interrupt 30

IP_31_R

Field IP_31 reader - Priority of interrupt 31

IP_31_W

Field IP_31 writer - Priority of interrupt 31

R

Register NVIC_IPR7 reader

W

Register NVIC_IPR7 writer