rp2040_pac::ppb

Module syst_csr

Use the SysTick Control and Status Register to enable the SysTick features.

Structs

SYST_CSR_SPEC

Use the SysTick Control and Status Register to enable the SysTick features.

Type Aliases

CLKSOURCE_R

Field CLKSOURCE reader - SysTick clock source. Always reads as one if SYST_CALIB reports NOREF.
Selects the SysTick timer clock source:
0 = External reference clock.
1 = Processor clock.

CLKSOURCE_W

Field CLKSOURCE writer - SysTick clock source. Always reads as one if SYST_CALIB reports NOREF.
Selects the SysTick timer clock source:
0 = External reference clock.
1 = Processor clock.

COUNTFLAG_R

Field COUNTFLAG reader - Returns 1 if timer counted to 0 since last time this was read. Clears on read by application or debugger.

ENABLE_R

Field ENABLE reader - Enable SysTick counter:
0 = Counter disabled.
1 = Counter enabled.

ENABLE_W

Field ENABLE writer - Enable SysTick counter:
0 = Counter disabled.
1 = Counter enabled.

R

Register SYST_CSR reader

TICKINT_R

Field TICKINT reader - Enables SysTick exception request:
0 = Counting down to zero does not assert the SysTick exception request.
1 = Counting down to zero to asserts the SysTick exception request.

TICKINT_W

Field TICKINT writer - Enables SysTick exception request:
0 = Counting down to zero does not assert the SysTick exception request.
1 = Counting down to zero to asserts the SysTick exception request.

W

Register SYST_CSR writer