rp2040_pac::psm

Module wdsel

Set to 1 if this peripheral should be reset when the watchdog fires.

Structs

WDSEL_SPEC

Set to 1 if this peripheral should be reset when the watchdog fires.

Type Aliases

BUSFABRIC_R

Field busfabric reader -

BUSFABRIC_W

Field busfabric writer -

CLOCKS_R

Field clocks reader -

CLOCKS_W

Field clocks writer -

PROC0_R

Field proc0 reader -

PROC0_W

Field proc0 writer -

PROC1_R

Field proc1 reader -

PROC1_W

Field proc1 writer -

R

Register WDSEL reader

RESETS_R

Field resets reader -

RESETS_W

Field resets writer -

ROM_R

Field rom reader -

ROM_W

Field rom writer -

ROSC_R

Field rosc reader -

ROSC_W

Field rosc writer -

SIO_R

Field sio reader -

SIO_W

Field sio writer -

SRAM0_R

Field sram0 reader -

SRAM0_W

Field sram0 writer -

SRAM1_R

Field sram1 reader -

SRAM1_W

Field sram1 writer -

SRAM2_R

Field sram2 reader -

SRAM2_W

Field sram2 writer -

SRAM3_R

Field sram3 reader -

SRAM3_W

Field sram3 writer -

SRAM4_R

Field sram4 reader -

SRAM4_W

Field sram4 writer -

SRAM5_R

Field sram5 reader -

SRAM5_W

Field sram5 writer -

VREG_AND_CHIP_RESET_R

Field vreg_and_chip_reset reader -

VREG_AND_CHIP_RESET_W

Field vreg_and_chip_reset writer -

W

Register WDSEL writer

XIP_R

Field xip reader -

XIP_W

Field xip writer -

XOSC_R

Field xosc reader -

XOSC_W

Field xosc writer -