rp2040_pac

Module pwm

Simple PWM

Re-exports

pub use self::ch::CH;

Modules

ch

Cluster Cluster CH%s, containing CH*_CC, CH*_CSR, CH*_CTR, CH*_DIV, CH*_TOP

en

This register aliases the CSR_EN bits for all channels.
Writing to this register allows multiple channels to be enabled
or disabled simultaneously, so they can run in perfect sync.
For each channel, there is only one physical EN register bit,
which can be accessed through here or CHx_CSR.

inte

Interrupt Enable

intf

Interrupt Force

intr

Raw Interrupts

ints

Interrupt status after masking & forcing

Structs

RegisterBlock

Register block

Type Aliases

EN

EN (rw) register accessor: This register aliases the CSR_EN bits for all channels.
Writing to this register allows multiple channels to be enabled
or disabled simultaneously, so they can run in perfect sync.
For each channel, there is only one physical EN register bit,
which can be accessed through here or CHx_CSR.

INTE

INTE (rw) register accessor: Interrupt Enable

INTF

INTF (rw) register accessor: Interrupt Force

INTR

INTR (rw) register accessor: Raw Interrupts

INTS

INTS (r) register accessor: Interrupt status after masking & forcing