rp2040_pac

Module resets

RESETS

Modules

reset

Reset control. If a bit is set it means the peripheral is in reset. 0 means the peripheral’s reset is deasserted.

reset_done

Reset done. If a bit is set then a reset done signal has been returned by the peripheral. This indicates that the peripheral’s registers are ready to be accessed.

wdsel

Watchdog select. If a bit is set then the watchdog will reset this peripheral when the watchdog fires.

Structs

RegisterBlock

Register block

Type Aliases

RESET

RESET (rw) register accessor: Reset control. If a bit is set it means the peripheral is in reset. 0 means the peripheral’s reset is deasserted.

RESET_DONE

RESET_DONE (r) register accessor: Reset done. If a bit is set then a reset done signal has been returned by the peripheral. This indicates that the peripheral’s registers are ready to be accessed.

WDSEL

WDSEL (rw) register accessor: Watchdog select. If a bit is set then the watchdog will reset this peripheral when the watchdog fires.