rp2040_pac::resets

Module reset

Reset control. If a bit is set it means the peripheral is in reset. 0 means the peripheral’s reset is deasserted.

Structs

RESET_SPEC

Reset control. If a bit is set it means the peripheral is in reset. 0 means the peripheral’s reset is deasserted.

Type Aliases

ADC_R

Field adc reader -

ADC_W

Field adc writer -

BUSCTRL_R

Field busctrl reader -

BUSCTRL_W

Field busctrl writer -

DMA_R

Field dma reader -

DMA_W

Field dma writer -

I2C0_R

Field i2c0 reader -

I2C0_W

Field i2c0 writer -

I2C1_R

Field i2c1 reader -

I2C1_W

Field i2c1 writer -

IO_BANK0_R

Field io_bank0 reader -

IO_BANK0_W

Field io_bank0 writer -

IO_QSPI_R

Field io_qspi reader -

IO_QSPI_W

Field io_qspi writer -

JTAG_R

Field jtag reader -

JTAG_W

Field jtag writer -

PADS_BANK0_R

Field pads_bank0 reader -

PADS_BANK0_W

Field pads_bank0 writer -

PADS_QSPI_R

Field pads_qspi reader -

PADS_QSPI_W

Field pads_qspi writer -

PIO0_R

Field pio0 reader -

PIO0_W

Field pio0 writer -

PIO1_R

Field pio1 reader -

PIO1_W

Field pio1 writer -

PLL_SYS_R

Field pll_sys reader -

PLL_SYS_W

Field pll_sys writer -

PLL_USB_R

Field pll_usb reader -

PLL_USB_W

Field pll_usb writer -

PWM_R

Field pwm reader -

PWM_W

Field pwm writer -

R

Register RESET reader

RTC_R

Field rtc reader -

RTC_W

Field rtc writer -

SPI0_R

Field spi0 reader -

SPI0_W

Field spi0 writer -

SPI1_R

Field spi1 reader -

SPI1_W

Field spi1 writer -

SYSCFG_R

Field syscfg reader -

SYSCFG_W

Field syscfg writer -

SYSINFO_R

Field sysinfo reader -

SYSINFO_W

Field sysinfo writer -

TBMAN_R

Field tbman reader -

TBMAN_W

Field tbman writer -

TIMER_R

Field timer reader -

TIMER_W

Field timer writer -

UART0_R

Field uart0 reader -

UART0_W

Field uart0 writer -

UART1_R

Field uart1 reader -

UART1_W

Field uart1 writer -

USBCTRL_R

Field usbctrl reader -

USBCTRL_W

Field usbctrl writer -

W

Register RESET writer