Module rp2040_pac::rosc::freqa

The FREQA & FREQB registers control the frequency by controlling the drive strength of each stage
The drive strength has 4 levels determined by the number of bits set
Increasing the number of bits set increases the drive strength and increases the oscillation frequency
0 bits set is the default drive strength
1 bit set doubles the drive strength
2 bits set triples drive strength
3 bits set quadruples drive strength

Structs

DS0_R

Field DS0 reader - Stage 0 drive strength

DS0_W

Field DS0 writer - Stage 0 drive strength

DS1_R

Field DS1 reader - Stage 1 drive strength

DS1_W

Field DS1 writer - Stage 1 drive strength

DS2_R

Field DS2 reader - Stage 2 drive strength

DS2_W

Field DS2 writer - Stage 2 drive strength

DS3_R

Field DS3 reader - Stage 3 drive strength

DS3_W

Field DS3 writer - Stage 3 drive strength

FREQA_SPEC

The FREQA & FREQB registers control the frequency by controlling the drive strength of each stage
The drive strength has 4 levels determined by the number of bits set
Increasing the number of bits set increases the drive strength and increases the oscillation frequency
0 bits set is the default drive strength
1 bit set doubles the drive strength
2 bits set triples drive strength
3 bits set quadruples drive strength

PASSWD_R

Field PASSWD reader - Set to 0x9696 to apply the settings
Any other value in this field will set all drive strengths to 0

PASSWD_W

Field PASSWD writer - Set to 0x9696 to apply the settings
Any other value in this field will set all drive strengths to 0

R

Register FREQA reader

W

Register FREQA writer

Enums

PASSWD_A

Set to 0x9696 to apply the settings
Any other value in this field will set all drive strengths to 0