Module rp2040_pac::syscfg

Register block for various chip control signals

Modules

dbgforce

Directly control the SWD debug port of either processor

mempowerdown

Control power downs to memories. Set high to power down memories.
Use with extreme caution

proc0_nmi_mask

Processor core 0 NMI source mask
Set a bit high to enable NMI from that IRQ

proc1_nmi_mask

Processor core 1 NMI source mask
Set a bit high to enable NMI from that IRQ

proc_config

Configuration for processors

proc_in_sync_bypass

For each bit, if 1, bypass the input synchronizer between that GPIO
and the GPIO input register in the SIO. The input synchronizers should
generally be unbypassed, to avoid injecting metastabilities into processors.
If you’re feeling brave, you can bypass to save two cycles of input
latency. This register applies to GPIO 0…29.

proc_in_sync_bypass_hi

For each bit, if 1, bypass the input synchronizer between that GPIO
and the GPIO input register in the SIO. The input synchronizers should
generally be unbypassed, to avoid injecting metastabilities into processors.
If you’re feeling brave, you can bypass to save two cycles of input
latency. This register applies to GPIO 30…35 (the QSPI IOs).

Structs

RegisterBlock

Register block

Type Definitions

DBGFORCE

DBGFORCE register accessor: an alias for Reg<DBGFORCE_SPEC>

MEMPOWERDOWN

MEMPOWERDOWN register accessor: an alias for Reg<MEMPOWERDOWN_SPEC>

PROC0_NMI_MASK

PROC0_NMI_MASK register accessor: an alias for Reg<PROC0_NMI_MASK_SPEC>

PROC1_NMI_MASK

PROC1_NMI_MASK register accessor: an alias for Reg<PROC1_NMI_MASK_SPEC>

PROC_CONFIG

PROC_CONFIG register accessor: an alias for Reg<PROC_CONFIG_SPEC>

PROC_IN_SYNC_BYPASS

PROC_IN_SYNC_BYPASS register accessor: an alias for Reg<PROC_IN_SYNC_BYPASS_SPEC>

PROC_IN_SYNC_BYPASS_HI

PROC_IN_SYNC_BYPASS_HI register accessor: an alias for Reg<PROC_IN_SYNC_BYPASS_HI_SPEC>