rp2040_pac::syscfg

Module dbgforce

Directly control the SWD debug port of either processor

Structs

DBGFORCE_SPEC

Directly control the SWD debug port of either processor

Type Aliases

PROC0_ATTACH_R

Field PROC0_ATTACH reader - Attach processor 0 debug port to syscfg controls, and disconnect it from external SWD pads.

PROC0_ATTACH_W

Field PROC0_ATTACH writer - Attach processor 0 debug port to syscfg controls, and disconnect it from external SWD pads.

PROC0_SWCLK_R

Field PROC0_SWCLK reader - Directly drive processor 0 SWCLK, if PROC0_ATTACH is set

PROC0_SWCLK_W

Field PROC0_SWCLK writer - Directly drive processor 0 SWCLK, if PROC0_ATTACH is set

PROC0_SWDI_R

Field PROC0_SWDI reader - Directly drive processor 0 SWDIO input, if PROC0_ATTACH is set

PROC0_SWDI_W

Field PROC0_SWDI writer - Directly drive processor 0 SWDIO input, if PROC0_ATTACH is set

PROC0_SWDO_R

Field PROC0_SWDO reader - Observe the value of processor 0 SWDIO output.

PROC1_ATTACH_R

Field PROC1_ATTACH reader - Attach processor 1 debug port to syscfg controls, and disconnect it from external SWD pads.

PROC1_ATTACH_W

Field PROC1_ATTACH writer - Attach processor 1 debug port to syscfg controls, and disconnect it from external SWD pads.

PROC1_SWCLK_R

Field PROC1_SWCLK reader - Directly drive processor 1 SWCLK, if PROC1_ATTACH is set

PROC1_SWCLK_W

Field PROC1_SWCLK writer - Directly drive processor 1 SWCLK, if PROC1_ATTACH is set

PROC1_SWDI_R

Field PROC1_SWDI reader - Directly drive processor 1 SWDIO input, if PROC1_ATTACH is set

PROC1_SWDI_W

Field PROC1_SWDI writer - Directly drive processor 1 SWDIO input, if PROC1_ATTACH is set

PROC1_SWDO_R

Field PROC1_SWDO reader - Observe the value of processor 1 SWDIO output.

R

Register DBGFORCE reader

W

Register DBGFORCE writer