rp2040_pac::syscfg

Module proc_config

Configuration for processors

Structs

PROC_CONFIG_SPEC

Configuration for processors

Type Aliases

PROC0_DAP_INSTID_R

Field PROC0_DAP_INSTID reader - Configure proc0 DAP instance ID.
Recommend that this is NOT changed until you require debug access in multi-chip environment
WARNING: do not set to 15 as this is reserved for RescueDP

PROC0_DAP_INSTID_W

Field PROC0_DAP_INSTID writer - Configure proc0 DAP instance ID.
Recommend that this is NOT changed until you require debug access in multi-chip environment
WARNING: do not set to 15 as this is reserved for RescueDP

PROC0_HALTED_R

Field PROC0_HALTED reader - Indication that proc0 has halted

PROC1_DAP_INSTID_R

Field PROC1_DAP_INSTID reader - Configure proc1 DAP instance ID.
Recommend that this is NOT changed until you require debug access in multi-chip environment
WARNING: do not set to 15 as this is reserved for RescueDP

PROC1_DAP_INSTID_W

Field PROC1_DAP_INSTID writer - Configure proc1 DAP instance ID.
Recommend that this is NOT changed until you require debug access in multi-chip environment
WARNING: do not set to 15 as this is reserved for RescueDP

PROC1_HALTED_R

Field PROC1_HALTED reader - Indication that proc1 has halted

R

Register PROC_CONFIG reader

W

Register PROC_CONFIG writer