rp2040_pac::usbctrl_regs

Module addr_endp

Device address and endpoint control

Structs

ADDR_ENDP_SPEC

Device address and endpoint control

Type Aliases

ADDRESS_R

Field ADDRESS reader - In device mode, the address that the device should respond to. Set in response to a SET_ADDR setup packet from the host. In host mode set to the address of the device to communicate with.

ADDRESS_W

Field ADDRESS writer - In device mode, the address that the device should respond to. Set in response to a SET_ADDR setup packet from the host. In host mode set to the address of the device to communicate with.

ENDPOINT_R

Field ENDPOINT reader - Device endpoint to send data to. Only valid for HOST mode.

ENDPOINT_W

Field ENDPOINT writer - Device endpoint to send data to. Only valid for HOST mode.

R

Register ADDR_ENDP reader

W

Register ADDR_ENDP writer