Module rp2040_pac::usbctrl_regs::buff_status

Buffer status register. A bit set here indicates that a buffer has completed on the endpoint (if the buffer interrupt is enabled). It is possible for 2 buffers to be completed, so clearing the buffer status bit may instantly re set it on the next clock cycle.

Structs

BUFF_STATUS_SPEC

Buffer status register. A bit set here indicates that a buffer has completed on the endpoint (if the buffer interrupt is enabled). It is possible for 2 buffers to be completed, so clearing the buffer status bit may instantly re set it on the next clock cycle.

Type Aliases

EP0_IN_R

Field EP0_IN reader -

EP0_IN_W

Field EP0_IN writer -

EP0_OUT_R

Field EP0_OUT reader -

EP0_OUT_W

Field EP0_OUT writer -

EP1_IN_R

Field EP1_IN reader -

EP1_IN_W

Field EP1_IN writer -

EP1_OUT_R

Field EP1_OUT reader -

EP1_OUT_W

Field EP1_OUT writer -

EP2_IN_R

Field EP2_IN reader -

EP2_IN_W

Field EP2_IN writer -

EP2_OUT_R

Field EP2_OUT reader -

EP2_OUT_W

Field EP2_OUT writer -

EP3_IN_R

Field EP3_IN reader -

EP3_IN_W

Field EP3_IN writer -

EP3_OUT_R

Field EP3_OUT reader -

EP3_OUT_W

Field EP3_OUT writer -

EP4_IN_R

Field EP4_IN reader -

EP4_IN_W

Field EP4_IN writer -

EP4_OUT_R

Field EP4_OUT reader -

EP4_OUT_W

Field EP4_OUT writer -

EP5_IN_R

Field EP5_IN reader -

EP5_IN_W

Field EP5_IN writer -

EP5_OUT_R

Field EP5_OUT reader -

EP5_OUT_W

Field EP5_OUT writer -

EP6_IN_R

Field EP6_IN reader -

EP6_IN_W

Field EP6_IN writer -

EP6_OUT_R

Field EP6_OUT reader -

EP6_OUT_W

Field EP6_OUT writer -

EP7_IN_R

Field EP7_IN reader -

EP7_IN_W

Field EP7_IN writer -

EP7_OUT_R

Field EP7_OUT reader -

EP7_OUT_W

Field EP7_OUT writer -

EP8_IN_R

Field EP8_IN reader -

EP8_IN_W

Field EP8_IN writer -

EP8_OUT_R

Field EP8_OUT reader -

EP8_OUT_W

Field EP8_OUT writer -

EP9_IN_R

Field EP9_IN reader -

EP9_IN_W

Field EP9_IN writer -

EP9_OUT_R

Field EP9_OUT reader -

EP9_OUT_W

Field EP9_OUT writer -

EP10_IN_R

Field EP10_IN reader -

EP10_IN_W

Field EP10_IN writer -

EP10_OUT_R

Field EP10_OUT reader -

EP10_OUT_W

Field EP10_OUT writer -

EP11_IN_R

Field EP11_IN reader -

EP11_IN_W

Field EP11_IN writer -

EP11_OUT_R

Field EP11_OUT reader -

EP11_OUT_W

Field EP11_OUT writer -

EP12_IN_R

Field EP12_IN reader -

EP12_IN_W

Field EP12_IN writer -

EP12_OUT_R

Field EP12_OUT reader -

EP12_OUT_W

Field EP12_OUT writer -

EP13_IN_R

Field EP13_IN reader -

EP13_IN_W

Field EP13_IN writer -

EP13_OUT_R

Field EP13_OUT reader -

EP13_OUT_W

Field EP13_OUT writer -

EP14_IN_R

Field EP14_IN reader -

EP14_IN_W

Field EP14_IN writer -

EP14_OUT_R

Field EP14_OUT reader -

EP14_OUT_W

Field EP14_OUT writer -

EP15_IN_R

Field EP15_IN reader -

EP15_IN_W

Field EP15_IN writer -

EP15_OUT_R

Field EP15_OUT reader -

EP15_OUT_W

Field EP15_OUT writer -

R

Register BUFF_STATUS reader

W

Register BUFF_STATUS writer