rp2040_pac::usbctrl_regs

Module sie_status

SIE status register

Structs

SIE_STATUS_SPEC

SIE status register

Enums

LINE_STATE_A

USB bus line state

Type Aliases

ACK_REC_R

Field ACK_REC reader - ACK received. Raised by both host and device.

ACK_REC_W

Field ACK_REC writer - ACK received. Raised by both host and device.

BIT_STUFF_ERROR_R

Field BIT_STUFF_ERROR reader - Bit Stuff Error. Raised by the Serial RX engine.

BIT_STUFF_ERROR_W

Field BIT_STUFF_ERROR writer - Bit Stuff Error. Raised by the Serial RX engine.

BUS_RESET_R

Field BUS_RESET reader - Device: bus reset received

BUS_RESET_W

Field BUS_RESET writer - Device: bus reset received

CONNECTED_R

Field CONNECTED reader - Device: connected

CONNECTED_W

Field CONNECTED writer - Device: connected

CRC_ERROR_R

Field CRC_ERROR reader - CRC Error. Raised by the Serial RX engine.

CRC_ERROR_W

Field CRC_ERROR writer - CRC Error. Raised by the Serial RX engine.

DATA_SEQ_ERROR_R

Field DATA_SEQ_ERROR reader - Data Sequence Error.

DATA_SEQ_ERROR_W

Field DATA_SEQ_ERROR writer - Data Sequence Error.

LINE_STATE_R

Field LINE_STATE reader - USB bus line state

NAK_REC_R

Field NAK_REC reader - Host: NAK received

NAK_REC_W

Field NAK_REC writer - Host: NAK received

R

Register SIE_STATUS reader

RESUME_R

Field RESUME reader - Host: Device has initiated a remote resume. Device: host has initiated a resume.

RESUME_W

Field RESUME writer - Host: Device has initiated a remote resume. Device: host has initiated a resume.

RX_OVERFLOW_R

Field RX_OVERFLOW reader - RX overflow is raised by the Serial RX engine if the incoming data is too fast.

RX_OVERFLOW_W

Field RX_OVERFLOW writer - RX overflow is raised by the Serial RX engine if the incoming data is too fast.

RX_TIMEOUT_R

Field RX_TIMEOUT reader - RX timeout is raised by both the host and device if an ACK is not received in the maximum time specified by the USB spec.

RX_TIMEOUT_W

Field RX_TIMEOUT writer - RX timeout is raised by both the host and device if an ACK is not received in the maximum time specified by the USB spec.

SETUP_REC_R

Field SETUP_REC reader - Device: Setup packet received

SETUP_REC_W

Field SETUP_REC writer - Device: Setup packet received

SPEED_R

Field SPEED reader - Host: device speed. Disconnected = 00, LS = 01, FS = 10

SPEED_W

Field SPEED writer - Host: device speed. Disconnected = 00, LS = 01, FS = 10

STALL_REC_R

Field STALL_REC reader - Host: STALL received

STALL_REC_W

Field STALL_REC writer - Host: STALL received

SUSPENDED_R

Field SUSPENDED reader - Bus in suspended state. Valid for device and host. Host and device will go into suspend if neither Keep Alive / SOF frames are enabled.

SUSPENDED_W

Field SUSPENDED writer - Bus in suspended state. Valid for device and host. Host and device will go into suspend if neither Keep Alive / SOF frames are enabled.

TRANS_COMPLETE_R

Field TRANS_COMPLETE reader - Transaction complete.

TRANS_COMPLETE_W

Field TRANS_COMPLETE writer - Transaction complete.

VBUS_DETECTED_R

Field VBUS_DETECTED reader - Device: VBUS Detected

VBUS_OVER_CURR_R

Field VBUS_OVER_CURR reader - VBUS over current detected

W

Register SIE_STATUS writer