Module rp2040_pac::vreg_and_chip_reset::chip_reset

Chip reset control and status

Structs

CHIP_RESET_SPEC

Chip reset control and status

HAD_POR_R

Field HAD_POR reader - Last reset was from the power-on reset or brown-out detection blocks

HAD_PSM_RESTART_R

Field HAD_PSM_RESTART reader - Last reset was from the debug port

HAD_RUN_R

Field HAD_RUN reader - Last reset was from the RUN pin

PSM_RESTART_FLAG_R

Field PSM_RESTART_FLAG reader - This is set by psm_restart from the debugger.
Its purpose is to branch bootcode to a safe mode when the debugger has issued a psm_restart in order to recover from a boot lock-up.
In the safe mode the debugger can repair the boot code, clear this flag then reboot the processor.

PSM_RESTART_FLAG_W

Field PSM_RESTART_FLAG writer - This is set by psm_restart from the debugger.
Its purpose is to branch bootcode to a safe mode when the debugger has issued a psm_restart in order to recover from a boot lock-up.
In the safe mode the debugger can repair the boot code, clear this flag then reboot the processor.

R

Register CHIP_RESET reader

W

Register CHIP_RESET writer