Module rp2040_pac::watchdog

WATCHDOG

Modules

ctrl
Watchdog control\n The rst_wdsel register determines which subsystems are reset when the watchdog is triggered.\n The watchdog can be triggered in software.

load
Load the watchdog timer. The maximum setting is 0xffffff which corresponds to 0xffffff / 2 ticks before triggering a watchdog reset (see errata RP2040-E1).

reason
Logs the reason for the last reset. Both bits are zero for the case of a hardware reset.

scratch0
Scratch register. Information persists through soft reset of the chip.

scratch1
Scratch register. Information persists through soft reset of the chip.

scratch2
Scratch register. Information persists through soft reset of the chip.

scratch3
Scratch register. Information persists through soft reset of the chip.

scratch4
Scratch register. Information persists through soft reset of the chip.

scratch5
Scratch register. Information persists through soft reset of the chip.

scratch6
Scratch register. Information persists through soft reset of the chip.

scratch7
Scratch register. Information persists through soft reset of the chip.

tick
Controls the tick generator

Structs

RegisterBlock
Register block

Type Definitions

CTRL
Watchdog control\n The rst_wdsel register determines which subsystems are reset when the watchdog is triggered.\n The watchdog can be triggered in software.

LOAD
Load the watchdog timer. The maximum setting is 0xffffff which corresponds to 0xffffff / 2 ticks before triggering a watchdog reset (see errata RP2040-E1).

REASON
Logs the reason for the last reset. Both bits are zero for the case of a hardware reset.

SCRATCH0
Scratch register. Information persists through soft reset of the chip.

SCRATCH1
Scratch register. Information persists through soft reset of the chip.

SCRATCH2
Scratch register. Information persists through soft reset of the chip.

SCRATCH3
Scratch register. Information persists through soft reset of the chip.

SCRATCH4
Scratch register. Information persists through soft reset of the chip.

SCRATCH5
Scratch register. Information persists through soft reset of the chip.

SCRATCH6
Scratch register. Information persists through soft reset of the chip.

SCRATCH7
Scratch register. Information persists through soft reset of the chip.

TICK
Controls the tick generator