Module rp2040_pac::watchdog

WATCHDOG

Modules

ctrl

Watchdog control
The rst_wdsel register determines which subsystems are reset when the watchdog is triggered.
The watchdog can be triggered in software.

load

Load the watchdog timer. The maximum setting is 0xffffff which corresponds to 0xffffff / 2 ticks before triggering a watchdog reset (see errata RP2040-E1).

reason

Logs the reason for the last reset. Both bits are zero for the case of a hardware reset.

scratch0

Scratch register. Information persists through soft reset of the chip.

scratch1

Scratch register. Information persists through soft reset of the chip.

scratch2

Scratch register. Information persists through soft reset of the chip.

scratch3

Scratch register. Information persists through soft reset of the chip.

scratch4

Scratch register. Information persists through soft reset of the chip.

scratch5

Scratch register. Information persists through soft reset of the chip.

scratch6

Scratch register. Information persists through soft reset of the chip.

scratch7

Scratch register. Information persists through soft reset of the chip.

tick

Controls the tick generator

Structs

RegisterBlock

Register block

Type Definitions

CTRL

CTRL register accessor: an alias for Reg<CTRL_SPEC>

LOAD

LOAD register accessor: an alias for Reg<LOAD_SPEC>

REASON

REASON register accessor: an alias for Reg<REASON_SPEC>

SCRATCH0

SCRATCH0 register accessor: an alias for Reg<SCRATCH0_SPEC>

SCRATCH1

SCRATCH1 register accessor: an alias for Reg<SCRATCH1_SPEC>

SCRATCH2

SCRATCH2 register accessor: an alias for Reg<SCRATCH2_SPEC>

SCRATCH3

SCRATCH3 register accessor: an alias for Reg<SCRATCH3_SPEC>

SCRATCH4

SCRATCH4 register accessor: an alias for Reg<SCRATCH4_SPEC>

SCRATCH5

SCRATCH5 register accessor: an alias for Reg<SCRATCH5_SPEC>

SCRATCH6

SCRATCH6 register accessor: an alias for Reg<SCRATCH6_SPEC>

SCRATCH7

SCRATCH7 register accessor: an alias for Reg<SCRATCH7_SPEC>

TICK

TICK register accessor: an alias for Reg<TICK_SPEC>