rp2040_pac::xip_ssi
Module spi_ctrlr0

SPI control

Structs

SPI_CTRLR0_SPEC

SPI control

Enums

INST_L_A

Instruction length (0/4/8/16b)

TRANS_TYPE_A

Address and instruction transfer format

Type Aliases

ADDR_L_R

Field ADDR_L reader - Address length (0b-60b in 4b increments)

ADDR_L_W

Field ADDR_L writer - Address length (0b-60b in 4b increments)

INST_DDR_EN_R

Field INST_DDR_EN reader - Instruction DDR transfer enable

INST_DDR_EN_W

Field INST_DDR_EN writer - Instruction DDR transfer enable

INST_L_R

Field INST_L reader - Instruction length (0/4/8/16b)

INST_L_W

Field INST_L writer - Instruction length (0/4/8/16b)

R

Register SPI_CTRLR0 reader

SPI_DDR_EN_R

Field SPI_DDR_EN reader - SPI DDR transfer enable

SPI_DDR_EN_W

Field SPI_DDR_EN writer - SPI DDR transfer enable

SPI_RXDS_EN_R

Field SPI_RXDS_EN reader - Read data strobe enable

SPI_RXDS_EN_W

Field SPI_RXDS_EN writer - Read data strobe enable

TRANS_TYPE_R

Field TRANS_TYPE reader - Address and instruction transfer format

TRANS_TYPE_W

Field TRANS_TYPE writer - Address and instruction transfer format

W

Register SPI_CTRLR0 writer

WAIT_CYCLES_R

Field WAIT_CYCLES reader - Wait cycles between control frame transmit and data reception (in SCLK cycles)

WAIT_CYCLES_W

Field WAIT_CYCLES writer - Wait cycles between control frame transmit and data reception (in SCLK cycles)

XIP_CMD_R

Field XIP_CMD reader - SPI Command to send in XIP mode (INST_L = 8-bit) or to append to Address (INST_L = 0-bit)

XIP_CMD_W

Field XIP_CMD writer - SPI Command to send in XIP mode (INST_L = 8-bit) or to append to Address (INST_L = 0-bit)