rp2040_pac::xosc

Module ctrl

Crystal Oscillator Control

Structs

CTRL_SPEC

Crystal Oscillator Control

Enums

ENABLE_A

On power-up this field is initialised to DISABLE and the chip runs from the ROSC.
If the chip has subsequently been programmed to run from the XOSC then setting this field to DISABLE may lock-up the chip. If this is a concern then run the clk_ref from the ROSC and enable the clk_sys RESUS feature.
The 12-bit code is intended to give some protection against accidental writes. An invalid setting will enable the oscillator.

FREQ_RANGE_A

Frequency range. This resets to 0xAA0 and cannot be changed.

Type Aliases

ENABLE_R

Field ENABLE reader - On power-up this field is initialised to DISABLE and the chip runs from the ROSC.
If the chip has subsequently been programmed to run from the XOSC then setting this field to DISABLE may lock-up the chip. If this is a concern then run the clk_ref from the ROSC and enable the clk_sys RESUS feature.
The 12-bit code is intended to give some protection against accidental writes. An invalid setting will enable the oscillator.

ENABLE_W

Field ENABLE writer - On power-up this field is initialised to DISABLE and the chip runs from the ROSC.
If the chip has subsequently been programmed to run from the XOSC then setting this field to DISABLE may lock-up the chip. If this is a concern then run the clk_ref from the ROSC and enable the clk_sys RESUS feature.
The 12-bit code is intended to give some protection against accidental writes. An invalid setting will enable the oscillator.

FREQ_RANGE_R

Field FREQ_RANGE reader - Frequency range. This resets to 0xAA0 and cannot be changed.

FREQ_RANGE_W

Field FREQ_RANGE writer - Frequency range. This resets to 0xAA0 and cannot be changed.

R

Register CTRL reader

W

Register CTRL writer